yet_another_parachain_runtime

Crate bridge_hub_test_utils

Module contains predefined test-case scenarios for “BridgeHub” Runtimes.

Modules

test_cases

Module contains predefined test-case scenarios for Runtime with bridging capabilities.

test_data

Generating test data, used by all tests.

Structs

CollatorSessionKey

CollatorSessionKeys

ExtBuilder

Basic builder based on balances, collators and pallet_session.

RuntimeHelper

SlotDurations

Enums

GovernanceOrigin

Enum representing governance origin/location.

XcmReceivedFrom

Traits

BasicParachainRuntime

A set of traits for a minimal parachain runtime, that may be used in conjunction with the ExtBuilder and the RuntimeHelper.

Functions

assert_metadata

assert_total

check_sane_fees_values

A helper function for comparing the actual value of a fee constant with its estimated value. The estimated value can be overestimated (overestimate_in_percent), and if the difference to the actual value is below margin_overestimate_diff_in_percent_for_lowering, we should lower the actual value.

diff_as_percent

ensure_opened_bridge

Helper function to open the bridge/lane for source and destination while ensuring all required balances are placed into the SA of the source.

mock_open_hrmp_channel

Helper function which emulates opening HRMP channel which is needed for XcmpQueue to pass.

open_bridge_with_extrinsic

Utility for opening bridge with dedicated pallet_xcm_bridge_hub’s extrinsic.

open_bridge_with_storage

Utility for opening bridge directly inserting data to the pallet_xcm_bridge_hub’s storage (used only for legacy purposes).

test_header

Get a header for testing.

Type Aliases

AccountIdOf

BalanceOf

RuntimeCallOf

RuntimeOriginOf

SessionKeysOf

ValidatorIdOf