yet_another_parachain_runtime

Crate cumulus_client_consensus_common

Modules

import_queue

(unstable) Composable utilities for constructing import queues for parachains.

Structs

ParachainBlockImport

Parachain specific block import.

ParachainCandidate

The result from building a collation.

ParentSearchResult

A potential parent block returned from find_parent_for_building

Enums

LevelLimit

Upper bound to the number of leaves allowed for each level of the blockchain.

ParentSearchParams

Constants

MAX_LEAVES_PER_LEVEL_SENSIBLE_DEFAULT

Value good enough to be used with parachains using the current backend implementation that ships with Substrate. This value may change in the future.

Traits

ParachainBlockImportMarker

Marker trait denoting a block import type that fits the parachain requirements.

ValidationCodeHashProvider

Provides the hash of validation code used for authoring/execution of blocks at a given hash.

Functions

fetch_included_from_relay_chain

Fetch the included block from the relay chain.

finalized_heads

Returns a stream that will yield finalized heads for the given para_id.

find_parent_for_building

Find the best parent block to build on.

get_relay_slot

Get the relay slot from a header.

get_relay_slot_and_timestamp

Get the relay slot and timestamp from a header.

load_abridged_host_configuration

Reads abridged host configuration from the relay chain storage at the given relay parent.

old_finalized_hash

Resolve the previously-finalized block hash: the parent of the first block in tree_route, or fallback_parent (the just-finalized header’s parent) when the route is empty or that header can’t be loaded.

spawn_parachain_consensus_tasks

Spawns the essential finalization tasks for parachain consensus.