yet_another_parachain_runtime

Crate cumulus_pallet_aura_ext

Cumulus extension pallet for AuRa

This pallet extends the Substrate AuRa pallet to make it compatible with parachains. It provides the Pallet, the Config and the GenesisConfig.

It is also required that the parachain runtime uses the provided BlockExecutor to properly check the constructed block on the relay chain.

cumulus_pallet_parachain_system::register_validate_block! {
    Runtime = Runtime,
    BlockExecutor = cumulus_pallet_aura_ext::BlockExecutor::<Runtime, Executive>,
}

Modules

consensus_hook

The definition of a FixedVelocityConsensusHook for consensus logic to manage block velocity.

migration

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

BlockExecutor

The block executor used when validating a PoV at the relay chain.

FixedVelocityConsensusHook

A consensus hook that enforces fixed block production velocity and unincluded segment capacity.

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Traits

Config

Configuration trait of this pallet.

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.