yet_another_parachain_runtime

Crate cumulus_pallet_parachain_system

cumulus-pallet-parachain-system is a base pallet for Cumulus-based parachains.

This pallet handles low-level details of being a parachain. Its responsibilities include:

ingestion of the parachain validation data;

ingestion and dispatch of incoming downward and lateral messages;

coordinating upgrades with the Relay Chain; and

communication of parachain outputs, such as sent messages, signaling an upgrade, etc.

Users must ensure that they register this pallet as an inherent provider.

Modules

consensus_hook

The definition of a ConsensusHook trait for consensus logic to manage the backlog of parachain blocks ready to submit to the relay chain, as well as some basic implementations.

migration

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

parachain_inherent

Cumulus parachain inherent related structures.

relay_state_snapshot

Relay chain state proof provides means for accessing part of relay chain storage for reads.

ump_constants

validate_block

A module that enables a runtime to work as parachain.

weights

Autogenerated weights for cumulus_pallet_parachain_system

Macros

register_validate_block

Structs

Ancestor

Ancestor of the block being currently executed, not yet included into the relay chain.

AnyRelayNumber

Provides an implementation of CheckAssociatedRelayNumber.

ExpectParentIncluded

A special consensus hook for handling the migration to asynchronous backing gracefully, even if collators haven’t been updated to provide the last included parent in the state proof yet.

GenesisConfig

Can be used to configure the genesis state of this pallet.

MessagingStateSnapshot

A snapshot of some messaging related state of relay chain pertaining to the current parachain.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

ParachainSetCode

Type that implements SetCode.

RelayChainState

Holds the most recent relay-parent state root and block number of the current parachain block.

RelayChainStateProof

A state proof extracted from the relay chain.

RelayNumberMonotonicallyIncreases

Provides an implementation of CheckAssociatedRelayNumber.

RelayNumberStrictlyIncreases

Provides an implementation of CheckAssociatedRelayNumber.

RelaychainDataProvider

Implements BlockNumberProvider and RelaychainStateProvider that returns relevant relay data fetched from validation data.

UsedBandwidth

Bandwidth used by a parachain block(s).

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

CheckAssociatedRelayNumber

Something that can check the associated relay block number.

Config

Configuration trait of this pallet.

ConsensusHook

The consensus hook for dealing with the unincluded segment.

OnSystemEvent

Something that should be informed about system related events.

RelaychainStateProvider

This exposes the RelayChainState to other runtime modules.

WeightInfo

Weight functions needed for cumulus_pallet_parachain_system.

Type Aliases

AggregatedUnincludedSegment

Storage field that keeps track of bandwidth used by the unincluded segment along with the latest HRMP watermark. Used for limiting the acceptance of new blocks with respect to relay chain constraints.

AnnouncedHrmpMessagesPerCandidate

The number of HRMP messages we observed in on_initialize and thus used that number for announcing the weight of on_initialize and on_finalize.

CustomValidationHeadData

A custom head data that should be returned as result of validate_block.

DidSetValidationCode

Were the validation data set to notify the relay chain?

HostConfiguration

The parachain host configuration that was obtained from the relay parent.

HrmpOutboundMessages

HRMP messages that were sent in a block.

HrmpWatermark

HRMP watermark that was set in a block.

LastDmqMqcHead

The last downward message queue chain head we have observed.

LastHrmpMqcHeads

The message queue chain heads we have observed per each channel incoming channel.

LastProcessedDownwardMessage

The last processed downward message.

LastProcessedHrmpMessage

The last processed HRMP message.

LastRelayChainBlockNumber

The relay chain block number associated with the last parachain block.

MaxDmpMessageLenOf

The max length of a DMP message.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NewValidationCode

Validation code that is set by the parachain and is to be communicated to collator and consequently the relay-chain.

PendingUpwardMessages

Upward messages that are still pending and not yet send to the relay chain.

PendingValidationCode

In case of a scheduled upgrade, this storage field contains the validation code to be applied.

ProcessedDownwardMessages

Number of downward messages processed in a block.

RelayStateProof

The state proof for the last relay parent block.

RelaychainBlockNumberProviderDeprecated

Implements BlockNumberProvider that returns relay chain block number fetched from validation data.

RelevantMessagingState

The snapshot of some state related to messaging relevant to the current parachain as per the relay parent.

ReservedDmpWeightOverride

The weight we reserve at the beginning of the block for processing DMP messages. This overrides the amount set in the Config trait.

ReservedXcmpWeightOverride

The weight we reserve at the beginning of the block for processing XCMP messages. This overrides the amount set in the Config trait.

UnincludedSegment

Latest included block descendants the runtime accepted. In other words, these are ancestors of the currently executing block which have not been included in the observed relay-chain state.

UpgradeGoAhead

Optional upgrade go-ahead signal from the relay-chain.

UpgradeRestrictionSignal

An option which indicates if the relay-chain restricts signalling a validation code upgrade. In other words, if this is Some and NewValidationCode is Some then the produced candidate will be invalid.

UpwardDeliveryFeeFactor

The factor to multiply the base delivery fee by for UMP.

UpwardMessages

Upward messages that were sent in a block.

ValidationData

The PersistedValidationData set for this block.