yet_another_parachain_runtime::cumulus_primitives_core::relay_chain::v9

Module well_known_keys

A declarations of storage keys where an external observer can find some interesting data.

Constants

ACTIVE_CONFIG

The currently active host configuration.

AUTHORITIES

The authorities for the current epoch.

CURRENT_BLOCK_RANDOMNESS

The current relay chain block randomness

CURRENT_SLOT

The current slot number.

EPOCH_INDEX

The current epoch index.

NEXT_AUTHORITIES

The authorities for the next epoch.

ONE_EPOCH_AGO_RANDOMNESS

The randomness for one epoch ago

TWO_EPOCHS_AGO_RANDOMNESS

The randomness for two epochs ago

Functions

dmq_mqc_head

The MQC head for the downward message queue of the given para. See more in the Dmp module.

hrmp_channels

The HRMP channel for the given identifier.

hrmp_egress_channel_index

The list of outbound channels for the given para.

hrmp_ingress_channel_index

The list of inbound channels for the given para.

para_head

Hash of the committed head data for a given registered para.

relay_dispatch_queue_remaining_capacity

The upward message dispatch queue remaining capacity for the given para id.

relay_dispatch_queue_sizeDeprecated

The upward message dispatch queue for the given para id.

relay_dispatch_queue_size_typedDeprecated

Type safe version of relay_dispatch_queue_size.

upgrade_go_ahead_signal

The signal that indicates whether the parachain should go-ahead with the proposed validation code upgrade.

upgrade_restriction_signal

The signal that indicates whether the parachain is disallowed to signal an upgrade at this relay-parent.