yet_another_parachain_runtime::emulated_integration_tests_common::dap_helpers

Function test_accumulate_forward_transfers_to_asset_hub

pub fn test_accumulate_forward_transfers_to_asset_hub<Sender, AH>(
    fund_sender: fn(<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, u128),
    get_relay_block: fn() -> u32,
    set_relay_block: fn(u32),
)
where
    Sender: Chain + TestExt,
    <Sender as Chain>::Runtime: Config<AccountId = <<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId> + Config<Balance = u128> + Config,
    <Sender as Chain>::RuntimeEvent: TryInto<Event<<Sender as Chain>::Runtime>>,
    Pallet<<Sender as Chain>::Runtime>: Hooks<u32>,
    <<Sender as Chain>::Runtime as Config>::MinTransferAmount: Get<u128>,
    <<Sender as Chain>::Runtime as Config>::TransferPeriod: Get<u32>,
    AH: Chain + TestExt,
    <AH as Chain>::Runtime: Config<AccountId = <<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId> + Config + Config<Balance = u128> + Config + Config,
    <AH as Chain>::RuntimeEvent: TryInto<Event<<AH as Chain>::Runtime>>,
    Pallet<<AH as Chain>::Runtime>: Hooks<u32>,

Tests that the accumulate-and-forward pallet accumulates native tokens, teleports them to the staging account of pallet-dap on AssetHub, and that pallet-dap’s on_idle subsequently drains and deactivates those funds into the main DAP buffer account.