yet_another_parachain_runtime::emulated_integration_tests_common::macros

Crate pallet_asset_conversion

Substrate Asset Conversion pallet

Substrate Asset Conversion pallet based on the Uniswap V2 logic.

Overview

This pallet allows you to:

create a liquidity pool for 2 assets

provide the liquidity and receive back an LP token

exchange the LP token back to assets

swap a specific amount of assets for another if there is a pool created, or

swap some assets for a specific amount of another.

query for an exchange price via a runtime call endpoint

query the size of a liquidity pool via a runtime api endpoint.

The quote_price_exact_tokens_for_tokens and quote_price_tokens_for_exact_tokens functions both take a path parameter of the route to take. If you want to swap from native asset to non-native asset 1, you would pass in a path of [DOT, 1] or [1, DOT]. If you want to swap from non-native asset 1 to non-native asset 2, you would pass in a path of [1, DOT, 2].

(For an example of configuring this pallet to use Location as an asset id, see the cumulus repo).

Here is an example state_call that asks for a quote of a pool of native versus asset 1:

curl -sS -H "Content-Type: application/json" -d \
'{"id":1, "jsonrpc":"2.0", "method": "state_call", "params": ["AssetConversionApi_quote_price_tokens_for_exact_tokens", "0x0101000000000000000000000011000000000000000000"]}' \
http://localhost:9933/

(This can be run against the kitchen sync node in the node folder of this repo.)

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_asset_conversion

Macros

_runtime_benchmarks_enabled

runtime_benchmarks_enabled

Structs

AccountIdConverter

PoolId to AccountId conversion.

AccountIdConverterNoSeed

PoolId to AccountId conversion without an addition arguments to the seed.

AddLiquidityAsset

A struct to represent an asset and its desired and minimum amounts for adding liquidity.

Ascending

Pool locator where the PoolId is a tuple of AssetKinds arranged in ascending order.

Chain

Pool locator that chains the First and Second implementations of PoolLocator.

GetReservesViewFunction

Returns the balance of each asset in the pool. The tuple result is in the order requested (not necessarily the same as pool order).

NativeOrWithIdFactory

Factory for creating a valid asset pairs with NativeOrWithId::Native always leading in the pair.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

PoolInfo

Stores the lp_token asset id a particular pool has been assigned.

QuotePriceExactTokensForTokensViewFunction

Gets a quote for swapping an exact amount of asset1 for asset2.

QuotePriceTokensForExactTokensViewFunction

Gets a quote for swapping amount of asset1 for an exact amount of asset2.

WithFirstAsset

Pool locator that mandates the inclusion of the specified FirstAsset in every asset pair.

Enums

Call

Pallet’s callable functions.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

AssetConversionApi

This runtime api allows people to query the size of the liquidity pools and quote prices for swaps.

BenchmarkHelper

Benchmark Helper

Config

Configuration trait of this pallet.

MutateLiquidity

Trait for providing methods to mutate liquidity pools. This includes creating pools, adding liquidity, and removing liquidity.

PoolLocator

Provides means to resolve the PoolId and AccountId from a pair of assets.

QuotePrice

Trait providing methods to quote swap prices between asset classes.

Swap

Trait for providing methods to swap between the various asset classes.

SwapCredit

Trait providing methods to swap between the various asset classes.

WeightInfo

Weight functions needed for pallet_asset_conversion.

Type Aliases

BalancePath

Represents a swap path with associated asset amounts indicating how much of the asset needs to be deposited to get the following asset’s amount withdrawn (this is inclusive of fees).

CreditOf

Credit of Config::Assets.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextPoolAssetId

Stores the PoolAssetId that is going to be used for the next lp token. This gets incremented whenever a new lp pool is created.

Pools

Map from PoolAssetId to PoolInfo. This establishes whether a pool has been officially created rather than people sending tokens directly to a pool’s public account.