yet_another_parachain_runtime::emulated_integration_tests_common::macros::pallet_asset_conversion

Trait BenchmarkHelper

pub trait BenchmarkHelper<AssetKind> {
    // Required method
    fn create_pair(seed1: u32, seed2: u32) -> (AssetKind, AssetKind);
}

Benchmark Helper

Required Methods

fn create_pair(seed1: u32, seed2: u32) -> (AssetKind, AssetKind)

Returns a valid assets pair for the pool creation.

When a specific asset, such as the native asset, is required in every pool, it should be returned for each odd-numbered seed.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.