yet_another_parachain_runtime::emulated_integration_tests_common::xcm_emulator::sp_tracing::tracing_subscriber::fmt

Struct Layer

pub struct Layer<S, N = DefaultFields, E = Format, W = fn() -> Stdout> { /* private fields */ }

A Layer that logs formatted representations of tracing events.

Examples

Constructing a layer with the default configuration:

use tracing_subscriber::{fmt, Registry};
use tracing_subscriber::prelude::*;

let subscriber = Registry::default()
    .with(fmt::Layer::default());

tracing::subscriber::set_global_default(subscriber).unwrap();

Overriding the layer’s behavior:

use tracing_subscriber::{fmt, Registry};
use tracing_subscriber::prelude::*;

let fmt_layer = fmt::layer()
   .with_target(false) // don't include event targets when logging
   .with_level(false); // don't include event levels when logging

let subscriber = Registry::default().with(fmt_layer);

Setting a custom event formatter:

use tracing_subscriber::fmt::{self, format, time};
use tracing_subscriber::prelude::*;

let fmt = format().with_timer(time::Uptime::default());
let fmt_layer = fmt::layer()
    .event_format(fmt)
    .with_target(false);

Implementations

impl<S> Layer<S>

pub fn new() -> Layer<S>

Returns a new Layer with the default configuration.

impl<S, N, E, W> Layer<S, N, E, W>
where S: Subscriber + for<'a> LookupSpan<'a>, N: for<'writer> FormatFields<'writer> + 'static, W: for<'writer> MakeWriter<'writer> + 'static,

pub fn event_format<E2>(self, e: E2) -> Layer<S, N, E2, W>
where E2: FormatEvent<S, N> + 'static,

Sets the event formatter that the layer being built will use to format events.

The event formatter may be any type implementing the FormatEvent trait, which is implemented for all functions taking a FmtContext, a Writer, and an Event.

Examples

Setting a type implementing FormatEvent as the formatter:

use tracing_subscriber::fmt::{self, format};

let layer = fmt::layer()
    .event_format(format().compact());

pub fn map_event_format<E2>(self, f: impl FnOnce(E) -> E2) -> Layer<S, N, E2, W>
where E2: FormatEvent<S, N> + 'static,

Updates the event formatter by applying a function to the existing event formatter.

This sets the event formatter that the layer being built will use to record fields.

Examples

Updating an event formatter:

let layer = tracing_subscriber::fmt::layer()
    .map_event_format(|e| e.compact());

impl<S, N, E, W> Layer<S, N, E, W>

pub fn with_writer<W2>(self, make_writer: W2) -> Layer<S, N, E, W2>
where W2: for<'writer> MakeWriter<'writer> + 'static,

Sets the MakeWriter that the layer being built will use to write events.

Examples

Using stderr rather than stdout:

use std::io;
use tracing_subscriber::fmt;

let layer = fmt::layer()
    .with_writer(io::stderr);

pub fn writer(&self) -> &W

Borrows the writer for this Layer.

pub fn writer_mut(&mut self) -> &mut W

Mutably borrows the writer for this Layer.

This method is primarily expected to be used with the reload::Handle::modify method.

Examples

let layer = fmt::layer().with_writer(non_blocking(std::io::stderr()));
let (layer, reload_handle) = reload::Layer::new(layer);
info!("This will be logged to stderr");
reload_handle.modify(|layer| *layer.writer_mut() = non_blocking(std::io::stdout()));
info!("This will be logged to stdout");

pub fn set_ansi(&mut self, ansi: bool)

Sets whether this layer should use ANSI terminal formatting escape codes (such as colors).

This method is primarily expected to be used with the reload::Handle::modify method when changing the writer.

pub fn with_test_writer(self) -> Layer<S, N, E, TestWriter>

Configures the layer to support libtest’s output capturing when used in unit tests.

See TestWriter for additional details.

Examples

Using TestWriter to let cargo test capture test output:

use std::io;
use tracing_subscriber::fmt;

let layer = fmt::layer()
    .with_test_writer();

pub fn with_ansi(self, ansi: bool) -> Layer<S, N, E, W>

Sets whether or not the formatter emits ANSI terminal escape codes for colors and other text formatting.

When the “ansi” crate feature flag is enabled, ANSI colors are enabled by default unless the NO_COLOR environment variable is set to a non-empty value. If the NO_COLOR environment variable is set to any non-empty value, then ANSI colors will be suppressed by default. The with_ansi and set_ansi methods can be used to forcibly enable ANSI colors, overriding any NO_COLOR environment variable.

Enabling ANSI escapes (calling with_ansi(true)) requires the “ansi” crate feature flag. Calling with_ansi(true) without the “ansi” feature flag enabled will panic if debug assertions are enabled, or print a warning otherwise.

This method itself is still available without the feature flag. This is to allow ANSI escape codes to be explicitly disabled without having to opt-in to the dependencies required to emit ANSI formatting. This way, code which constructs a formatter that should never emit ANSI escape codes can ensure that they are not used, regardless of whether or not other crates in the dependency graph enable the “ansi” feature flag.

pub fn log_internal_errors(self, log_internal_errors: bool) -> Layer<S, N, E, W>

Sets whether to write errors from FormatEvent to the writer. Defaults to true.

By default, fmt::Layer will write any FormatEvent-internal errors to the writer. These errors are unlikely and will only occur if there is a bug in the FormatEvent implementation or its dependencies.

If writing to the writer fails, the error message is printed to stderr as a fallback.

pub fn map_writer<W2>(self, f: impl FnOnce(W) -> W2) -> Layer<S, N, E, W2>
where W2: for<'writer> MakeWriter<'writer> + 'static,

Updates the MakeWriter by applying a function to the existing MakeWriter.

This sets the MakeWriter that the layer being built will use to write events.

Examples

Redirect output to stderr if level is <= WARN:

use tracing::Level;
use tracing_subscriber::fmt::{self, writer::MakeWriterExt};

let stderr = std::io::stderr.with_max_level(Level::WARN);
let layer = fmt::layer()
    .map_writer(move |w| stderr.or_else(w));

impl<S, N, L, T, W> Layer<S, N, Format<L, T>, W>
where N: for<'writer> FormatFields<'writer> + 'static,

pub fn with_timer<T2>(self, timer: T2) -> Layer<S, N, Format<L, T2>, W>

Use the given timer for span and event timestamps.

See the time module for the provided timer implementations.

Note that using the "time“” feature flag enables the additional time formatters UtcTime and LocalTime, which use the time crate to provide more sophisticated timestamp formatting options.

pub fn without_time(self) -> Layer<S, N, Format<L, ()>, W>

Do not emit timestamps with spans and event.

pub fn with_span_events(self, kind: FmtSpan) -> Layer<S, N, Format<L, T>, W>

Configures how synthesized events are emitted at points in the span lifecycle.

The following options are available:

FmtSpan::NONE: No events will be synthesized when spans are created, entered, exited, or closed. Data from spans will still be included as the context for formatted events. This is the default.

FmtSpan::NEW: An event will be synthesized when spans are created.

FmtSpan::ENTER: An event will be synthesized when spans are entered.

FmtSpan::EXIT: An event will be synthesized when spans are exited.

FmtSpan::CLOSE: An event will be synthesized when a span closes. If timestamps are enabled for this formatter, the generated event will contain fields with the span’s busy time (the total time for which it was entered) and idle time (the total time that the span existed but was not entered).

FmtSpan::ACTIVE: Events will be synthesized when spans are entered or exited.

FmtSpan::FULL: Events will be synthesized whenever a span is created, entered, exited, or closed. If timestamps are enabled, the close event will contain the span’s busy and idle time, as described above.

The options can be enabled in any combination. For instance, the following will synthesize events whenever spans are created and closed:

use tracing_subscriber::fmt;
use tracing_subscriber::fmt::format::FmtSpan;

let subscriber = fmt()
    .with_span_events(FmtSpan::NEW | FmtSpan::CLOSE)
    .finish();

Note that the generated events will only be part of the log output by this formatter; they will not be recorded by other Subscribers or by Layers added to this subscriber.

pub fn with_target(self, display_target: bool) -> Layer<S, N, Format<L, T>, W>

Sets whether or not an event’s target is displayed.

pub fn with_file(self, display_filename: bool) -> Layer<S, N, Format<L, T>, W>

Sets whether or not an event’s source code file path is displayed.

pub fn with_line_number( self, display_line_number: bool, ) -> Layer<S, N, Format<L, T>, W>

Sets whether or not an event’s source code line number is displayed.

pub fn with_level(self, display_level: bool) -> Layer<S, N, Format<L, T>, W>

Sets whether or not an event’s level is displayed.

pub fn with_thread_ids( self, display_thread_ids: bool, ) -> Layer<S, N, Format<L, T>, W>

Sets whether or not the thread ID of the current thread is displayed when formatting events.

pub fn with_thread_names( self, display_thread_names: bool, ) -> Layer<S, N, Format<L, T>, W>

Sets whether or not the name of the current thread is displayed when formatting events.

pub fn compact(self) -> Layer<S, N, Format<Compact, T>, W>
where N: for<'writer> FormatFields<'writer> + 'static,

Sets the layer being built to use a less verbose formatter.

pub fn pretty(self) -> Layer<S, Pretty, Format<Pretty, T>, W>

Sets the layer being built to use an excessively pretty, human-readable formatter.

impl<S, N, E, W> Layer<S, N, E, W>

pub fn fmt_fields<N2>(self, fmt_fields: N2) -> Layer<S, N2, E, W>
where N2: for<'writer> FormatFields<'writer> + 'static,

Sets the field formatter that the layer being built will use to record fields.

pub fn map_fmt_fields<N2>(self, f: impl FnOnce(N) -> N2) -> Layer<S, N2, E, W>
where N2: for<'writer> FormatFields<'writer> + 'static,

Updates the field formatter by applying a function to the existing field formatter.

This sets the field formatter that the layer being built will use to record fields.

Examples

Updating a field formatter:

use tracing_subscriber::field::MakeExt;
let layer = tracing_subscriber::fmt::layer()
    .map_fmt_fields(|f| f.debug_alt());