yet_another_parachain_runtime::emulated_integration_tests_common::xcm_emulator::sp_tracing::tracing_subscriber::fmt

Struct Subscriber

pub struct Subscriber<N = DefaultFields, E = Format, F = LevelFilter, W = fn() -> Stdout> { /* private fields */ }

A Subscriber that logs formatted representations of tracing events.

This consists of an inner Formatter wrapped in a layer that performs filtering.

Implementations

impl Subscriber

pub const DEFAULT_MAX_LEVEL: LevelFilter = LevelFilter::INFO

The maximum verbosity level that is enabled by a Subscriber by default.

This can be overridden with the SubscriberBuilder::with_max_level method.

pub fn builder() -> SubscriberBuilder

Returns a new SubscriberBuilder for configuring a format subscriber.

pub fn new() -> Subscriber

Returns a new format subscriber with the default configuration.