yet_another_parachain_runtime::emulated_integration_tests_common::xcm_simulator

Struct XcmExecutor

pub struct XcmExecutor<Config>
where
    Config: Config,
{ /* private fields */ }

The XCM executor.

Implementations

impl<Config> XcmExecutor<Config>
where Config: Config,

pub fn holding(&self) -> &AssetsInHolding

pub fn set_holding(&mut self, v: AssetsInHolding)

pub fn holding_limit(&self) -> &usize

pub fn set_holding_limit(&mut self, v: usize)

pub fn origin(&self) -> &Option<Location>

pub fn set_origin(&mut self, v: Option<Location>)

pub fn original_origin(&self) -> &Location

pub fn set_original_origin(&mut self, v: Location)

pub fn trader(&self) -> &<Config as Config>::Trader

pub fn set_trader(&mut self, v: <Config as Config>::Trader)

pub fn error(&self) -> &Option<(u32, Error)>

pub fn set_error(&mut self, v: Option<(u32, Error)>)

pub fn total_surplus(&self) -> &Weight

pub fn set_total_surplus(&mut self, v: Weight)

pub fn total_refunded(&self) -> &Weight

pub fn set_total_refunded(&mut self, v: Weight)

pub fn error_handler(&self) -> &Xcm<<Config as Config>::RuntimeCall>

pub fn set_error_handler(&mut self, v: Xcm<<Config as Config>::RuntimeCall>)

pub fn error_handler_weight(&self) -> &Weight

pub fn set_error_handler_weight(&mut self, v: Weight)

pub fn appendix(&self) -> &Xcm<<Config as Config>::RuntimeCall>

pub fn set_appendix(&mut self, v: Xcm<<Config as Config>::RuntimeCall>)

pub fn appendix_weight(&self) -> &Weight

pub fn set_appendix_weight(&mut self, v: Weight)

pub fn transact_status(&self) -> &MaybeErrorCode

pub fn set_transact_status(&mut self, v: MaybeErrorCode)

pub fn fees_mode(&self) -> &FeesMode

pub fn set_fees_mode(&mut self, v: FeesMode)

pub fn fees(&self) -> &AssetsInHolding

pub fn set_fees(&mut self, value: AssetsInHolding)

pub fn topic(&self) -> &Option<[u8; 32]>

pub fn set_topic(&mut self, v: Option<[u8; 32]>)

pub fn asset_claimer(&self) -> Option<Location>

pub fn set_message_weight(&mut self, weight: Weight)

pub fn already_paid_fees(&self) -> bool

impl<Config> XcmExecutor<Config>
where Config: Config,

pub fn new( origin: impl Into<Location>, message_id: [u8; 32], ) -> XcmExecutor<Config>

pub fn post_process(self, xcm_weight: Weight) -> Outcome

Execute any final operations after having executed the XCM message. This includes refunding surplus weight, trapping extra holding funds, and returning any errors during execution.

pub fn bench_process( &mut self, xcm: Xcm<<Config as Config>::RuntimeCall>, ) -> Result<(), ExecutorError>

pub fn bench_post_process(self, xcm_weight: Weight) -> Outcome