yet_another_parachain_runtime::ethereum_standards

Module IERC20

@dev Interface combining the ERC-20 standard, its metadata extension, and EIP-2612 permit. Note: Due to ABI generation constraints, all interfaces are merged into a single contract.

interface IERC20 {
    event Transfer(address indexed from, address indexed to, uint256 value);
    event Approval(address indexed owner, address indexed spender, uint256 value);
    function totalSupply() external view returns (uint256);
    function balanceOf(address account) external view returns (uint256);
    function transfer(address to, uint256 value) external returns (bool);
    function allowance(address owner, address spender) external view returns (uint256);
    function approve(address spender, uint256 value) external returns (bool);
    function transferFrom(address from, address to, uint256 value) external returns (bool);
    function name() external view returns (string memory);
    function symbol() external view returns (string memory);
    function decimals() external view returns (uint8);
    function permit(address owner, address spender, uint256 value, uint256 deadline, uint8 v, bytes32 r, bytes32 s) external;
    function nonces(address owner) external view returns (uint256);
    function DOMAIN_SEPARATOR() external view returns (bytes32);
}

Structs

Approval

@dev Emitted when the allowance of a spender for an owner is set by a call to {approve}. value is the new allowance. Event with signature Approval(address,address,uint256) and selector 0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925.

DOMAIN_SEPARATORCall

@dev Returns the domain separator used in the encoding of the signature for {permit}, as defined by {EIP712}. Function with signature DOMAIN_SEPARATOR() and selector 0x3644e515.

DOMAIN_SEPARATORReturn

@dev Returns the domain separator used in the encoding of the signature for {permit}, as defined by {EIP712}. Container type for the return parameters of the DOMAIN_SEPARATOR() function.

Transfer

@dev Emitted when value tokens are moved from one account (from) to another (to).

allowanceCall

@dev Returns the remaining number of tokens that spender will be allowed to spend on behalf of owner through {transferFrom}. This is zero by default.

allowanceReturn

@dev Returns the remaining number of tokens that spender will be allowed to spend on behalf of owner through {transferFrom}. This is zero by default.

approveCall

@dev Sets a value amount of tokens as the allowance of spender over the caller’s tokens.

approveReturn

@dev Sets a value amount of tokens as the allowance of spender over the caller’s tokens.

balanceOfCall

@dev Returns the value of tokens owned by account. Function with signature balanceOf(address) and selector 0x70a08231.

balanceOfReturn

@dev Returns the value of tokens owned by account. Container type for the return parameters of the balanceOf(address) function.

decimalsCall

@dev Returns the decimals places of the token. Function with signature decimals() and selector 0x313ce567.

decimalsReturn

@dev Returns the decimals places of the token. Container type for the return parameters of the decimals() function.

nameCall

@dev Returns the name of the token. Function with signature name() and selector 0x06fdde03.

nameReturn

@dev Returns the name of the token. Container type for the return parameters of the name() function.

noncesCall

@dev Returns the current nonce for owner. This value must be included whenever a signature is generated for {permit}.

noncesReturn

@dev Returns the current nonce for owner. This value must be included whenever a signature is generated for {permit}.

permitCall

@dev Sets value as the allowance of spender over owner’s tokens, given owner’s signed approval.

permitReturn

@dev Sets value as the allowance of spender over owner’s tokens, given owner’s signed approval.

symbolCall

@dev Returns the symbol of the token. Function with signature symbol() and selector 0x95d89b41.

symbolReturn

@dev Returns the symbol of the token. Container type for the return parameters of the symbol() function.

totalSupplyCall

@dev Returns the value of tokens in existence. Function with signature totalSupply() and selector 0x18160ddd.

totalSupplyReturn

@dev Returns the value of tokens in existence. Container type for the return parameters of the totalSupply() function.

transferCall

@dev Moves a value amount of tokens from the caller’s account to to.

transferFromCall

@dev Moves a value amount of tokens from from to to using the allowance mechanism. value is then deducted from the caller’s allowance.

transferFromReturn

@dev Moves a value amount of tokens from from to to using the allowance mechanism. value is then deducted from the caller’s allowance.

transferReturn

@dev Moves a value amount of tokens from the caller’s account to to.

Enums

IERC20Calls

Container for all the IERC20 function calls.

IERC20Events

Container for all the IERC20 events.