yet_another_parachain_runtime::frame_election_provider_support

Trait ElectionProvider

pub trait ElectionProvider {
    type AccountId;
    type BlockNumber;
    type Error: Debug + PartialEq;
    type MaxWinnersPerPage: Get<u32>;
    type MaxBackersPerWinner: Get<u32>;
    type MaxBackersPerWinnerFinal: Get<u32>;
    type Pages: Get<u32>;
    type DataProvider: ElectionDataProvider<AccountId = Self::AccountId, BlockNumber = Self::BlockNumber>;

    // Required methods
    fn elect(
        page: u32,
    ) -> Result<BoundedSupports<Self::AccountId, Self::MaxWinnersPerPage, Self::MaxBackersPerWinner>, Self::Error>;
    fn duration() -> Self::BlockNumber;
    fn start() -> Result<(), Self::Error>;
    fn status() -> Result<bool, ()>;

    // Provided methods
    fn msp() -> u32 { ... }
    fn lsp() -> u32 { ... }
    fn desired_targets_checked() -> Result<u32, &'static str> { ... }
    fn duration_with_export() -> Self::BlockNumber
       where Self::BlockNumber: From<u32> + Add<Output = Self::BlockNumber> { ... }
    fn asap() { ... }
}

Something that can compute the result of an election and pass it back to the caller in a paged way.

Required Associated Types

type AccountId

The account ID identifier;

type BlockNumber

The block number type.

type Error: Debug + PartialEq

The error type returned by the provider;

type MaxWinnersPerPage: Get<u32>

The maximum number of winners per page in results returned by this election provider.

A winner is an AccountId that is part of the final election result.

type MaxBackersPerWinner: Get<u32>

The maximum number of backers that a single page may have in results returned by this election provider.

A backer is an AccountId that “backs” one or more winners. For example, in the context of nominated proof of stake, a backer is a voter that nominates a winner validator in the election result.

type MaxBackersPerWinnerFinal: Get<u32>

Same as Self::MaxBackersPerWinner, but across all pages.

If Self::Pages is set to 0, a reasonable value is Self::MaxBackersPerWinner. For multi-page elections, a reasonable value is the range of Self::MaxBackersPerWinner to Self::Pages * Self::MaxBackersPerWinner.

type Pages: Get<u32>

The number of pages that this election provider supports.

type DataProvider: ElectionDataProvider<AccountId = Self::AccountId, BlockNumber = Self::BlockNumber>

The data provider of the election.

Required Methods

fn elect( page: u32, ) -> Result<BoundedSupports<Self::AccountId, Self::MaxWinnersPerPage, Self::MaxBackersPerWinner>, Self::Error>

Elect a new set of winners.

A complete election may require multiple calls to ElectionProvider::elect if ElectionProvider::Pages is higher than one.

The result is returned in a target major format, namely as vector of supports.

fn duration() -> Self::BlockNumber

Return the duration of your election.

This excludes the duration of the export. For that, use Self::duration_with_export.

fn start() -> Result<(), Self::Error>

Signal that the election should start

fn status() -> Result<bool, ()>

Indicate whether this election provider is currently ongoing an asynchronous election.

Err(()) should signal that we are not doing anything, and elect should def. not be called. Ok(false) means we are doing something, but work is still ongoing. elect should not be called. Ok(true) means we are done and ready for a call to elect.

Provided Methods

fn msp() -> u32

The index of the most significant page that this election provider supports.

fn lsp() -> u32

The index of the least significant page that this election provider supports.

fn desired_targets_checked() -> Result<u32, &'static str>

checked call to Self::DataProvider::desired_targets() ensuring the value never exceeds Self::MaxWinnersPerPage.

fn duration_with_export() -> Self::BlockNumber
where Self::BlockNumber: From<u32> + Add<Output = Self::BlockNumber>,

Return the duration of your election, including the export.

fn asap()

Signal the election provider that we are about to call elect asap, and it should prepare itself.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.