yet_another_parachain_runtime::frame_election_provider_support

Trait NposSolver

pub trait NposSolver {
    type AccountId: IdentifierT;
    type Accuracy: PerThing128;
    type Error: Debug + PartialEq;

    // Required methods
    fn solve(
        to_elect: usize,
        targets: Vec<Self::AccountId>,
        voters: Vec<(Self::AccountId, u64, impl Clone + IntoIterator<Item = Self::AccountId>)>,
    ) -> Result<ElectionResult<Self::AccountId, Self::Accuracy>, Self::Error>;
    fn weight<T>(voters: u32, targets: u32, vote_degree: u32) -> Weight
       where T: WeightInfo;
}

Something that can compute the result to an NPoS solution.

Required Associated Types

type AccountId: IdentifierT

The account identifier type of this solver.

type Accuracy: PerThing128

The accuracy of this solver. This will affect the accuracy of the output.

type Error: Debug + PartialEq

The error type of this implementation.

Required Methods

fn solve( to_elect: usize, targets: Vec<Self::AccountId>, voters: Vec<(Self::AccountId, u64, impl Clone + IntoIterator<Item = Self::AccountId>)>, ) -> Result<ElectionResult<Self::AccountId, Self::Accuracy>, Self::Error>

Solve an NPoS solution with the given voters, targets, and select to_elect count of targets.

fn weight<T>(voters: u32, targets: u32, vote_degree: u32) -> Weight
where T: WeightInfo,

Measure the weight used in the calculation of the solver.

voters is the number of voters.

targets is the number of targets.

vote_degree is the degree ie the maximum numbers of votes per voter.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.