yet_another_parachain_runtime

Crate frame_support

Support code for the runtime.

Note on Tuple Traits

Many of the traits defined in traits have auto-implementations on tuples as well. Usually, the tuple is a function of number of pallets in the runtime. By default, the traits are implemented for tuples of up to 64 items.

Modules

crypto

Utilities for dealing with crypto primitives. Sometimes we need to use these from inside WASM contracts, where crypto calculations have weak performance.

dispatch

Dispatch system. Contains a macro for defining runtime modules and generating values representing lazy module function calls.

dispatch_context

Provides functions to interact with the dispatch context.

dynamic_params

Experimental macros for defining dynamic params that can be used in pallet configs.

errorDeprecated

generate_genesis_config

Helper for generating the RuntimeGenesisConfig instance for presets. Helper macro allowing to construct JSON representation of partially initialized structs.

genesis_builder_helper

Helper for implementing GenesisBuilder runtime API Helper functions for implementing sp_genesis_builder::GenesisBuilder for runtimes.

inherent

instances

Some instance placeholder to be used in frame_support::pallet attribute macro.

macro_magic

Macro Magic ๐ช

migrations

pallet_macros

Contains macro stubs for all of the pallet:: macros

pallet_prelude

Prelude to be used alongside pallet macro, for ease of use.

sp_runtime

Substrate Runtime Primitives.

testing_prelude

Prelude to be used for pallet testing, for ease of use.

view_functions

Traits for querying pallet view functions.

weights

Re-exports sp-weights public API, and contains benchmarked weight constants specific to FRAME.

Macros

__create_tt_macro

Internal macro used by frame_support to create tt-call-compliant macros

__generate_dummy_part_checker

Internal macro use by frame_support to generate dummy part checker for old pallet declaration

_std_enabled

_try_runtime_and_std_not_enabled

_try_runtime_enabled

_try_runtime_or_std_enabled

assert_err

Assert an expression returns an error specified.

assert_err_ignore_postinfo

Assert an expression returns an error specified.

assert_err_with_weight

Assert an expression returns error with the given weight.

assert_error_encoded_size

Assert that the maximum encoding size does not exceed the value defined in MAX_MODULE_ERROR_ENCODED_SIZE during compilation.

assert_noop

Evaluate an expression, assert it returns an expected Err value and that runtime storage has not been mutated (i.e. expression is a no-operation).

assert_ok

Panic if an expression doesnโt evaluate to Ok.

assert_storage_noop

Evaluate any expression and assert that runtime storage has not been mutated (i.e. expression is a storage no-operation).

build_struct_json_patch

Creates a JSON patch for given struct_type, supporting recursive field initialization.

construct_runtime

Construct a runtime, with the given name and the given pallets.

crate_to_crate_version

Macro used internally in FRAME to generate the crate version for a pallet.

defensive

Generic function to mark an execution path as ONLY defensive.

defensive_assert

Trigger a defensive failure if a condition is not met.

ensure

Evaluate $x:expr and if not true return Err($y:expr).

fail

Return Err of the expression: return Err($expression);.

hypothetically

Do something hypothetically by rolling back any changes afterwards.

hypothetically_ok

Assert something to be hypothetically Ok, without actually committing it.

impl_ensure_origin_with_arg_ignoring_arg

Simple macro to explicitly implement EnsureOriginWithArg to be used on any type which implements EnsureOrigin. This is quick and dirty, so you must use the type parameters O (the origin type), T (the argument type) and AccountId (if you are using the O: .. form).

impl_filter_stack

impl_for_tuples_attr

Puts the impl_for_tuples attribute above the given code.

match_and_insert

Macro that inserts some tokens after the first match of some pattern.

match_typeDeprecated

Create a type which implements the Contains trait for a particular type with syntax similar to matches!.

match_types

Create a type which implements the Contains trait for a particular type with syntax similar to matches!.

ord_parameter_types

Macro for easily creating a new implementation of both the Get and Contains traits. Use exactly as with parameter_types, only the type must be Ord.

parameter_types

Create new implementations of the Get trait.

parameter_types_impl_thread_local

runtime_print

Print out a formatted message.

std_enabled

try_runtime_and_std_not_enabled

try_runtime_enabled

try_runtime_or_std_enabled

Structs

Blake2_128

Hash storage keys with blake2 128

Blake2_256

Hash storage keys with blake2 256

Blake2_128Concat

Hash storage keys with concat(blake2_128(key), key)

BoundedBTreeMap

A bounded map based on a B-Tree.

BoundedBTreeSet

A bounded set based on a B-Tree.

BoundedSlice

A bounded slice.

BoundedVec

A bounded vector.

Identity

Store the key directly.

PalletId

A pallet identifier. These are per pallet and should be stored in a registry somewhere.

StorageNoopGuard

Asserts that no storage changes took place between con- and destruction of Self.

Twox64Concat

Hash storage keys with concat(twox64(key), key)

Twox128

Hash storage keys with twox 128

Twox256

Hash storage keys with twox 256

WeakBoundedVec

A weakly bounded vector.

Enums

Never

A type that cannot be instantiated.

Constants

LOG_TARGET

A unified log target for support operations.

MAX_EXTRINSIC_DEPTH

Maximum nesting level for extrinsics.

MAX_MODULE_ERROR_ENCODED_SIZE

The number of bytes of the module-specific error field defined in ModuleError. In FRAME, this is the maximum encoded size of a pallet error type.

Traits

Callable

Serializable version of pallet dispatchable.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Hashable

IterableStorageDoubleMap

A strongly-typed double map in storage whose secondary keys and values can be iterated over.

IterableStorageMap

A strongly-typed map in storage whose keys and values can be iterated over.

IterableStorageNMap

A strongly-typed map with arbitrary number of keys in storage whose keys and values can be iterated over.

Parameter

A type that can be used as a parameter in a dispatchable function.

Printable

Trait for things which can be printed from the runtime.

ReversibleStorageHasher

Hasher to use to hash keys to insert to storage.

Serialize

A data structure that can be serialized into any data format supported by Serde.

StorageDoubleMap

An implementation of a map with a two keys.

StorageHasher

Hasher to use to hash keys to insert to storage.

StorageMap

A strongly-typed map in storage.

StorageNMap

An implementation of a map with an arbitrary number of keys.

StoragePrefixedMap

Trait for maps that store all its value after a unique prefix.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

print

Print something that implements Printable from the runtime.

Type Aliases

ConsensusEngineId

Consensus engine unique ID.

Attribute Macros

derive_impl

This attribute can be used to derive a full implementation of a trait based on a local partial impl and an external impl containing defaults that can be overridden in the local impl.

pallet

register_default_impl

Attach this attribute to an impl statement that you want to use with #[derive_impl(..)].

require_transactional

runtime

Construct a runtime, with the given name and the given pallets.

storage_alias

transactional

Execute the annotated function in a new storage transaction.

Derive Macros

CloneNoBound

Derive [Clone] but do not bound any generic.

DebugNoBound

Derive [Debug] but do not bound any generics.

DefaultNoBound

derive Default but do no bound any generic. Docs are at frame_support::DefaultNoBound.

Deserialize

EqNoBound

DeriveEq but do no bound any generic.

OrdNoBound

Derive [Ord] but do no bound any generic. Docs are at frame_support::OrdNoBound.

PalletError

PartialEqNoBound

Derive [PartialEq] but do not bound any generic.

PartialOrdNoBound

Derive [PartialOrd] but do not bound any generic. Docs are at frame_support::PartialOrdNoBound.

RuntimeDebugNoBound

Derive [Debug], if std is enabled it uses frame_support::DebugNoBound, if std is not enabled it just returns "<wasm:stripped>". This behaviour is useful to prevent bloating the runtime WASM blob from unneeded code.

Serialize