yet_another_parachain_runtime::frame_support

Module migrations

Structs

EnterSafeModeOnFailedMigration

Enter safe mode on a failed runtime upgrade.

FreezeChainOnFailedMigration

Do now allow any transactions to be processed after a runtime upgrade failed.

MigrationId

A generic migration identifier that can be used by MBMs.

RemovePallet

RemovePallet is a utility struct used to remove all storage items associated with a specific pallet.

RemoveStorage

RemoveStorage is a utility struct used to remove a storage item from a specific pallet.

VersionedMigration

Handles storage migration pallet versioning.

Enums

FailedMigrationHandling

How to proceed after a runtime upgrade failed.

SteppedMigrationError

Error that can occur during a SteppedMigration.

VersionedPostUpgradeData

A helper enum to wrap the pre_upgrade bytes like an Option before passing them to post_upgrade. This enum is used rather than an Option to make the API clearer to the developer.

Traits

FailedMigrationHandler

Handles a failed runtime migration.

MigrationStatusHandler

Notification handler for status updates regarding Multi-Block-Migrations.

MultiStepMigrator

Something that can do multi step migrations.

PalletVersionToStorageVersionHelper

Trait used by migrate_from_pallet_version_to_storage_version to do the actual migration.

SteppedMigration

A migration that can proceed in multiple steps.

SteppedMigrations

Multiple SteppedMigration.

StoreInCodeStorageVersion

Can store the in-code pallet version on-chain.

Functions

migrate_from_pallet_version_to_storage_version

Migrate from the PalletVersion struct to the new StorageVersion struct.