yet_another_parachain_runtime::frame_support::migrations

Trait SteppedMigrations

pub trait SteppedMigrations {
    // Required methods
    fn len() -> u32;
    fn nth_id(n: u32) -> Option<Vec<u8>>;
    fn nth_max_steps(n: u32) -> Option<Option<u32>>;
    fn nth_step(
        n: u32,
        cursor: Option<Vec<u8>>,
        meter: &mut WeightMeter,
    ) -> Option<Result<Option<Vec<u8>>, SteppedMigrationError>>;
    fn nth_transactional_step(
        n: u32,
        cursor: Option<Vec<u8>>,
        meter: &mut WeightMeter,
    ) -> Option<Result<Option<Vec<u8>>, SteppedMigrationError>>;
    fn nth_pre_upgrade(n: u32) -> Option<Result<Vec<u8>, DispatchError>>;
    fn nth_post_upgrade(
        n: u32,
        _state: Vec<u8>,
    ) -> Option<Result<(), DispatchError>>;
    fn cursor_max_encoded_len() -> usize;
    fn identifier_max_encoded_len() -> usize;

    // Provided method
    fn integrity_test() -> Result<(), &'static str> { ... }
}

Multiple SteppedMigration.

Required Methods

fn len() -> u32

The number of migrations that Self aggregates.

fn nth_id(n: u32) -> Option<Vec<u8>>

The nth SteppedMigration::id.

Is guaranteed to return Some if n < Self::len(). Calling this with any index larger or equal to Self::len() MUST return None.

fn nth_max_steps(n: u32) -> Option<Option<u32>>

The SteppedMigration::max_steps of the nth migration.

Is guaranteed to return Some if n < Self::len().

fn nth_step( n: u32, cursor: Option<Vec<u8>>, meter: &mut WeightMeter, ) -> Option<Result<Option<Vec<u8>>, SteppedMigrationError>>

Do a SteppedMigration::step on the nth migration.

Is guaranteed to return Some if n < Self::len().

fn nth_transactional_step( n: u32, cursor: Option<Vec<u8>>, meter: &mut WeightMeter, ) -> Option<Result<Option<Vec<u8>>, SteppedMigrationError>>

Do a SteppedMigration::transactional_step on the nth migration.

Is guaranteed to return Some if n < Self::len().

fn nth_pre_upgrade(n: u32) -> Option<Result<Vec<u8>, DispatchError>>

Call the pre-upgrade hooks of the nth migration.

Returns None if the index is out of bounds.

fn nth_post_upgrade( n: u32, _state: Vec<u8>, ) -> Option<Result<(), DispatchError>>

Call the post-upgrade hooks of the nth migration.

Returns None if the index is out of bounds.

fn cursor_max_encoded_len() -> usize

The maximal encoded length across all cursors.

fn identifier_max_encoded_len() -> usize

The maximal encoded length across all identifiers.

Provided Methods

fn integrity_test() -> Result<(), &'static str>

Assert the integrity of the migrations.

Should be executed as part of a test prior to runtime usage. May or may not need externalities.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.