yet_another_parachain_runtime::frame_support

Module pallet_prelude

Prelude to be used alongside pallet macro, for ease of use.

Macros

defensive

Generic function to mark an execution path as ONLY defensive.

defensive_assert

Trigger a defensive failure if a condition is not met.

ensure

Evaluate $x:expr and if not true return Err($y:expr).

Structs

Blake2_128

Hash storage keys with blake2 128

Blake2_256

Hash storage keys with blake2 256

Blake2_128Concat

Hash storage keys with concat(blake2_128(key), key)

BoundedBTreeMap

A bounded map based on a B-Tree.

BoundedBTreeSet

A bounded set based on a B-Tree.

BoundedVec

A bounded vector.

ConstU32

Const getter for a basic type.

ConstUint

Const getter for unsigned integers.

CountedStorageMap

A wrapper around a StorageMap and a StorageValue (with the value being u32) to keep track of how many items are in a map, without needing to iterate all the values.

CountedStorageNMap

A wrapper around a StorageNMap and a StorageValue (with the value being u32) to keep track of how many items are in a map, without needing to iterate all the values.

GetDefault

Implement Get by returning Default for any type that implements Default.

Identity

Store the key directly.

InherentData

Inherent data to include in a block.

MakeFatalError

Auxiliary to make any given error resolve to is_fatal_error() == true for IsFatalError.

NMapKey

A type used exclusively by storage maps as their key type.

OptionQuery

Implements QueryKindTrait with Query type being Option<_>.

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

ResultQuery

Implements QueryKindTrait with Query type being Result<Value, PalletError>.

StorageDoubleMap

A type representing a double map in storage. This structure associates a pair of keys with a value of a specified type stored on-chain.

StorageMap

A type representing a map in storage. A storage map is a mapping of keys to values of a given type stored on-chain.

StorageNMap

A type representing an NMap in storage. This structure associates an arbitrary number of keys with a value of a specified type stored on-chain.

StorageValue

A type representing a value in storage. A storage value is a single value of a given type stored on-chain.

StorageVersion

The storage version of a pallet.

Twox64Concat

Hash storage keys with concat(twox64(key), key)

Twox128

Hash storage keys with twox 128

Twox256

Hash storage keys with twox 256

ValidTransaction

Information concerning a valid transaction.

ValueQuery

Implements QueryKindTrait with Query type being Value.

WeakBoundedVec

A weakly bounded vector.

Weight

Enums

DispatchClass

A generalized group of dispatch types.

DispatchError

Reason why a dispatch call failed.

InvalidTransaction

An invalid transaction validity.

Pays

Explicit enum to denote if a transaction pays fee or not.

TransactionSource

The source of the transaction.

TransactionValidityError

Errors that can occur while checking the validity of a transaction.

UnknownTransaction

An unknown transaction validity.

Constants

MAX_MODULE_ERROR_ENCODED_SIZE

The number of bytes of the module-specific error field defined in ModuleError. In FRAME, this is the maximum encoded size of a pallet error type.

Traits

Authorize

A trait to allow calls to authorize themselves from the origin None.

BuildGenesisConfig

A trait to define the build function of a genesis config for both runtime and pallets.

CheckedAdd

Performs addition that returns None instead of wrapping around on overflow.

CheckedConversion

Convenience type to work around the highly unergonomic syntax needed to invoke the functions of overloaded generic traits, in this case TryFrom and TryInto.

CheckedDiv

Performs division that returns None instead of panicking on division by zero and instead of wrapping around on underflow and overflow.

CheckedMul

Performs multiplication that returns None instead of wrapping around on underflow or overflow.

CheckedShl

Performs a left shift that returns None on shifts larger than or equal to the type width.

CheckedShr

Performs a right shift that returns None on shifts larger than or equal to the type width.

CheckedSub

Performs subtraction that returns None instead of wrapping around on underflow.

Decode

Trait that allows zero-copy read of value-references from slices in LE format.

DecodeWithMemTracking

Marker trait used for identifying types that call the [Input::on_before_alloc_mem] hook while decoding.

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

EnsureOrigin

Some sort of check on the origin is performed by this object.

Get

A trait for querying a single value from a type.

GetStorageVersion

Provides information about a pallet’s storage versions.

Hooks

The pallet hooks trait. This is merely an umbrella trait for:

IsType

Trait to be used when types are exactly same.

MaxEncodedLen

Items implementing MaxEncodedLen have a statically known maximum encoded size.

MaybeSerializeDeserialize

A type that implements Serialize, DeserializeOwned and Debug when in std environment or serde feature is activated.

Member

A type that can be used in runtime structures.

One

Defines a multiplicative identity element for Self.

OriginTrait

Methods available on frame_system::Config::RuntimeOrigin.

PalletInfoAccess

Provides information about the pallet itself and its setup in the runtime.

Parameter

A type that can be used as a parameter in a dispatchable function.

ProvideInherent

A pallet that provides or verifies an inherent extrinsic will implement this trait.

StorageInfoTrait

A trait to give information about storage.

StorageList

A non-continuous container type.

Task

A general-purpose trait which defines a type of service work (i.e., work to performed by an off-chain worker) including methods for enumerating, validating, indexing, and running tasks of this type.

TypeInfo

Implementors return their meta type information.

TypedGet

A trait for querying a single value from a type defined in the trait.

ValidateUnsigned

Provide validation for unsigned extrinsics.

Zero

Defines an additive identity element for Self.

Type Aliases

DispatchResult

Un-augmented version of DispatchResultWithPostInfo that can be returned from dispatchable functions and is automatically converted to the augmented type. Should be used whenever the PostDispatchInfo does not need to be overwritten. As this should be the common case it is the implicit return type when none is specified.

DispatchResultWithPostInfo

The return type of a Dispatchable in frame. When returned explicitly from a dispatchable function it allows overriding the default PostDispatchInfo returned from a dispatch.

InherentIdentifier

An identifier for an inherent.

TransactionLongevity

Minimum number of blocks a transaction will remain valid for. TransactionLongevity::max_value() means “forever”.

TransactionPriority

Priority for a transaction. Additive. Higher is better.

TransactionTag

Tag for a transaction. No two transactions with the same tag should be placed on-chain.

TransactionValidity

Information on a transaction’s validity and, if valid, on how it relates to other transactions.

TransactionValidityWithRefund

Information on a transaction’s validity and, if valid, on how it relates to other transactions and some refund for the operation.

ValidateResult

Shortcut for the result value of the validate function.

Attribute Macros

authorize

Allows to authorize some general transactions with specific dispatchable functions (dispatchable functions a.k.a. calls).

call

call_index

Each dispatchable may also be annotated with the #[pallet::call_index($idx)] attribute, which explicitly defines the codec index for the dispatchable function in the Call enum.

compact

composite_enum

config

constant

disable_frame_system_supertrait_check

disable_try_decode_storage

error

event

extra_constants

feeless_if

generate_deposit

genesis_build

genesis_config

getter

hooks

import_section

include_metadata

inherent

inject_runtime_type

The optional attribute #[inject_runtime_type] can be attached to RuntimeCall, RuntimeEvent, RuntimeOrigin or PalletInfo in an impl statement that has #[register_default_impl] attached to indicate that this item is generated by construct_runtime.

no_default

no_default_bounds

origin

pallet_section

register_default_impl

Attach this attribute to an impl statement that you want to use with #[derive_impl(..)].

storage

storage_prefix

storage_version

task_condition

Allows defining conditions for a task to run.

task_index

Allows defining an index for a task.

task_list

Allows defining an iterator over available work items for a task.

task_weight

Allows defining the weight of a task.

tasks_experimental

Allows you to define some service work that can be recognized by a script or an off-chain worker.

type_value

unbounded

validate_unsigned

view_functions

weight

weight_of_authorize

Allows to define the weight of the authorize function.

whitelist_storage

Derive Macros

CloneNoBound

Derive [Clone] but do not bound any generic.

DebugNoBound

Derive [Debug] but do not bound any generics.

Decode

Derive parity_scale_codec::Decode for struct and enum.

DecodeWithMemTracking

Derive parity_scale_codec::DecodeWithMemTracking for struct and enum.

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.

EqNoBound

DeriveEq but do no bound any generic.

MaxEncodedLen

Derive parity_scale_codec::MaxEncodedLen for struct and enum.

PartialEqNoBound

Derive [PartialEq] but do not bound any generic.

RuntimeDebug

RuntimeDebugNoBound

Derive [Debug], if std is enabled it uses frame_support::DebugNoBound, if std is not enabled it just returns "<wasm:stripped>". This behaviour is useful to prevent bloating the runtime WASM blob from unneeded code.

TypeInfo