yet_another_parachain_runtime::frame_support::sp_runtime

Crate app_crypto

Traits and macros for constructing application specific strongly typed crypto wrappers.

Modules

bandersnatch

Bandersnatch VRF application crypto types.

bls381

BLS12-381 crypto applications.

ecdsa

Ecdsa crypto types.

ecdsa_bls381

ECDSA and BLS12-381 paired crypto applications.

ed25519

Ed25519 crypto types.

key_types

Known key types; this also functions as a global registry of key types for projects wishing to avoid collisions with each other.

sp_core

Shareable Substrate types.

sr25519

Sr25519 crypto types.

Macros

app_crypto

Declares Public, Pair, Signature and ProofOfPossession types which are functionally equivalent to the corresponding types defined by $module but are new application-specific types whose identifier is $key_type.

app_crypto_pair_common

Declares Pair type which is functionally equivalent to $pair, but is new application-specific type whose identifier is $key_type. It is a common part shared between full_crypto and non full_crypto environments.

with_pair

Generate the given code if the pair type is available.

wrap

Implement bidirectional From and on-way AsRef/AsMut for two types, $inner and $outer.

Structs

CryptoTypeId

An identifier for a specific cryptographic algorithm used by a key pair

KeyTypeId

An identifier for a type of cryptographic key.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Enums

DeriveError

An error when deriving a key.

DeriveJunction

A since derivation junction description. It is the single parameter used when creating a new secret key from an existing secret key and, in the case of SoftRaw and SoftIndex a new public key from an existing public key.

SecretStringError

An error with the interpretation of a secret.

Traits

AppCrypto

Application-specific cryptographic object.

AppPair

Application-specific key pair.

AppPublic

Application-specific public key.

AppSignature

Application-specific signature and Proof Of Possession

BoundToRuntimeAppPublic

Something that is bound to a fixed RuntimeAppPublic.

ByteArray

Trait used for types that are really just a fixed-length array.

CryptoType

Type which has a particular kind of crypto associated with it.

Deref

Used for immutable dereferencing operations, like *v.

Derive

Derivable key trait.

IsWrappedBy

One type is wrapped by another.

MaybeHash

Type which implements Hash in std, not when no-std (std variant).

Pair

Trait suitable for typical cryptographic PKI key pair type.

ProofOfPossessionGenerator

Pair which is able to generate proof of possession.

ProofOfPossessionVerifier

Pair which is able to verify proof of possession.

Public

Trait suitable for cryptographic public keys.

RuntimeAppPublic

Runtime interface for an application’s public key.

RuntimePublic

Runtime interface for a public key.

Signature

Trait suitable for cryptographic signatures.

Ss58Codec

Key that can be encoded to/from SS58.

UncheckedFrom

Similar to From, except that the onus is on the part of the caller to ensure that data passed in makes sense. Basically, you’re not guaranteed to get anything sensible out.

Wraps

Opposite of IsWrappedBy - denotes a type which is a simple wrapper around another type.

Derive Macros

RuntimeDebug