yet_another_parachain_runtime::frame_support::sp_runtime::app_crypto::ed25519

Struct CryptoBytes

#[repr(transparent)]
pub struct CryptoBytes<const N: usize, T = ()>(pub [u8; N], _);

Generic byte array holding some crypto-related raw data.

The type is generic over a constant length N and a โtagโ T which can be used to specialize the byte array without requiring newtypes.

The tag T is held in a PhantomData<fn() ->T>, a trick allowing CryptoBytes to be Send and Sync regardless of T properties (ref).

Tuple Fields
0: [u8; N]
Implementations

impl<const N: usize, T> CryptoBytes<N, T>

pub fn from_raw(inner: [u8; N]) -> CryptoBytes<N, T>

Construct from raw array.

pub fn to_raw(self) -> [u8; N]

Construct from raw array.

pub fn as_array_ref(&self) -> &[u8; N]

Return a slice filled with raw data.

impl<T> CryptoBytes<32, T>

pub fn from_h256(x: H256) -> CryptoBytes<32, T>

A new instance from an H256.

impl<T> CryptoBytes<64, T>

pub fn from_h512(x: H512) -> CryptoBytes<64, T>

A new instance from an H512.

impl<TAG> CryptoBytes<sp_core::::ecdsa::GenericPublic::{constant#0}, (PublicTag, TAG)>

pub fn from_full( full: &[u8], ) -> Result<CryptoBytes<sp_core::::ecdsa::GenericPublic::{constant#0}, (PublicTag, TAG)>, ()>

Create a new instance from the given full public key.

This will convert the full public key into the compressed format.

impl<PUBLIC> CryptoBytes<sp_core::::ecdsa::GenericSignature::{constant#0}, (SignatureTag, PUBLIC)>
where PUBLIC: From<PublicKey>,

pub fn recover_prehashed(&self, message: &[u8; 32]) -> Option<PUBLIC>

Recover the public key from this signature and a pre-hashed message.

impl CryptoBytes<sp_core::::ecdsa::GenericSignature::{constant#0}, (SignatureTag, CryptoBytes<sp_core::::ecdsa::GenericPublic::{constant#0}, (PublicTag, EcdsaTag)>)>

pub fn recover<M>( &self, message: M, ) -> Option<CryptoBytes<sp_core::::ecdsa::GenericPublic::{constant#0}, (PublicTag, EcdsaTag)>>
where M: AsRef<[u8]>,

Recover the public key from this signature and a message.

impl CryptoBytes<sp_core::::ecdsa::GenericSignature::{constant#0}, (SignatureTag, CryptoBytes<sp_core::::ecdsa::GenericPublic::{constant#0}, (PublicTag, EcdsaKeccakTag)>)>

pub fn recover<M>( &self, message: M, ) -> Option<CryptoBytes<sp_core::::ecdsa::GenericPublic::{constant#0}, (PublicTag, EcdsaKeccakTag)>>
where M: AsRef<[u8]>,

Recover the public key from this signature and a message.