yet_another_parachain_runtime::frame_support::sp_runtime::app_crypto

Crate sp_core

Shareable Substrate types.

Modules

bandersnatch

VRFs backed by Bandersnatch, an elliptic curve built over BLS12-381 scalar field.

bls

BLS (Boneh–Lynn–Shacham) Signature along with efficiently verifiable Chaum-Pedersen proof API. Signatures are implemented according to Efficient Aggregatable BLS Signatures with Chaum-Pedersen Proofs Hash-to-BLS-curve is using Simplified SWU for AB == 0 RFC 9380 Sect 6.6.3. Chaum-Pedersen proof uses the same hash-to-field specified in RFC 9380 for the field of the BLS curve.

bounded

Collection types that have an upper limit on how many elements that they can contain, and supporting traits that aid in defining the limit.

const_hex2array

Provides a const function for converting a hex string to a u8 array at compile time, when used in the proper context.

crypto

Cryptographic utilities.

crypto_bytes

Generic byte array which can be specialized with a marker type.

ecdsa

Simple ECDSA secp256k1 API.

ed25519

Simple Ed25519 API.

hash

A fixed hash type.

hashing

Hashing Functions.

hexdisplay

Wrapper type for byte collections that outputs hex.

offchain

Offchain workers types

paired_crypto

API for using a pair of crypto schemes together.

proof_of_possession

Utilities for proving possession of a particular public key

storage

Primitive types for storage related stuff.

testing

Types that should only be used for testing!

traits

Shareable Substrate traits.

uint

An unsigned fixed-size integer.

Macros

bounded_btree_map

Build a bounded btree-map from the given literals.

bounded_vec

Build a bounded vec from the given literals.

defer

Executes the given code when the current scope is dropped.

generate_feature_enabled_macro

Generates a macro for checking if a certain feature is enabled.

hex2array

Provides a const array from given string literal.

impl_maybe_marker

Macro for creating Maybe* marker traits.

impl_maybe_marker_std_or_serde

Macro for creating Maybe* marker traits.

map

Initialize a key-value collection from array.

parameter_types

Create new implementations of the Get trait.

wasm_export_functions

Macro for exporting functions from wasm in with the expected signature for using it with the wasm executor. This is useful for tests where you need to call a function in wasm.

Structs

Blake2Hasher

Concrete implementation of Hasher using Blake2b 256-bit hashes

Bytes

Hex-serialized shim for Vec<u8>.

ConstBool

Const getter for a basic type.

ConstI8

Const getter for a basic type.

ConstI16

Const getter for a basic type.

ConstI32

Const getter for a basic type.

ConstI64

Const getter for a basic type.

ConstI128

Const getter for a basic type.

ConstInt

Const getter for signed integers.

ConstU8

Const getter for a basic type.

ConstU16

Const getter for a basic type.

ConstU32

Const getter for a basic type.

ConstU64

Const getter for a basic type.

ConstU128

Const getter for a basic type.

ConstUint

Const getter for unsigned integers.

GetDefault

Implement Get by returning Default for any type that implements Default.

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H512

Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

KeccakHasher

Concrete implementation of Hasher using Keccak 256-bit hashes

OpaqueMetadata

Stores the encoded RuntimeMetadata for the native side as opaque type.

OpaquePeerId

Simple blob to hold a PeerId without committing to its format.

U256

Little-endian large integer type 256-bit unsigned integer.

U512

Little-endian large integer type 512-bits unsigned integer.

Enums

DeriveJunction

A since derivation junction description. It is the single parameter used when creating a new secret key from an existing secret key and, in the case of SoftRaw and SoftIndex a new public key from an existing public key.

LogLevelFilter

Log level filter that expresses which log levels should be filtered.

RuntimeInterfaceLogLevel

A log level matching the one from log crate.

Void

The void type - it cannot exist.

Constants

MAX_POSSIBLE_ALLOCATION

The maximum number of bytes that can be allocated at one time.

Traits

ByteArray

Trait used for types that are really just a fixed-length array.

Decode

Trait that allows zero-copy read of value-references from slices in LE format.

DecodeWithMemTracking

Marker trait used for identifying types that call the [Input::on_before_alloc_mem] hook while decoding.

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

Get

A trait for querying a single value from a type.

Hasher

Trait describing an object that can hash a slice of bytes. Used to abstract other types over the hashing algorithm. Defines a single hash method and an Out associated type with the necessary bounds.

MaxEncodedLen

Items implementing MaxEncodedLen have a statically known maximum encoded size.

Pair

Trait suitable for typical cryptographic PKI key pair type.

Public

Trait suitable for cryptographic public keys.

TryCollect

Try and collect into a collection C.

TypeId

Provide a simple 4 byte identifier for a type.

TypedGet

A trait for querying a single value from a type defined in the trait.

Functions

blake2_64

Do a Blake2 64-bit hash and return result.

blake2_128

Do a Blake2 128-bit hash and return result.

blake2_256

Do a Blake2 256-bit hash and return result.

blake2_512

Do a Blake2 512-bit hash and return result.

blake2_512_into

Do a Blake2 512-bit hash and place result in dest.

convert_hash

Hash conversion. Used to convert between unbound associated hash types in traits, implemented by the same hash type. Panics if used to convert between different hash types.

keccak_256

Do a keccak 256-bit hash and return result.

keccak_512

Do a keccak 512-bit hash and return result.

sha2_256

Do a sha2 256-bit hash and return result.

twox_64

Do a XX 64-bit hash and return result.

twox_64_into

Do a XX 64-bit hash and place result in dest.

twox_128

Do a XX 128-bit hash and return result.

twox_256

Do a XX 256-bit hash and return result.

twox_128_into

Do a XX 128-bit hash and place result in dest.

twox_256_into

Do a XX 256-bit hash and place result in dest.

Derive Macros

Decode

Derive parity_scale_codec::Decode for struct and enum.

DecodeWithMemTracking

Derive parity_scale_codec::DecodeWithMemTracking for struct and enum.

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.

MaxEncodedLen

Derive parity_scale_codec::MaxEncodedLen for struct and enum.

RuntimeDebug