yet_another_parachain_runtime::frame_support::sp_runtime::app_crypto::sp_core::crypto

Trait FromEntropy

pub trait FromEntropy: Sized {
    // Required method
    fn from_entropy(input: &mut impl Input) -> Result<Self, Error>;
}

Create random values of Self given a stream of entropy.

Required Methods

fn from_entropy(input: &mut impl Input) -> Result<Self, Error>

Create a random value of Self given a stream of random bytes on input. May only fail if input has an error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.