yet_another_parachain_runtime::frame_support::sp_runtime::app_crypto::sp_core::crypto

Trait Wraps

pub trait Wraps: Sized {
    type Inner: IsWrappedBy<Self>;

    // Provided method
    fn as_inner_ref(&self) -> &Self::Inner { ... }
}

Opposite of IsWrappedBy - denotes a type which is a simple wrapper around another type.

Required Associated Types

type Inner: IsWrappedBy<Self>

The inner type it is wrapping.

Provided Methods

fn as_inner_ref(&self) -> &Self::Inner

Get a reference to the inner type that is wrapped.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.