yet_another_parachain_runtime::frame_support::sp_runtime::app_crypto::sp_core::crypto

Trait Zeroize

pub trait Zeroize {
    // Required method
    fn zeroize(&mut self);
}

Trait for securely erasing values from memory.

Required Methods

fn zeroize(&mut self)

Zero out this object from memory using Rust intrinsics which ensure the zeroization operation is not “optimized away” by the compiler.