yet_another_parachain_runtime::frame_support::sp_runtime::app_crypto::sp_core

Trait TypedGet

pub trait TypedGet {
    type Type;

    // Required method
    fn get() -> Self::Type;
}

A trait for querying a single value from a type defined in the trait.

It is not required that the value is constant.

Required Associated Types

type Type

The type which is returned.

Required Methods

fn get() -> Self::Type

Return the current value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.