yet_another_parachain_runtime::frame_support::sp_runtime::app_crypto::sp_core::traits

Trait SpawnEssentialNamed

pub trait SpawnEssentialNamed:
    DynClone
    + Send
    + Sync {
    // Required methods
    fn spawn_essential_blocking(
        &self,
        name: &'static str,
        group: Option<&'static str>,
        future: Pin<Box<dyn Future<Output = ()> + Send>>,
    );
    fn spawn_essential(
        &self,
        name: &'static str,
        group: Option<&'static str>,
        future: Pin<Box<dyn Future<Output = ()> + Send>>,
    );
}

Something that can spawn essential tasks (blocking and non-blocking) with an assigned name and optional group.

Essential tasks are special tasks that should take down the node when they end.

Required Methods

fn spawn_essential_blocking( &self, name: &'static str, group: Option<&'static str>, future: Pin<Box<dyn Future<Output = ()> + Send>>, )

Spawn the given blocking future.

The given group and name is used to identify the future in tracing.

fn spawn_essential( &self, name: &'static str, group: Option<&'static str>, future: Pin<Box<dyn Future<Output = ()> + Send>>, )

Spawn the given non-blocking future.

The given group and name is used to identify the future in tracing.