yet_another_parachain_runtime::frame_support::sp_runtime::app_crypto

Trait ByteArray

pub trait ByteArray:
    AsRef<[u8]>
    + AsMut<[u8]>
    + for<'a> TryFrom<&'a [u8], Error = ()> {
    const LEN: usize;

    // Provided methods
    fn from_slice(data: &[u8]) -> Result<Self, ()> { ... }
    fn to_raw_vec(&self) -> Vec<u8>  { ... }
    fn as_slice(&self) -> &[u8]  { ... }
}

Trait used for types that are really just a fixed-length array.

Required Associated Constants

const LEN: usize

The “length” of the values of this type, which is always the same.

Provided Methods

fn from_slice(data: &[u8]) -> Result<Self, ()>

A new instance from the given slice that should be Self::LEN bytes long.

fn to_raw_vec(&self) -> Vec<u8>

Return a Vec<u8> filled with raw data.

fn as_slice(&self) -> &[u8]

Return a slice filled with raw data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.