yet_another_parachain_runtime::frame_support::sp_runtime::app_crypto

Trait RuntimePublic

pub trait RuntimePublic: Sized {
    type Signature: Debug + Eq + PartialEq + Clone;
    type ProofOfPossession: Debug + Eq + PartialEq + Clone;

    // Required methods
    fn all(key_type: KeyTypeId) -> Vec<Self>;
    fn generate_pair(key_type: KeyTypeId, seed: Option<Vec<u8>>) -> Self;
    fn sign<M>(&self, key_type: KeyTypeId, msg: &M) -> Option<Self::Signature>
       where M: AsRef<[u8]>;
    fn verify<M>(&self, msg: &M, signature: &Self::Signature) -> bool
       where M: AsRef<[u8]>;
    fn generate_proof_of_possession(
        &mut self,
        key_type: KeyTypeId,
        owner: &[u8],
    ) -> Option<Self::ProofOfPossession>;
    fn verify_proof_of_possession(
        &self,
        owner: &[u8],
        pop: &Self::ProofOfPossession,
    ) -> bool;
    fn to_raw_vec(&self) -> Vec<u8> ;
}

Runtime interface for a public key.

Required Associated Types

type Signature: Debug + Eq + PartialEq + Clone

The signature that will be generated when signing with the corresponding private key.

type ProofOfPossession: Debug + Eq + PartialEq + Clone

The Proof Of Possession the corresponding private key.

Required Methods

fn all(key_type: KeyTypeId) -> Vec<Self>

Returns all public keys for the given key type in the keystore.

fn generate_pair(key_type: KeyTypeId, seed: Option<Vec<u8>>) -> Self

Generate a public/private pair for the given key type with an optional seed and store it in the keystore.

The seed needs to be valid utf8.

Returns the generated public key.

fn sign<M>(&self, key_type: KeyTypeId, msg: &M) -> Option<Self::Signature>
where M: AsRef<[u8]>,

Sign the given message with the corresponding private key of this public key.

The private key will be requested from the keystore using the given key type.

Returns the signature or None if the private key could not be found or some other error occurred.

fn verify<M>(&self, msg: &M, signature: &Self::Signature) -> bool
where M: AsRef<[u8]>,

Verify that the given signature matches the given message using this public key.

fn generate_proof_of_possession( &mut self, key_type: KeyTypeId, owner: &[u8], ) -> Option<Self::ProofOfPossession>

Generates the necessary proof(s) usually as a signature or list of signatures, for the corresponding public key to be accepted as legitimate by the network.

This includes attestation of the owner of the public key in the form of signing the owner’s identity. It might also includes other signatures such as signature obtained by signing the corresponding public in different context than normal signatures in case of BLS key pair.

The owner is an arbitrary byte array representing the identity of the owner of the key which has been signed by the private key in process of generating the proof.

The private key will be requested from the keystore using the given key type.

Returns the proof of possession or None if it failed or is not able to do so.

fn verify_proof_of_possession( &self, owner: &[u8], pop: &Self::ProofOfPossession, ) -> bool

Verifies that the given proof is valid for the corresponding public key. The proof is usually a signature or list of signatures, for the corresponding public key to be accepted by the network. It include attestation of the owner of the public key in the form signing the owner’s identity but might also includes other signatures.

The owner is an arbitrary byte array representing the identity of the owner of the key which has been signed by the private key in process of generating the proof.

Returns true if the proof is deemed correct by the cryto type.

fn to_raw_vec(&self) -> Vec<u8>

Returns Self as raw vec.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.