yet_another_parachain_runtime::frame_system::pallet

Type Alias Origin

pub type Origin<T> = RawOrigin<<T as Config>::AccountId>;

Exposed trait-generic origin type.

Aliased Type

enum Origin<T> {
    Root,
    Signed(<T as Config>::AccountId),
    None,
    Authorized,
}

Variants

Root

The system itself ordained this dispatch to happen: this is the highest privilege level.

Signed(<T as Config>::AccountId)

It is signed by some public key and we provide the AccountId.

None

It is signed by nobody, can be either:

included and agreed upon by the validators anyway,

or unsigned transaction validated by a pallet.

Authorized

It is signed by nobody, the extrinsic is authorized by the runtime.

Authorization logic is defined by pallets. See trait Authorize and attribute macro authorize for more details.