yet_another_parachain_runtime

Crate pallet_alliance

Alliance Pallet

The Alliance Pallet provides a collective that curates a list of accounts and URLs, deemed by the voting members to be unscrupulous actors. The Alliance

provides a set of ethics against bad behavior, and

provides recognition and influence for those teams that contribute something back to the ecosystem.

Overview

The network initializes the Alliance via a Root call. After that, anyone with an approved identity and website can join as an Ally. The MembershipManager origin can elevate Allies to Fellows, giving them voting rights within the Alliance.

Voting members of the Alliance maintain a list of accounts and websites. Members can also vote to update the Alliance’s rule and make announcements.

Terminology

Rule: The IPFS CID (hash) of the Alliance rules for the community to read and the Alliance members to enforce. Similar to a Charter or Code of Conduct.

Announcement: An IPFS CID of some content that the Alliance want to announce.

Member: An account that is already in the group of the Alliance, including two types: Fellow, or Ally. A member can also be kicked by the MembershipManager origin or retire by itself.

Fellow: An account who is elevated from Ally by other Fellows.

Ally: An account who would like to join the Alliance. To become a voting member (Fellow), it will need approval from the MembershipManager origin. Any account can join as an Ally either by placing a deposit or by nomination from a voting member.

Unscrupulous List: A list of bad websites and addresses; items can be added or removed by voting members.

Interface

Dispatchable Functions

For General Users

join_alliance - Join the Alliance as an Ally. This requires a slashable deposit.

For Members (All)

give_retirement_notice - Give a retirement notice and start a retirement period required to pass in order to retire.

retire - Retire from the Alliance and release the caller’s deposit.

For Voting Members

propose - Propose a motion.

vote - Vote on a motion.

close - Close a motion with enough votes or that has expired.

set_rule - Initialize or update the Alliance’s rule by IPFS CID.

announce - Make announcement by IPFS CID.

nominate_ally - Nominate a non-member to become an Ally, without deposit.

elevate_ally - Approve an ally to become a Fellow.

kick_member - Kick a member and slash its deposit.

add_unscrupulous_items - Add some items, either accounts or websites, to the list of unscrupulous items.

remove_unscrupulous_items - Remove some items from the list of unscrupulous items.

abdicate_fellow_status - Abdicate one’s voting rights, demoting themself to Ally.

Root Calls

init_members - Initialize the Alliance, onboard fellows and allies.

disband - Disband the Alliance, remove all active members and unreserve deposits.

Modules

migration

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_alliance

Structs

Cid

Representation of a CID.

DisbandWitness

Witness data for the disband call.

GenesisConfig

Can be used to configure the genesis state of this pallet.

Instance1

Instance1 to be used for instantiable pallets defined with the #[pallet] macro. Instances 2-16 are also available but are hidden from docs.

Multihash

A Multihash instance that only supports the basic functionality and no hashing.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

SubstrateWeight

Weights for pallet_alliance using the Substrate node and recommended hardware.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

MemberRole

The various roles that a member can hold.

UnscrupulousItem

The type of item that may be deemed unscrupulous.

Version

The version of the CID.

Constants

LOG_TARGET

The log target of this pallet.

Traits

Config

Configuration trait of this pallet.

IdentityVerifier

Interface required for identity verification.

ProposalProvider

The provider of a collective action interface, for example an instance of pallet-collective.

WeightInfo

Weight functions needed for pallet_alliance.

Type Aliases

Announcements

The current IPFS CIDs of any announcements.

DepositOf

Maps members to their candidacy deposit.

Members

Maps member type to members of each type.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

ProposalIndex

Simple index type for proposal counting.

RetiringMembers

A set of members who gave a retirement notice. They can retire after the end of retirement period stored as a future block number.

Rule

The IPFS CID of the alliance rule. Fellows can propose a new rule with a super-majority.

UnscrupulousAccounts

The current list of accounts deemed unscrupulous. These accounts non grata cannot submit candidacy.

UnscrupulousWebsites

The current list of websites deemed unscrupulous.