yet_another_parachain_runtime::pallet_alliance

Trait Config

pub trait Config<I = ()>: Config + Config
where
    Self::RuntimeEvent: From<Event<Self, I>>,
    I: 'static,
{

    type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>;
    type Proposal: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin, PostInfo = PostDispatchInfo> + From<Call<Self>> + From<Call<Self, I>> + GetDispatchInfo + IsSubType<Call<Self, I>> + IsType<Self::RuntimeCall>;
    type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type MembershipManager: EnsureOrigin<Self::RuntimeOrigin>;
    type AnnouncementOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type Slashed: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type InitializeMembers: InitializeMembers<Self::AccountId>;
    type MembershipChanged: ChangeMembers<Self::AccountId>;
    type IdentityVerifier: IdentityVerifier<Self::AccountId>;
    type ProposalProvider: ProposalProvider<Self::AccountId, Self::Hash, Self::Proposal>;
    type MaxProposals: Get<u32>;
    type MaxFellows: Get<u32>;
    type MaxAllies: Get<u32>;
    type MaxUnscrupulousItems: Get<u32>;
    type MaxWebsiteUrlLength: Get<u32>;
    type AllyDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaxAnnouncementsCount: Get<u32>;
    type MaxMembersCount: Get<u32>;
    type WeightInfo: WeightInfo;
    type RetirementPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>

The overarching event type.

type Proposal: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin, PostInfo = PostDispatchInfo> + From<Call<Self>> + From<Call<Self, I>> + GetDispatchInfo + IsSubType<Call<Self, I>> + IsType<Self::RuntimeCall>

The runtime call dispatch type.

type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin for admin-level operations, like setting the Alliance’s rules.

type MembershipManager: EnsureOrigin<Self::RuntimeOrigin>

Origin that manages entry and forcible discharge from the Alliance.

type AnnouncementOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin for making announcements and adding/removing unscrupulous items.

type Currency: ReservableCurrency<Self::AccountId>

The currency used for deposits.

type Slashed: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

What to do with slashed funds.

type InitializeMembers: InitializeMembers<Self::AccountId>

What to do with initial voting members of the Alliance.

type MembershipChanged: ChangeMembers<Self::AccountId>

What to do when a member has been added or removed.

type IdentityVerifier: IdentityVerifier<Self::AccountId>

The identity verifier of an Alliance member.

type ProposalProvider: ProposalProvider<Self::AccountId, Self::Hash, Self::Proposal>

The provider of the proposal operation.

type MaxProposals: Get<u32>

Maximum number of proposals allowed to be active in parallel.

type MaxFellows: Get<u32>

The maximum number of Fellows supported by the pallet. Used for weight estimation.

NOTE:

Benchmarks will need to be re-run and weights adjusted if this changes.

This pallet assumes that dependencies keep to the limit without enforcing it.

type MaxAllies: Get<u32>

The maximum number of Allies supported by the pallet. Used for weight estimation.

NOTE:

Benchmarks will need to be re-run and weights adjusted if this changes.

This pallet assumes that dependencies keep to the limit without enforcing it.

type MaxUnscrupulousItems: Get<u32>

The maximum number of the unscrupulous items supported by the pallet.

type MaxWebsiteUrlLength: Get<u32>

The maximum length of a website URL.

type AllyDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The deposit required for submitting candidacy.

type MaxAnnouncementsCount: Get<u32>

The maximum number of announcements.

type MaxMembersCount: Get<u32>

The maximum number of members per member role.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type RetirementPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The number of blocks a member must wait between giving a retirement notice and retiring. Supposed to be greater than time required to kick_member.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.