yet_another_parachain_runtime::pallet_asset_tx_payment

Trait HandleCredit

pub trait HandleCredit<AccountId, B>
where
    B: Balanced<AccountId>,
{
    // Required method
    fn handle_credit(
        credit: Imbalance<<B as Inspect<AccountId>>::AssetId, <B as Inspect<AccountId>>::Balance, <B as Balanced<AccountId>>::OnDropCredit, <B as Balanced<AccountId>>::OnDropDebt>,
    );
}

Allows specifying what to do with the withdrawn asset fees.

Required Methods

fn handle_credit( credit: Imbalance<<B as Inspect<AccountId>>::AssetId, <B as Inspect<AccountId>>::Balance, <B as Balanced<AccountId>>::OnDropCredit, <B as Balanced<AccountId>>::OnDropDebt>, )

Implement to determine what to do with the withdrawn asset fees. Default for CreditOf from the assets pallet is to burn and decrease total issuance.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.