yet_another_parachain_runtime::pallet_bounties

Trait Config

pub trait Config<I = ()>:
    Config
    + Config<I>
    + Config
where
    Self::RuntimeEvent: From<Event<Self, I>>,
    I: 'static,
{

    type BountyDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type BountyDepositPayoutDelay: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type BountyUpdatePeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type CuratorDepositMultiplier: Get<Permill>;
    type CuratorDepositMax: Get<Option<<Self::Currency as Currency<Self::AccountId>>::Balance>>;
    type CuratorDepositMin: Get<Option<<Self::Currency as Currency<Self::AccountId>>::Balance>>;
    type BountyValueMinimum: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type DataDepositPerByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>;
    type MaximumReasonLength: Get<u32>;
    type WeightInfo: WeightInfo;
    type ChildBountyManager: ChildBountyManager<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type OnSlash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type BountyDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit for placing a bounty proposal.

type BountyDepositPayoutDelay: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

The delay period for which a bounty beneficiary need to wait before claim the payout.

type BountyUpdatePeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

The time limit for a curator to act before a bounty expires.

The period that starts when a curator is approved, during which they must execute or update the bounty via extend_bounty_expiry. If missed, the bounty expires, and the curator may be slashed. If BlockNumberFor::MAX, bounties stay active indefinitely, removing the need for extend_bounty_expiry.

type CuratorDepositMultiplier: Get<Permill>

The curator deposit is calculated as a percentage of the curator fee.

This deposit has optional upper and lower bounds with CuratorDepositMax and CuratorDepositMin.

type CuratorDepositMax: Get<Option<<Self::Currency as Currency<Self::AccountId>>::Balance>>

Maximum amount of funds that should be placed in a deposit for making a proposal.

type CuratorDepositMin: Get<Option<<Self::Currency as Currency<Self::AccountId>>::Balance>>

Minimum amount of funds that should be placed in a deposit for making a proposal.

type BountyValueMinimum: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

Minimum value for a bounty.

type DataDepositPerByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit per byte within the tip report reason or bounty description.

type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>

The overarching event type.

type MaximumReasonLength: Get<u32>

Maximum acceptable reason length.

Benchmarks depend on this value, be sure to update weights file when changing this value

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type ChildBountyManager: ChildBountyManager<<Self::Currency as Currency<Self::AccountId>>::Balance>

The child bounty manager.

type OnSlash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for the unbalanced decrease when slashing for a rejected bounty.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.