yet_another_parachain_runtime::pallet_bridge_messages

Trait CallSubType

pub trait CallSubType<T, I>: IsSubType<<Pallet<T, I> as Callable<T>>::RuntimeCall>
where
    T: Config<I, RuntimeCall = Self>,
    I: 'static,
{
    // Required methods
    fn receive_messages_proof_info(
        &self,
    ) -> Option<ReceiveMessagesProofInfo<<T as Config<I>>::LaneId>>;
    fn receive_messages_delivery_proof_info(
        &self,
    ) -> Option<ReceiveMessagesDeliveryProofInfo<<T as Config<I>>::LaneId>>;
    fn call_info(&self) -> Option<MessagesCallInfo<<T as Config<I>>::LaneId>>;
    fn call_info_for(
        &self,
        lane_id: <T as Config<I>>::LaneId,
    ) -> Option<MessagesCallInfo<<T as Config<I>>::LaneId>>;
    fn check_obsolete_call(
        &self,
    ) -> Result<ValidTransaction, TransactionValidityError>;
}

Trait representing a call that is a sub type of pallet_bridge_messages::Call.

Required Methods

fn receive_messages_proof_info( &self, ) -> Option<ReceiveMessagesProofInfo<<T as Config<I>>::LaneId>>

Create a new instance of ReceiveMessagesProofInfo from a ReceiveMessagesProof call.

fn receive_messages_delivery_proof_info( &self, ) -> Option<ReceiveMessagesDeliveryProofInfo<<T as Config<I>>::LaneId>>

Create a new instance of ReceiveMessagesDeliveryProofInfo from a ReceiveMessagesDeliveryProof call.

fn call_info(&self) -> Option<MessagesCallInfo<<T as Config<I>>::LaneId>>

Create a new instance of MessagesCallInfo from a ReceiveMessagesProof or a ReceiveMessagesDeliveryProof call.

fn call_info_for( &self, lane_id: <T as Config<I>>::LaneId, ) -> Option<MessagesCallInfo<<T as Config<I>>::LaneId>>

Create a new instance of MessagesCallInfo from a ReceiveMessagesProof or a ReceiveMessagesDeliveryProof call, if the call is for the provided lane.

fn check_obsolete_call( &self, ) -> Result<ValidTransaction, TransactionValidityError>

Ensures that a ReceiveMessagesProof or a ReceiveMessagesDeliveryProof call:

does not deliver already delivered messages. We require all messages in the ReceiveMessagesProof call to be undelivered;

does not submit empty ReceiveMessagesProof call with zero messages, unless the lane needs to be unblocked by providing relayer rewards proof;

brings no new delivery confirmations in a ReceiveMessagesDeliveryProof call. We require at least one new delivery confirmation in the unrewarded relayers set;

does not violate some basic (easy verifiable) messages pallet rules obsolete (like submitting a call when a pallet is halted or delivering messages when a dispatcher is inactive).

If one of above rules is violated, the transaction is treated as invalid.