yet_another_parachain_runtime

Crate pallet_bridge_parachains

Parachains finality module.

This module needs to be deployed with GRANDPA module, which is syncing relay chain blocks. The main entry point of this module is submit_parachain_heads, which accepts storage proof of some parachain Heads entries from bridged relay chain. It requires corresponding relay headers to be already synced.

Modules

benchmarking

Parachains finality pallet benchmarking.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_bridge_parachains

weights_ext

Weight-related utilities.

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Instance1

Instance1 to be used for instantiable pallets defined with the #[pallet] macro. Instances 2-16 are also available but are hidden from docs.

MaybeMaxParachains

Returns maximal number of parachains, supported by the pallet.

MaybeMaxTotalParachainHashes

Returns total number of all parachains hashes/heads, stored by the pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

ParachainHeaders

Single parachain header chain adapter.

SubmitParachainHeadsHelper

Helper struct that provides methods for working with the SubmitParachainHeads call.

VerifiedSubmitParachainHeadsInfo

Verified SubmitParachainHeadsInfo.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Constants

LOG_TARGET

The target that will be used when publishing logs related to this pallet.

Traits

BoundedBridgeGrandpaConfig

Convenience trait for defining BridgedChain bounds.

CallSubType

Trait representing a call that is a sub type of this pallet’s call.

Config

Configuration trait of this pallet.

WeightInfo

Weight functions needed for pallet_bridge_parachains.

WeightInfoExt

Extended weight info.

Functions

initialize_for_benchmarks

(Re)initialize pallet with given header for using it in pallet-bridge-messages benchmarks.

Type Aliases

GrandpaPalletOf

Bridge GRANDPA pallet that is used to verify parachain proofs.

ImportedParaHeads

State roots of parachain heads which have been imported into the pallet.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PalletOperatingMode

The current operating mode of the pallet.

PalletOwner

Optional pallet owner.

ParasInfo

Parachains info.

StoredParaHeadDataOf

Stored parachain head data of given parachains pallet.

WeightInfoOf

Weight info of the given parachains pallet.