yet_another_parachain_runtime::pallet_broker

Type Alias SaleInfoRecordOf

pub type SaleInfoRecordOf<T> = SaleInfoRecord<<<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance, <<<T as Config>::Coretime as CoretimeInterface>::RelayChainBlockNumberProvider as BlockNumberProvider>::BlockNumber>;

Aliased Type

struct SaleInfoRecordOf<T> {
    pub sale_start: <<<T as Config>::Coretime as CoretimeInterface>::RelayChainBlockNumberProvider as BlockNumberProvider>::BlockNumber,
    pub leadin_length: <<<T as Config>::Coretime as CoretimeInterface>::RelayChainBlockNumberProvider as BlockNumberProvider>::BlockNumber,
    pub end_price: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    pub region_begin: u32,
    pub region_end: u32,
    pub ideal_cores_sold: u16,
    pub cores_offered: u16,
    pub first_core: u16,
    pub sellout_price: Option<<<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance>,
    pub cores_sold: u16,
}

Fields
sale_start: <<<T as Config>::Coretime as CoretimeInterface>::RelayChainBlockNumberProvider as BlockNumberProvider>::BlockNumber

The relay block number at which the sale will/did start.

leadin_length: <<<T as Config>::Coretime as CoretimeInterface>::RelayChainBlockNumberProvider as BlockNumberProvider>::BlockNumber

The length in blocks of the Leadin Period (where the price is decreasing).

end_price: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance

The price of Bulk Coretime after the Leadin Period.

region_begin: u32

The first timeslice of the Regions which are being sold in this sale.

region_end: u32

The timeslice on which the Regions which are being sold in the sale terminate. (i.e. One after the last timeslice which the Regions control.)

ideal_cores_sold: u16

The number of cores we want to sell, ideally. Selling this amount would result in no change to the price for the next sale.

cores_offered: u16

Number of cores which are/have been offered for sale.

first_core: u16

The index of the first core which is for sale. Core of Regions which are sold have incrementing indices from this.

sellout_price: Option<<<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance>

The price at which cores have been sold out.

Will only be None if no core was offered for sale.

cores_sold: u16

Number of cores which have been sold; never more than cores_offered.