yet_another_parachain_runtime

Crate pallet_child_bounties

Child Bounties Pallet ( pallet-child-bounties )

Child Bounty

NOTE: This pallet is tightly coupled with pallet-treasury and pallet-bounties.

With child bounties, a large bounty proposal can be divided into smaller chunks, for parallel execution, and for efficient governance and tracking of spent funds. A child bounty is a smaller piece of work, extracted from a parent bounty. A curator is assigned after the child bounty is created by the parent bounty curator, to be delegated with the responsibility of assigning a payout address once the specified set of tasks is completed.

Interface

Dispatchable Functions

Child Bounty protocol:

add_child_bounty - Add a child bounty for a parent bounty to for dividing the work in smaller tasks.

propose_curator - Assign an account to a child bounty as candidate curator.

accept_curator - Accept a child bounty assignment from the parent bounty curator, setting a curator deposit.

award_child_bounty - Close and pay out the specified amount for the completed work.

claim_child_bounty - Claim a specific child bounty amount from the payout address.

unassign_curator - Unassign an accepted curator from a specific child bounty.

close_child_bounty - Cancel the child bounty for a specific treasury amount and close the bounty.

Modules

migration

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_child_bounties

Structs

ChildBounty

A child bounty proposal.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

ChildBountyStatus

The status of a child-bounty.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

WeightInfo

Weight functions needed for pallet_child_bounties.

Type Aliases

AccountIdLookupOf

BalanceOf

BlockNumberFor

BountiesError

BountyIndex

ChildBounties

Child bounties that have been added.

ChildBountyCount

DEPRECATED: Replaced with ParentTotalChildBounties storage item keeping dedicated counts for each parent bounty. Number of total child bounties. Will be removed in May 2025.

ChildBountyDescriptionsV1

The description of each child-bounty. Indexed by (parent_id, child_id).

ChildrenCuratorFees

The cumulative child-bounty curator fee for each parent bounty.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

ParentChildBounties

Number of active child bounties per parent bounty. Map of parent bounty index to number of child bounties.

ParentTotalChildBounties

Number of total child bounties per parent bounty, including completed bounties.

V0ToV1ChildBountyIds

The mapping of the child bounty ids from storage version V0 to the new V1 version.