yet_another_parachain_runtime

Crate pallet_collective

Collective system: Members of a set of account IDs can make their collective feelings known through dispatched calls from one of two specialized origins.

The membership can be provided in one of two ways: either directly, using the Root-dispatchable function set_members, or indirectly, through implementing the ChangeMembers. The pallet assumes that the amount of members stays at or below MaxMembers for its weight calculations, but enforces this neither in set_members nor in change_members_sorted.

A “prime” member may be set to help determine the default vote behavior based on chain config. If PrimeDefaultVote is used, the prime vote acts as the default vote in case of any abstentions after the voting period. If MoreThanMajorityThenPrimeDefaultVote is used, then abstentions will first follow the majority of the collective voting, and then the prime member.

Voting happens through motions comprising a proposal (i.e. a curried dispatchable) plus a number of approvals required for it to pass and be called. Motions are open for members to vote on for a minimum period given by MotionDuration. As soon as the needed number of approvals is given, the motion is closed and executed. If the number of approvals is not reached during the voting period, then close may be called by any account in order to force the end the motion explicitly. If a prime member is defined then their vote is used in place of any abstentions and the proposal is executed if there are enough approvals counting the new votes.

If there are not, or if no prime is set, then the motion is dropped without being executed.

Modules

deposit

Types implementing various cost strategies for a given proposal count.

migrations

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_collective

Structs

EnsureMember

EnsureMembers

EnsureProportionAtLeast

EnsureProportionMoreThan

GenesisConfig

Can be used to configure the genesis state of this pallet.

Instance1

Instance1 to be used for instantiable pallets defined with the #[pallet] macro. Instances 2-16 are also available but are hidden from docs.

MoreThanMajorityThenPrimeDefaultVote

First see if yes vote are over majority of the whole collective. If so, set the default vote as yes. Otherwise, use the prime member’s vote as the default vote.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

PrimeDefaultVote

Set the prime member’s vote as the default vote.

Votes

Info for keeping track of a motion being voted on.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

HoldReason

A reason for the pallet placing a hold on funds.

RawOrigin

Origin for the collective module.

Traits

Config

Configuration trait of this pallet.

DefaultVote

Default voting strategy when a member is inactive.

WeightInfo

Weight functions needed for pallet_collective.

Functions

ensure_members

Ensure that the origin o represents at least n members. Returns Ok or an Err otherwise.

Type Aliases

CostOf

Consideration cost created for publishing and storing a proposal.

MemberCount

A number of members.

Members

The current members of the collective. This is stored sorted (just by value).

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

Origin

Origin for the collective pallet.

Prime

The prime member that helps determine the default vote behavior in case of abstentions.

ProposalCount

Proposals so far.

ProposalIndex

Simple index type for proposal counting.

ProposalOf

Actual proposal for a given hash, if it’s current.

Proposals

The hashes of the active proposals.

Voting

Votes on a given proposal, if it is ongoing.