yet_another_parachain_runtime::pallet_contracts::chain_extension

Trait ChainExtension

pub trait ChainExtension<C>
where
    C: Config,
{
    // Required method
    fn call<E>(
        &mut self,
        env: Environment<'_, '_, E, InitState>,
    ) -> Result<RetVal, DispatchError>
       where E: Ext<T = C>;

    // Provided method
    fn enabled() -> bool { ... }
}

A trait used to extend the set of contract callable functions.

In order to create a custom chain extension this trait must be implemented and supplied to the pallet contracts configuration trait as the associated type of the same name. Consult the module documentation for a general explanation of chain extensions.

Lifetime

The extension will be Default initialized at the beginning of each call (not per call stack) and dropped afterwards. Hence any value held inside the extension can be used as a per-call scratch buffer.

Required Methods

fn call<E>( &mut self, env: Environment<'_, '_, E, InitState>, ) -> Result<RetVal, DispatchError>
where E: Ext<T = C>,

Call the chain extension logic.

This is the only function that needs to be implemented in order to write a chain extensions. It is called whenever a contract calls the seal_call_chain_extension imported wasm function.

Parameters

env: Access to the remaining arguments and the execution environment.

Return

In case of Err the contract execution is immediately suspended and the passed error is returned to the caller. Otherwise the value of RetVal determines the exit behaviour.

Note

The Self::call can be invoked within a read-only context, where any state-changing calls are disallowed. This information can be obtained using env.ext().is_read_only(). It is crucial for the implementer to handle this scenario appropriately.

Provided Methods

fn enabled() -> bool

Determines whether chain extensions are enabled for this chain.

The default implementation returns true. Therefore it is not necessary to overwrite this function when implementing a chain extension. In case of false the deployment of a contract that references seal_call_chain_extension will be denied and calling this function will return NoChainExtension without first calling into call.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.