yet_another_parachain_runtime::pallet_democracy::pallet

Trait Config

pub trait Config:
    Sized
    + Config
    + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{

    type WeightInfo: WeightInfo;
    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type Scheduler: Named<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number, Self::RuntimeCall, Self::PalletsOrigin, Hasher = Self::Hashing>;
    type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage;
    type Currency: ReservableCurrency<Self::AccountId> + LockableCurrency<Self::AccountId, Moment = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type EnactmentPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type LaunchPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type VotingPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type VoteLockingPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type MinimumDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type InstantAllowed: Get<bool>;
    type FastTrackVotingPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type CooloffPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type MaxVotes: Get<u32>;
    type MaxProposals: Get<u32>;
    type MaxDeposits: Get<u32>;
    type MaxBlacklisted: Get<u32>;
    type ExternalOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type ExternalMajorityOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type ExternalDefaultOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type SubmitOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>;
    type FastTrackOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type InstantOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type CancellationOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type BlacklistOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type CancelProposalOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type VetoOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>;
    type PalletsOrigin: From<RawOrigin<Self::AccountId>>;
    type Slash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type WeightInfo: WeightInfo

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

type Scheduler: Named<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number, Self::RuntimeCall, Self::PalletsOrigin, Hasher = Self::Hashing>

The Scheduler.

type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage

The Preimage provider.

type Currency: ReservableCurrency<Self::AccountId> + LockableCurrency<Self::AccountId, Moment = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Currency type for this pallet.

type EnactmentPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The period between a proposal being approved and enacted.

It should generally be a little more than the unstake period to ensure that voting stakers have an opportunity to remove themselves from the system in the case where they are on the losing side of a vote.

type LaunchPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

How often (in blocks) new public referenda are launched.

type VotingPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

How often (in blocks) to check for new votes.

type VoteLockingPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The minimum period of vote locking.

It should be no shorter than enactment period to ensure that in the case of an approval, those successful voters are locked into the consequences that their votes entail.

type MinimumDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The minimum amount to be used as a deposit for a public referendum proposal.

type InstantAllowed: Get<bool>

Indicator for whether an emergency origin is even allowed to happen. Some chains may want to set this permanently to false, others may want to condition it on things such as an upgrade having happened recently.

type FastTrackVotingPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Minimum voting period allowed for a fast-track referendum.

type CooloffPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Period in blocks where an external proposal may not be re-submitted after being vetoed.

type MaxVotes: Get<u32>

The maximum number of votes for an account.

Also used to compute weight, an overly big value can lead to extrinsic with very big weight: see delegate for instance.

type MaxProposals: Get<u32>

The maximum number of public proposals that can exist at any time.

type MaxDeposits: Get<u32>

The maximum number of deposits a public proposal may have at any time.

type MaxBlacklisted: Get<u32>

The maximum number of items which can be blacklisted.

type ExternalOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which the next tabled referendum may be forced. This is a normal โsuper-majority-requiredโ referendum.

type ExternalMajorityOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which the next tabled referendum may be forced; this allows for the tabling of a majority-carries referendum.

type ExternalDefaultOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which the next tabled referendum may be forced; this allows for the tabling of a negative-turnout-bias (default-carries) referendum.

type SubmitOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>

Origin from which the new proposal can be made.

The success variant is the account id of the depositor.

type FastTrackOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which the next majority-carries (or more permissive) referendum may be tabled to vote according to the FastTrackVotingPeriod asynchronously in a similar manner to the emergency origin. It retains its threshold method.

type InstantOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which the next majority-carries (or more permissive) referendum may be tabled to vote immediately and asynchronously in a similar manner to the emergency origin. It retains its threshold method.

type CancellationOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which any referendum may be cancelled in an emergency.

type BlacklistOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which proposals may be blacklisted.

type CancelProposalOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which a proposal may be cancelled and its backers slashed.

type VetoOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>

Origin for anyone able to veto proposals.

type PalletsOrigin: From<RawOrigin<Self::AccountId>>

Overarching type of all pallets origins.

type Slash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for the unbalanced reduction when slashing a preimage deposit.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.