yet_another_parachain_runtime::pallet_election_provider_multi_block

Trait Config

pub trait Config: Config + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{

    type UnsignedPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type SignedPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type SignedValidationPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type VoterSnapshotPerBlock: Get<u32>;
    type TargetSnapshotPerBlock: Get<u32>;
    type Pages: Get<u32>;
    type DataProvider: ElectionDataProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type MinerConfig: MinerConfig<Pages = Self::Pages, AccountId = Self::AccountId, MaxVotesPerVoter = <Self::DataProvider as ElectionDataProvider>::MaxVotesPerVoter, VoterSnapshotPerBlock = Self::VoterSnapshotPerBlock, TargetSnapshotPerBlock = Self::TargetSnapshotPerBlock, MaxBackersPerWinner = <Self::Verifier as Verifier>::MaxBackersPerWinner, MaxWinnersPerPage = <Self::Verifier as Verifier>::MaxWinnersPerPage>;
    type Fallback: InstantElectionProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, DataProvider = Self::DataProvider, MaxBackersPerWinner = <Self::Verifier as Verifier>::MaxBackersPerWinner, MaxWinnersPerPage = <Self::Verifier as Verifier>::MaxWinnersPerPage>;
    type Verifier: Verifier<Solution = <Self::MinerConfig as MinerConfig>::Solution, AccountId = Self::AccountId> + AsynchronousVerifier;
    type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type AreWeDone: Get<Phase<Self>>;
    type WeightInfo: WeightInfo;
    type OnRoundRotation: OnRoundRotation;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type UnsignedPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Duration of the unsigned phase.

type SignedPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Duration of the signed phase.

type SignedValidationPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Duration of the singed validation phase.

The duration of this should not be less than T::Pages, and there is no point in it being more than SignedPhase::MaxSubmission::get() * T::Pages. TODO: integrity test for it.

type VoterSnapshotPerBlock: Get<u32>

The number of snapshot voters to fetch per block.

type TargetSnapshotPerBlock: Get<u32>

The number of snapshot targets to fetch per block.

type Pages: Get<u32>

The number of pages.

The snapshot is created with this many keys in the storage map.

The solutions may contain at MOST this many pages, but less pages are acceptable as well.

type DataProvider: ElectionDataProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Something that will provide the election data.

type MinerConfig: MinerConfig<Pages = Self::Pages, AccountId = Self::AccountId, MaxVotesPerVoter = <Self::DataProvider as ElectionDataProvider>::MaxVotesPerVoter, VoterSnapshotPerBlock = Self::VoterSnapshotPerBlock, TargetSnapshotPerBlock = Self::TargetSnapshotPerBlock, MaxBackersPerWinner = <Self::Verifier as Verifier>::MaxBackersPerWinner, MaxWinnersPerPage = <Self::Verifier as Verifier>::MaxWinnersPerPage>

The miner configuration.

These configurations are passed to crate::unsigned::miner::BaseMiner. An external miner implementation should implement this trait, and use the said BaseMiner.

type Fallback: InstantElectionProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, DataProvider = Self::DataProvider, MaxBackersPerWinner = <Self::Verifier as Verifier>::MaxBackersPerWinner, MaxWinnersPerPage = <Self::Verifier as Verifier>::MaxWinnersPerPage>

The fallback type used for the election.

type Verifier: Verifier<Solution = <Self::MinerConfig as MinerConfig>::Solution, AccountId = Self::AccountId> + AsynchronousVerifier

The verifier pallet’s interface.

type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin that can perform administration operations on this pallet.

type AreWeDone: Get<Phase<Self>>

An indicator of whether we should move to do the crate::types::Phase::Done or not? This is called at the end of the election process.

Common implementation is ProceedRegardlessOf or RevertToSignedIfNotQueuedOf.

type WeightInfo: WeightInfo

The weight of the pallet.

type OnRoundRotation: OnRoundRotation

Single type that implement super::OnRoundRotation to do something when the round ends.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.