yet_another_parachain_runtime::pallet_election_provider_multi_phase

Trait Config

pub trait Config:
    Config
    + CreateBare<Call<Self>>
    + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{

    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent> + TryInto<Event<Self>>;
    type Currency: ReservableCurrency<Self::AccountId> + Currency<Self::AccountId>;
    type EstimateCallFee: EstimateCallFee<Call<Self>, <Self::Currency as Currency<Self::AccountId>>::Balance>;
    type UnsignedPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type SignedPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type BetterSignedThreshold: Get<Perbill>;
    type OffchainRepeat: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type MinerTxPriority: Get<u64>;
    type MinerConfig: MinerConfig<AccountId = Self::AccountId, MaxVotesPerVoter = <Self::DataProvider as ElectionDataProvider>::MaxVotesPerVoter, MaxWinners = Self::MaxWinners, MaxBackersPerWinner = Self::MaxBackersPerWinner>;
    type SignedMaxSubmissions: Get<u32>;
    type SignedMaxWeight: Get<Weight>;
    type SignedMaxRefunds: Get<u32>;
    type SignedRewardBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type SignedDepositByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type SignedDepositWeight: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaxWinners: Get<u32>;
    type MaxBackersPerWinner: Get<u32>;
    type SignedDepositBase: Convert<usize, <Self::Currency as Currency<Self::AccountId>>::Balance>;
    type ElectionBounds: Get<ElectionBounds>;
    type SlashHandler: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type RewardHandler: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::PositiveImbalance>;
    type DataProvider: ElectionDataProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type Fallback: InstantElectionProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, DataProvider = Self::DataProvider, MaxBackersPerWinner = Self::MaxBackersPerWinner, MaxWinnersPerPage = Self::MaxWinners>;
    type GovernanceFallback: InstantElectionProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, DataProvider = Self::DataProvider, MaxWinnersPerPage = Self::MaxWinners, MaxBackersPerWinner = Self::MaxBackersPerWinner>;
    type Solver: NposSolver<AccountId = Self::AccountId>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type BenchmarkingConfig: BenchmarkingConfig;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent> + TryInto<Event<Self>>

type Currency: ReservableCurrency<Self::AccountId> + Currency<Self::AccountId>

Currency type.

type EstimateCallFee: EstimateCallFee<Call<Self>, <Self::Currency as Currency<Self::AccountId>>::Balance>

Something that can predict the fee of a call. Used to sensibly distribute rewards.

type UnsignedPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Duration of the unsigned phase.

type SignedPhase: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Duration of the signed phase.

type BetterSignedThreshold: Get<Perbill>

The minimum amount of improvement to the solution score that defines a solution as “better” in the Signed phase.

type OffchainRepeat: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The repeat threshold of the offchain worker.

For example, if it is 5, that means that at least 5 blocks will elapse between attempts to submit the worker’s solution.

type MinerTxPriority: Get<u64>

The priority of the unsigned transaction submitted in the unsigned-phase

type MinerConfig: MinerConfig<AccountId = Self::AccountId, MaxVotesPerVoter = <Self::DataProvider as ElectionDataProvider>::MaxVotesPerVoter, MaxWinners = Self::MaxWinners, MaxBackersPerWinner = Self::MaxBackersPerWinner>

Configurations of the embedded miner.

Any external software implementing this can use the unsigned::Miner type provided, which can mine new solutions and trim them accordingly.

type SignedMaxSubmissions: Get<u32>

Maximum number of signed submissions that can be queued.

It is best to avoid adjusting this during an election, as it impacts downstream data structures. In particular, SignedSubmissionIndices<T> is bounded on this value. If you update this value during an election, you must ensure that SignedSubmissionIndices.len() is less than or equal to the new value. Otherwise, attempts to submit new solutions may cause a runtime panic.

type SignedMaxWeight: Get<Weight>

Maximum weight of a signed solution.

If Config::MinerConfig is being implemented to submit signed solutions (outside of this pallet), then MinerConfig::solution_weight is used to compare against this value.

type SignedMaxRefunds: Get<u32>

The maximum amount of unchecked solutions to refund the call fee for.

type SignedRewardBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

Base reward for a signed solution

type SignedDepositByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

Per-byte deposit for a signed solution.

type SignedDepositWeight: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

Per-weight deposit for a signed solution.

type MaxWinners: Get<u32>

Maximum number of winners that an election supports.

Note: This must always be greater or equal to T::DataProvider::desired_targets().

type MaxBackersPerWinner: Get<u32>

Maximum number of voters that can support a winner in an election solution.

This is needed to ensure election computation is bounded.

type SignedDepositBase: Convert<usize, <Self::Currency as Currency<Self::AccountId>>::Balance>

Something that calculates the signed deposit base based on the signed submissions queue size.

type ElectionBounds: Get<ElectionBounds>

The maximum number of electing voters and electable targets to put in the snapshot.

type SlashHandler: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for the slashed deposits.

type RewardHandler: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::PositiveImbalance>

Handler for the rewards.

type DataProvider: ElectionDataProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Something that will provide the election data.

type Fallback: InstantElectionProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, DataProvider = Self::DataProvider, MaxBackersPerWinner = Self::MaxBackersPerWinner, MaxWinnersPerPage = Self::MaxWinners>

Configuration for the fallback.

type GovernanceFallback: InstantElectionProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, DataProvider = Self::DataProvider, MaxWinnersPerPage = Self::MaxWinners, MaxBackersPerWinner = Self::MaxBackersPerWinner>

Configuration of the governance-only fallback.

As a side-note, it is recommend for test-nets to use type ElectionProvider = BoundedExecution<_> if the test-net is not expected to have thousands of nominators.

type Solver: NposSolver<AccountId = Self::AccountId>

OCW election solution miner algorithm implementation.

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that can control this pallet. Note that any action taken by this origin (such) as providing an emergency solution is not checked. Thus, it must be a trusted origin.

type BenchmarkingConfig: BenchmarkingConfig

The configuration of benchmarking.

type WeightInfo: WeightInfo

The weight of the pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.