yet_another_parachain_runtime::pallet_election_provider_multi_phase

Trait MinerConfig

pub trait MinerConfig {
    type AccountId: Ord + Clone + Codec + Debug;
    type Solution: Codec + Default + PartialEq + Eq + Clone + Debug + Ord + NposSolution + TypeInfo;
    type MaxVotesPerVoter;
    type MaxLength: Get<u32>;
    type MaxWeight: Get<Weight>;
    type MaxWinners: Get<u32>;
    type MaxBackersPerWinner: Get<u32>;

    // Required method
    fn solution_weight(
        voters: u32,
        targets: u32,
        active_voters: u32,
        degree: u32,
    ) -> Weight;
}

Configurations for a miner that comes with this pallet.

Required Associated Types

type AccountId: Ord + Clone + Codec + Debug

The account id type.

type Solution: Codec + Default + PartialEq + Eq + Clone + Debug + Ord + NposSolution + TypeInfo

The solution that the miner is mining.

type MaxVotesPerVoter

Maximum number of votes per voter in the snapshots.

type MaxLength: Get<u32>

Maximum length of the solution that the miner is allowed to generate.

Solutions are trimmed to respect this.

type MaxWeight: Get<Weight>

Maximum weight of the solution that the miner is allowed to generate.

Solutions are trimmed to respect this.

The weight is computed using solution_weight.

type MaxWinners: Get<u32>

The maximum number of winners that can be elected in the single page supported by this pallet.

type MaxBackersPerWinner: Get<u32>

The maximum number of backers per winner in the last solution.

Required Methods

fn solution_weight( voters: u32, targets: u32, active_voters: u32, degree: u32, ) -> Weight

Something that can compute the weight of a solution.

This weight estimate is then used to trim the solution, based on MinerConfig::MaxWeight.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.