yet_another_parachain_runtime

Crate pallet_grandpa

GRANDPA Consensus module for runtime.

This manages the GRANDPA authority set ready for the native code. These authorities are only for GRANDPA finality, not for consensus overall.

In the future, it will also handle misbehavior reports, and on-chain finality notifications.

For full integration with GRANDPA, the GrandpaApi should be implemented. The necessary items are re-exported via the fg_primitives crate.

Modules

fg_primitives

Primitives for GRANDPA integration, suitable for WASM compilation.

migrations

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

DefaultForState

EquivocationOffence

GRANDPA equivocation offence report.

EquivocationReportSystem

GRANDPA equivocation offence report system.

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

StoredPendingChange

A stored pending change. Limit is the bound for next_authorities

TimeSlot

A round number and set id which point on the time of an offence.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

StoredState

Current state of the GRANDPA authority set. State transitions must happen in the same order of states defined below, e.g. Paused implies a prior PendingPause.

Traits

Config

Configuration trait of this pallet.

WeightInfo

Functions

__type_value_for_default_for_state

Type Aliases

Authorities

The current list of authorities.

AuthorityId

Identity of a Grandpa authority.

AuthorityList

A list of Grandpa authorities with associated weights.

AuthorityWeight

The weight of an authority.

BoundedAuthorityList

Bounded version of AuthorityList, Limit being the bound

CurrentSetId

The number of changes (both in terms of keys and underlying economic responsibilities) in the “set” of Grandpa validators from genesis.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextForced

next block number where we can force a change.

PendingChange

Pending change: (signaled at, scheduled change).

SetIdSession

A mapping from grandpa set ID to the index of the most recent session for which its members were responsible.

Stalled

true if we are currently stalled.

State

State of the current authority set.