yet_another_parachain_runtime::pallet_identity

Trait Config

pub trait Config: Config + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{

    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type BasicDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type ByteDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type UsernameDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type SubAccountDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaxSubAccounts: Get<u32>;
    type IdentityInformation: IdentityInformationProvider;
    type MaxRegistrars: Get<u32>;
    type Slashed: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type RegistrarOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type OffchainSignature: Verify<Signer = Self::SigningPublicKey> + Parameter;
    type SigningPublicKey: IdentifyAccount<AccountId = Self::AccountId>;
    type UsernameAuthorityOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type PendingUsernameExpiration: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type UsernameGracePeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type MaxSuffixLength: Get<u32>;
    type MaxUsernameLength: Get<u32>;
    type BenchmarkHelper: BenchmarkHelper<Self::SigningPublicKey, Self::OffchainSignature>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

The overarching event type.

type Currency: ReservableCurrency<Self::AccountId>

The currency trait.

type BasicDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit for a registered identity.

type ByteDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit per encoded byte for a registered identity.

type UsernameDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit per registered username. This value should change only in runtime upgrades with proper migration of existing deposits.

type SubAccountDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit for a registered subaccount. This should account for the fact that one storage item’s value will increase by the size of an account ID, and there will be another trie item whose value is the size of an account ID plus 32 bytes.

type MaxSubAccounts: Get<u32>

The maximum number of sub-accounts allowed per identified account.

type IdentityInformation: IdentityInformationProvider

Structure holding information about an identity.

type MaxRegistrars: Get<u32>

Maximum number of registrars allowed in the system. Needed to bound the complexity of, e.g., updating judgements.

type Slashed: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

What to do with slashed funds.

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which may forcibly set or remove a name. Root can always do this.

type RegistrarOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which may add or remove registrars. Root can always do this.

type OffchainSignature: Verify<Signer = Self::SigningPublicKey> + Parameter

Signature type for pre-authorizing usernames off-chain.

Can verify whether an Self::SigningPublicKey created a signature.

type SigningPublicKey: IdentifyAccount<AccountId = Self::AccountId>

Public key that corresponds to an on-chain Self::AccountId.

type UsernameAuthorityOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which may add or remove username authorities. Root can always do this.

type PendingUsernameExpiration: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The number of blocks within which a username grant must be accepted.

type UsernameGracePeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The number of blocks that must pass to enable the permanent deletion of a username by its respective authority.

type MaxSuffixLength: Get<u32>

The maximum length of a suffix.

type MaxUsernameLength: Get<u32>

The maximum length of a username, including its suffix and any system-added delimiters.

type BenchmarkHelper: BenchmarkHelper<Self::SigningPublicKey, Self::OffchainSignature>

A set of helper functions for benchmarking. The default configuration () uses the SR25519 signature schema.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.