yet_another_parachain_runtime

Crate pallet_im_online

I’m online Pallet

If the local node is a validator (i.e. contains an authority key), this pallet gossips a heartbeat transaction with each new session. The heartbeat functions as a simple mechanism to signal that the node is online in the current era.

Received heartbeats are tracked for one era and reset with each new era. The pallet exposes two public functions to query if a heartbeat has been received in the current era or session.

The heartbeat is a signed transaction, which was signed using the session key and includes the recent best block number of the local validators chain. It is submitted as an Unsigned Transaction via off-chain workers.

Config

Call

Pallet

Interface

Public Functions

is_online - True if the validator sent a heartbeat in the current session.

Usage

use pallet_im_online::{self as im_online};

#[frame_support::pallet]
pub mod pallet {
	use super::*;
	use frame_support::pallet_prelude::*;
	use frame_system::pallet_prelude::*;

	#[pallet::pallet]
	pub struct Pallet<T>(_);

	#[pallet::config]
	pub trait Config: frame_system::Config + im_online::Config {}

	#[pallet::call]
	impl<T: Config> Pallet<T> {
		#[pallet::weight(0)]
		pub fn is_online(origin: OriginFor<T>, authority_index: u32) -> DispatchResult {
			let _sender = ensure_signed(origin)?;
			let _is_online = <im_online::Pallet<T>>::is_online(authority_index);
			Ok(())
		}
	}
}

Dependencies

This pallet depends on the Session pallet.

Modules

ed25519

migration

Storage migrations for the im-online pallet.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

sr25519

weights

Autogenerated weights for pallet_im_online

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Heartbeat

Heartbeat which is sent/received.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

UnresponsivenessOffence

An offence that is filed if a validator didn’t send a heartbeat message.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

WeightInfo

Weight functions needed for pallet_im_online.

Type Aliases

AuthIndex

AuthoredBlocks

For each session index, we keep a mapping of ValidatorId<T> to the number of blocks authored by the given authority.

HeartbeatAfter

The block number after which it’s ok to send heartbeats in the current session.

IdentificationTuple

A tuple of (ValidatorId, Identification) where Identification is the full identification of ValidatorId.

Keys

The current set of keys that may issue a heartbeat.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

ReceivedHeartbeats

For each session index, we keep a mapping of SessionIndex and AuthIndex.

ValidatorId

A type for representing the validator id in a session.