yet_another_parachain_runtime

Crate pallet_lottery

A lottery pallet that uses participation in the network to purchase tickets.

With this pallet, you can configure a lottery, which is a pot of money that users contribute to, and that is reallocated to a single user at the end of the lottery period. Just like a normal lottery system, to participate, you need to “buy a ticket”, which is used to fund the pot.

The unique feature of this lottery system is that tickets can only be purchased by making a “valid call” dispatched through this pallet. By configuring certain calls to be valid for the lottery, you can encourage users to make those calls on your network. An example of how this could be used is to set validator nominations as a valid lottery call. If the lottery is set to repeat every month, then users would be encouraged to re-nominate validators every month. A user can only purchase one ticket per valid call per lottery.

This pallet can be configured to use dynamically set calls or statically set calls. Call validation happens through the ValidateCall implementation. This pallet provides one implementation of this using the CallIndices storage item. You can also make your own implementation at the runtime level which can contain much more complex logic, such as validation of the parameters, which this pallet alone cannot do.

This pallet uses the modulus operator to pick a random winner. It is known that this might introduce a bias if the random number chosen in a range that is not perfectly divisible by the total number of participants. The MaxGenerateRandom configuration can help mitigate this by generating new numbers until we hit the limit or we find a “fair” number. This is best effort only.

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_lottery

Structs

LotteryConfig

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

ValidateCall

WeightInfo

Weight functions needed for pallet_lottery.

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.