yet_another_parachain_runtime::pallet_mmr::primitives

Trait MmrApi

pub trait MmrApi<Block, Hash, BlockNumber>: Core<Block>
where
    Block: Block,
    Hash: Codec,
    BlockNumber: Codec,
{
    // Provided methods
    fn mmr_root(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
    ) -> Result<Result<Hash, Error>, ApiError> { ... }
    fn mmr_leaf_count(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
    ) -> Result<Result<u64, Error>, ApiError> { ... }
    fn generate_proof(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
        block_numbers: Vec<BlockNumber>,
        best_known_block_number: Option<BlockNumber>,
    ) -> Result<Result<(Vec<EncodableOpaqueLeaf>, LeafProof<Hash>), Error>, ApiError> { ... }
    fn generate_ancestry_proof(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
        prev_block_number: BlockNumber,
        best_known_block_number: Option<BlockNumber>,
    ) -> Result<Result<AncestryProof<Hash>, Error>, ApiError> { ... }
    fn verify_proof(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
        leaves: Vec<EncodableOpaqueLeaf>,
        proof: LeafProof<Hash>,
    ) -> Result<Result<(), Error>, ApiError> { ... }
    fn verify_proof_stateless(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
        root: Hash,
        leaves: Vec<EncodableOpaqueLeaf>,
        proof: LeafProof<Hash>,
    ) -> Result<Result<(), Error>, ApiError> { ... }
}

API to interact with MMR pallet.

Provided Methods

fn mmr_root( &self, __runtime_api_at_param__: <Block as Block>::Hash, ) -> Result<Result<Hash, Error>, ApiError>

Return the on-chain MMR root hash.

fn mmr_leaf_count( &self, __runtime_api_at_param__: <Block as Block>::Hash, ) -> Result<Result<u64, Error>, ApiError>

Return the number of MMR blocks in the chain.

fn generate_proof( &self, __runtime_api_at_param__: <Block as Block>::Hash, block_numbers: Vec<BlockNumber>, best_known_block_number: Option<BlockNumber>, ) -> Result<Result<(Vec<EncodableOpaqueLeaf>, LeafProof<Hash>), Error>, ApiError>

Generate MMR proof for a series of block numbers. If best_known_block_number = Some(n), use historical MMR state at given block height n. Else, use current MMR state.

fn generate_ancestry_proof( &self, __runtime_api_at_param__: <Block as Block>::Hash, prev_block_number: BlockNumber, best_known_block_number: Option<BlockNumber>, ) -> Result<Result<AncestryProof<Hash>, Error>, ApiError>

Generates a proof that the prev_block_number is part of the canonical chain at best_known_block_number.

fn verify_proof( &self, __runtime_api_at_param__: <Block as Block>::Hash, leaves: Vec<EncodableOpaqueLeaf>, proof: LeafProof<Hash>, ) -> Result<Result<(), Error>, ApiError>

Verify MMR proof against on-chain MMR for a batch of leaves.

Note this function will use on-chain MMR root hash and check if the proof matches the hash. Note, the leaves should be sorted such that corresponding leaves and leaf indices have the same position in both the leaves vector and the leaf_indices vector contained in the LeafProof

fn verify_proof_stateless( &self, __runtime_api_at_param__: <Block as Block>::Hash, root: Hash, leaves: Vec<EncodableOpaqueLeaf>, proof: LeafProof<Hash>, ) -> Result<Result<(), Error>, ApiError>

Verify MMR proof against given root hash for a batch of leaves.

Note this function does not require any on-chain storage - the proof is verified against given MMR root hash.

Note, the leaves should be sorted such that corresponding leaves and leaf indices have the same position in both the leaves vector and the leaf_indices vector contained in the LeafProof