yet_another_parachain_runtime::pallet_mmr

Trait LeafDataProvider

pub trait LeafDataProvider {
    type LeafData: FullLeaf + Decode;

    // Required method
    fn leaf_data() -> Self::LeafData;
}

A provider of the MMR’s leaf data.

Required Associated Types

type LeafData: FullLeaf + Decode

A type that should end up in the leaf of MMR.

Required Methods

fn leaf_data() -> Self::LeafData

The method to return leaf data that should be placed in the leaf node appended MMR at this block.

This is being called by the on_initialize method of this pallet at the very beginning of each block.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.