yet_another_parachain_runtime::pallet_nomination_pools::adapter

Trait StakeStrategy

pub trait StakeStrategy {
    type Balance: Balance;
    type AccountId: Clone + Debug;
    type CoreStaking: StakingInterface<Balance = Self::Balance, AccountId = Self::AccountId>;

    // Required methods
    fn strategy_type() -> StakeStrategyType;
    fn transferable_balance(
        pool_account: Pool<Self::AccountId>,
        member_account: Member<Self::AccountId>,
    ) -> Self::Balance;
    fn total_balance(
        pool_account: Pool<Self::AccountId>,
    ) -> Option<Self::Balance>;
    fn member_delegation_balance(
        member_account: Member<Self::AccountId>,
    ) -> Option<Self::Balance>;
    fn pledge_bond(
        who: Member<Self::AccountId>,
        pool_account: Pool<Self::AccountId>,
        reward_account: &Self::AccountId,
        amount: Self::Balance,
        bond_type: BondType,
    ) -> Result<(), DispatchError>;
    fn member_withdraw(
        who: Member<Self::AccountId>,
        pool_account: Pool<Self::AccountId>,
        amount: Self::Balance,
        num_slashing_spans: u32,
    ) -> Result<(), DispatchError>;
    fn dissolve(
        pool_account: Pool<Self::AccountId>,
    ) -> Result<(), DispatchError>;
    fn pending_slash(pool_account: Pool<Self::AccountId>) -> Self::Balance;
    fn member_slash(
        who: Member<Self::AccountId>,
        pool_account: Pool<Self::AccountId>,
        amount: Self::Balance,
        maybe_reporter: Option<Self::AccountId>,
    ) -> Result<(), DispatchError>;
    fn migrate_nominator_to_agent(
        pool_account: Pool<Self::AccountId>,
        reward_account: &Self::AccountId,
    ) -> Result<(), DispatchError>;
    fn migrate_delegation(
        pool: Pool<Self::AccountId>,
        delegator: Member<Self::AccountId>,
        value: Self::Balance,
    ) -> Result<(), DispatchError>;

    // Provided methods
    fn bonding_duration() -> u32 { ... }
    fn current_era() -> u32 { ... }
    fn minimum_nominator_bond() -> Self::Balance { ... }
    fn active_stake(pool_account: Pool<Self::AccountId>) -> Self::Balance { ... }
    fn total_stake(pool_account: Pool<Self::AccountId>) -> Self::Balance { ... }
    fn pool_strategy(pool_account: Pool<Self::AccountId>) -> StakeStrategyType { ... }
    fn nominate(
        pool_account: Pool<Self::AccountId>,
        validators: Vec<Self::AccountId>,
    ) -> Result<(), DispatchError> { ... }
    fn chill(pool_account: Pool<Self::AccountId>) -> Result<(), DispatchError> { ... }
    fn unbond(
        pool_account: Pool<Self::AccountId>,
        amount: Self::Balance,
    ) -> Result<(), DispatchError> { ... }
    fn withdraw_unbonded(
        pool_account: Pool<Self::AccountId>,
        num_slashing_spans: u32,
    ) -> Result<bool, DispatchError> { ... }
    fn nominations(
        pool_account: Pool<Self::AccountId>,
    ) -> Option<Vec<Self::AccountId>> { ... }
    fn remove_as_agent(_pool: Pool<Self::AccountId>) { ... }

}

An adapter trait that can support multiple staking strategies.

Depending on which staking strategy we want to use, the staking logic can be slightly different. Refer the two possible strategies currently: TransferStake and DelegateStake for more detail.

Required Associated Types

type Balance: Balance

type AccountId: Clone + Debug

type CoreStaking: StakingInterface<Balance = Self::Balance, AccountId = Self::AccountId>

Required Methods

fn strategy_type() -> StakeStrategyType

The type of staking strategy of the current adapter.

fn transferable_balance( pool_account: Pool<Self::AccountId>, member_account: Member<Self::AccountId>, ) -> Self::Balance

Balance that can be transferred from pool account to member.

This is part of the pool balance that can be withdrawn.

fn total_balance(pool_account: Pool<Self::AccountId>) -> Option<Self::Balance>

Total balance of the pool including amount that is actively staked.

fn member_delegation_balance( member_account: Member<Self::AccountId>, ) -> Option<Self::Balance>

Amount of tokens delegated by the member.

fn pledge_bond( who: Member<Self::AccountId>, pool_account: Pool<Self::AccountId>, reward_account: &Self::AccountId, amount: Self::Balance, bond_type: BondType, ) -> Result<(), DispatchError>

Pledge amount towards pool_account and update the pool bond. Also see StakingInterface::bond.

fn member_withdraw( who: Member<Self::AccountId>, pool_account: Pool<Self::AccountId>, amount: Self::Balance, num_slashing_spans: u32, ) -> Result<(), DispatchError>

Withdraw funds from pool account to member account.

fn dissolve(pool_account: Pool<Self::AccountId>) -> Result<(), DispatchError>

Dissolve the pool account.

fn pending_slash(pool_account: Pool<Self::AccountId>) -> Self::Balance

Check if there is any pending slash for the pool.

fn member_slash( who: Member<Self::AccountId>, pool_account: Pool<Self::AccountId>, amount: Self::Balance, maybe_reporter: Option<Self::AccountId>, ) -> Result<(), DispatchError>

Slash the member account with amount against pending slashes for the pool.

fn migrate_nominator_to_agent( pool_account: Pool<Self::AccountId>, reward_account: &Self::AccountId, ) -> Result<(), DispatchError>

Migrate pool account from being a direct nominator to a delegated agent.

This is useful for migrating a pool account from StakeStrategyType::Transfer to StakeStrategyType::Delegate.

fn migrate_delegation( pool: Pool<Self::AccountId>, delegator: Member<Self::AccountId>, value: Self::Balance, ) -> Result<(), DispatchError>

Migrate member balance from pool account to member account.

This is useful for a pool account that migrated from StakeStrategyType::Transfer to StakeStrategyType::Delegate. Its members can then migrate their delegated balance back to their account.

Internally, the member funds that are locked in the pool account are transferred back and locked in the member account.

Provided Methods

fn bonding_duration() -> u32

See StakingInterface::bonding_duration.

fn current_era() -> u32

See StakingInterface::current_era.

fn minimum_nominator_bond() -> Self::Balance

See StakingInterface::minimum_nominator_bond.

fn active_stake(pool_account: Pool<Self::AccountId>) -> Self::Balance

See StakingInterface::active_stake.

fn total_stake(pool_account: Pool<Self::AccountId>) -> Self::Balance

See StakingInterface::total_stake.

fn pool_strategy(pool_account: Pool<Self::AccountId>) -> StakeStrategyType

Which strategy the pool account is using.

This can be different from the Self::strategy_type of the adapter if the pool has not migrated to the new strategy yet.

fn nominate( pool_account: Pool<Self::AccountId>, validators: Vec<Self::AccountId>, ) -> Result<(), DispatchError>

See StakingInterface::nominate.

fn chill(pool_account: Pool<Self::AccountId>) -> Result<(), DispatchError>

See StakingInterface::chill.

fn unbond( pool_account: Pool<Self::AccountId>, amount: Self::Balance, ) -> Result<(), DispatchError>

See StakingInterface::unbond.

fn withdraw_unbonded( pool_account: Pool<Self::AccountId>, num_slashing_spans: u32, ) -> Result<bool, DispatchError>

See StakingInterface::withdraw_unbonded.

fn nominations( pool_account: Pool<Self::AccountId>, ) -> Option<Vec<Self::AccountId>>

List of validators nominated by the pool account.

fn remove_as_agent(_pool: Pool<Self::AccountId>)

Remove the pool account as agent.

Useful for migrating pool account from a delegated agent to a direct nominator. Only used in tests and benchmarks.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.