yet_another_parachain_runtime::pallet_preimage

Trait Config

pub trait Config: Config + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{
    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type Currency: ReservableCurrency<Self::AccountId>;
    type ManagerOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Consideration: Consideration<Self::AccountId, Footprint>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

The overarching event type.

type WeightInfo: WeightInfo

The Weight information for this pallet.

type Currency: ReservableCurrency<Self::AccountId>

Currency type for this pallet.

type ManagerOrigin: EnsureOrigin<Self::RuntimeOrigin>

An origin that can request a preimage be placed on-chain without a deposit or fee, or manage existing preimages.

type Consideration: Consideration<Self::AccountId, Footprint>

A means of providing some cost while data is stored on-chain.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.