yet_another_parachain_runtime::pallet_ranked_collective

Trait Config

pub trait Config<I = ()>: Config + Config
where
    Self::RuntimeEvent: From<Event<Self, I>>,
    I: 'static,
{

    type WeightInfo: WeightInfo;
    type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>;
    type AddOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type RemoveOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = u16>;
    type PromoteOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = u16>;
    type DemoteOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = u16>;
    type ExchangeOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Polls: Polling<Tally<Self, I, Pallet<Self, I>>, Votes = u32, Moment = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type MinRankOfClass: Convert<<Self::Polls as Polling<Tally<Self, I, Pallet<Self, I>>>>::Class, u16>;
    type MemberSwappedHandler: RankedMembersSwapHandler<<Pallet<Self, I> as RankedMembers>::AccountId, <Pallet<Self, I> as RankedMembers>::Rank>;
    type VoteWeight: Convert<u16, u32>;
    type MaxMemberCount: MaybeConvert<u16, u32>;
    type BenchmarkSetup: BenchmarkSetup<Self::AccountId>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>

The runtime event type.

type AddOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin required to add a member.

type RemoveOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = u16>

The origin required to remove a member.

The success value indicates the maximum rank from which the removal may be.

type PromoteOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = u16>

The origin required to promote a member. The success value indicates the maximum rank to which the promotion may be.

type DemoteOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = u16>

The origin required to demote a member. The success value indicates the maximum rank from which the demotion may be.

type ExchangeOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin that can swap the account of a member.

type Polls: Polling<Tally<Self, I, Pallet<Self, I>>, Votes = u32, Moment = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The polling system used for our voting.

type MinRankOfClass: Convert<<Self::Polls as Polling<Tally<Self, I, Pallet<Self, I>>>>::Class, u16>

Convert the tally class into the minimum rank required to vote on the poll. If Polls::Class is the same type as Rank, then Identity can be used here to mean “a rank of at least the poll class”.

type MemberSwappedHandler: RankedMembersSwapHandler<<Pallet<Self, I> as RankedMembers>::AccountId, <Pallet<Self, I> as RankedMembers>::Rank>

An external handler that will be notified when two members are swapped.

type VoteWeight: Convert<u16, u32>

Convert a rank_delta into a number of votes the rank gets.

Rank_delta is defined as the number of ranks above the minimum required to take part in the poll.

type MaxMemberCount: MaybeConvert<u16, u32>

The maximum number of members for a given rank in the collective.

The member at rank x contributes to the count at rank x and all ranks below it. Therefore, the limit m at rank x sets the maximum total member count for rank x and all ranks above. The None indicates no member count limit for the given rank.

type BenchmarkSetup: BenchmarkSetup<Self::AccountId>

Setup a member for benchmarking.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.