yet_another_parachain_runtime::pallet_recovery::pallet

Trait Config

pub trait Config: Config + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{
    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin, PostInfo = PostDispatchInfo> + GetDispatchInfo + From<Call<Self>>;
    type BlockNumberProvider: BlockNumberProvider;
    type Currency: ReservableCurrency<Self::AccountId>;
    type ConfigDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type FriendDepositFactor: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaxFriends: Get<u32>;
    type RecoveryDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

The overarching event type.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin, PostInfo = PostDispatchInfo> + GetDispatchInfo + From<Call<Self>>

The overarching call type.

type BlockNumberProvider: BlockNumberProvider

Query the current block number.

Must return monotonically increasing values when called from consecutive blocks. Can be configured to return either:

the local block number of the runtime via frame_system::Pallet

a remote block number, eg from the relay chain through RelaychainDataProvider

an arbitrary value through a custom implementation of the trait

There is currently no migration provided to “hot-swap” block number providers and it may result in undefined behavior when doing so. Parachains are therefore best off setting this to their local block number provider if they have the pallet already deployed.

Suggested values:

Solo- and Relay-chains: frame_system::Pallet

Parachains that may produce blocks sparingly or only when needed (on-demand):

already have the pallet deployed: frame_system::Pallet

are freshly deploying this pallet: RelaychainDataProvider

Parachains with a reliably block production rate (PLO or bulk-coretime):

already have the pallet deployed: frame_system::Pallet

are freshly deploying this pallet: no strong recommendation. Both local and remote providers can be used. Relay provider can be a bit better in cases where the parachain is lagging its block production to avoid clock skew.

type Currency: ReservableCurrency<Self::AccountId>

The currency mechanism.

type ConfigDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The base amount of currency needed to reserve for creating a recovery configuration.

This is held for an additional storage item whose value size is 2 + sizeof(BlockNumber, Balance) bytes.

type FriendDepositFactor: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount of currency needed per additional user when creating a recovery configuration.

This is held for adding sizeof(AccountId) bytes more into a pre-existing storage value.

type MaxFriends: Get<u32>

The maximum amount of friends allowed in a recovery configuration.

NOTE: The threshold programmed in this Pallet uses u16, so it does not really make sense to have a limit here greater than u16::MAX. But also, that is a lot more than you should probably set this value to anyway…

type RecoveryDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The base amount of currency needed to reserve for starting a recovery.

This is primarily held for deterring malicious recovery attempts, and should have a value large enough that a bad actor would choose not to place this deposit. It also acts to fund additional storage item whose value size is sizeof(BlockNumber, Balance + T * AccountId) bytes. Where T is a configurable threshold.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.