yet_another_parachain_runtime::pallet_referenda

Trait Config

pub trait Config<I = ()>:
    Sized
    + Config
    + Config
where
    Self::RuntimeEvent: From<Event<Self, I>>,
    I: 'static,
{

    type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + From<Call<Self, I>> + IsType<Self::RuntimeCall> + From<Call<Self>>;
    type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type Scheduler: Anon<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber, Self::RuntimeCall, <Self::RuntimeOrigin as OriginTrait>::PalletsOrigin, Hasher = Self::Hashing> + Named<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber, Self::RuntimeCall, <Self::RuntimeOrigin as OriginTrait>::PalletsOrigin, Hasher = Self::Hashing>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type SubmitOrigin: EnsureOriginWithArg<Self::RuntimeOrigin, <Self::RuntimeOrigin as OriginTrait>::PalletsOrigin, Success = Self::AccountId>;
    type CancelOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type KillOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Slash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type Votes: AtLeast32BitUnsigned + Copy + Parameter + Member + MaxEncodedLen;
    type Tally: VoteTally<Self::Votes, <Self::Tracks as TracksInfo<<Self::Currency as Currency<Self::AccountId>>::Balance, <Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>>::Id> + Clone + Codec + Eq + Debug + TypeInfo + MaxEncodedLen;
    type SubmissionDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaxQueued: Get<u32>;
    type UndecidingTimeout: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type AlarmInterval: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type Tracks: TracksInfo<<Self::Currency as Currency<Self::AccountId>>::Balance, <Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber, RuntimeOrigin = <Self::RuntimeOrigin as OriginTrait>::PalletsOrigin>;
    type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage;
    type BlockNumberProvider: BlockNumberProvider;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + From<Call<Self, I>> + IsType<Self::RuntimeCall> + From<Call<Self>>

type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type Scheduler: Anon<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber, Self::RuntimeCall, <Self::RuntimeOrigin as OriginTrait>::PalletsOrigin, Hasher = Self::Hashing> + Named<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber, Self::RuntimeCall, <Self::RuntimeOrigin as OriginTrait>::PalletsOrigin, Hasher = Self::Hashing>

The Scheduler.

type Currency: ReservableCurrency<Self::AccountId>

Currency type for this pallet.

type SubmitOrigin: EnsureOriginWithArg<Self::RuntimeOrigin, <Self::RuntimeOrigin as OriginTrait>::PalletsOrigin, Success = Self::AccountId>

Origin from which proposals may be submitted.

type CancelOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which any vote may be cancelled.

type KillOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which any vote may be killed.

type Slash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for the unbalanced reduction when slashing a preimage deposit.

type Votes: AtLeast32BitUnsigned + Copy + Parameter + Member + MaxEncodedLen

The counting type for votes. Usually just balance.

type Tally: VoteTally<Self::Votes, <Self::Tracks as TracksInfo<<Self::Currency as Currency<Self::AccountId>>::Balance, <Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>>::Id> + Clone + Codec + Eq + Debug + TypeInfo + MaxEncodedLen

The tallying type.

type SubmissionDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The minimum amount to be used as a deposit for a public referendum proposal.

type MaxQueued: Get<u32>

Maximum size of the referendum queue for a single track.

type UndecidingTimeout: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

The number of blocks after submission that a referendum must begin being decided by. Once this passes, then anyone may cancel the referendum.

type AlarmInterval: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

Quantization level for the referendum wakeup scheduler. A higher number will result in fewer storage reads/writes needed for smaller voters, but also result in delays to the automatic referendum status changes. Explicit servicing instructions are unaffected.

type Tracks: TracksInfo<<Self::Currency as Currency<Self::AccountId>>::Balance, <Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber, RuntimeOrigin = <Self::RuntimeOrigin as OriginTrait>::PalletsOrigin>

Information concerning the different referendum tracks.

type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage

The preimage provider.

type BlockNumberProvider: BlockNumberProvider

Provider for the block number.

Normally this is the frame_system pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.