yet_another_parachain_runtime::pallet_revive::evm

Struct StateOverrideSet

pub struct StateOverrideSet(pub BTreeMap<H160, StateOverride>);

A mapping from account addresses to their state overrides, used to temporarily modify account state during eth_call and similar simulation methods without affecting on-chain data.

Each entry maps an Address to a StateOverride that specifies which parts of the account’s state to replace for the duration of the call.

Conforms to the Geth state override set specification.

Tuple Fields
0: BTreeMap<H160, StateOverride>