yet_another_parachain_runtime::pallet_revive::precompiles::alloy::primitives

Module aliases

Type aliases for common primitive types.

Type Aliases

B8

1-byte fixed byte-array type.

B16

2-byte fixed byte-array type.

B32

4-byte fixed byte-array type.

B64

8-byte fixed byte-array type.

B96

12-byte fixed byte-array type.

B128

16-byte fixed byte-array type.

B192

24-byte fixed byte-array type.

B224

28-byte fixed byte-array type.

B256

32-byte fixed byte-array type.

B512

64-byte fixed byte-array type.

B1024

128-byte fixed byte-array type.

B2048

256-byte fixed byte-array type.

BlockHash

A block hash.

BlockNumber

A block number.

BlockTimestamp

A block timestamp.

ChainId

Chain identifier type (introduced in EIP-155).

I0

The 0-bit signed integer type, capable of representing 0.

I1

The 1-bit signed integer type, capable of representing 0 and -1.

I8

8-bit signed integer type, consisting of 1, 64-bit limbs.

I16

16-bit signed integer type, consisting of 1, 64-bit limbs.

I24

24-bit signed integer type, consisting of 1, 64-bit limbs.

I32

32-bit signed integer type, consisting of 1, 64-bit limbs.

I40

40-bit signed integer type, consisting of 1, 64-bit limbs.

I48

48-bit signed integer type, consisting of 1, 64-bit limbs.

I56

56-bit signed integer type, consisting of 1, 64-bit limbs.

I64

64-bit signed integer type, consisting of 1, 64-bit limbs.

I72

72-bit signed integer type, consisting of 2, 64-bit limbs.

I80

80-bit signed integer type, consisting of 2, 64-bit limbs.

I88

88-bit signed integer type, consisting of 2, 64-bit limbs.

I96

96-bit signed integer type, consisting of 2, 64-bit limbs.

I104

104-bit signed integer type, consisting of 2, 64-bit limbs.

I112

112-bit signed integer type, consisting of 2, 64-bit limbs.

I120

120-bit signed integer type, consisting of 2, 64-bit limbs.

I128

128-bit signed integer type, consisting of 2, 64-bit limbs.

I136

136-bit signed integer type, consisting of 3, 64-bit limbs.

I144

144-bit signed integer type, consisting of 3, 64-bit limbs.

I152

152-bit signed integer type, consisting of 3, 64-bit limbs.

I160

160-bit signed integer type, consisting of 3, 64-bit limbs.

I168

168-bit signed integer type, consisting of 3, 64-bit limbs.

I176

176-bit signed integer type, consisting of 3, 64-bit limbs.

I184

184-bit signed integer type, consisting of 3, 64-bit limbs.

I192

192-bit signed integer type, consisting of 3, 64-bit limbs.

I200

200-bit signed integer type, consisting of 4, 64-bit limbs.

I208

208-bit signed integer type, consisting of 4, 64-bit limbs.

I216

216-bit signed integer type, consisting of 4, 64-bit limbs.

I224

224-bit signed integer type, consisting of 4, 64-bit limbs.

I232

232-bit signed integer type, consisting of 4, 64-bit limbs.

I240

240-bit signed integer type, consisting of 4, 64-bit limbs.

I248

248-bit signed integer type, consisting of 4, 64-bit limbs.

I256

256-bit signed integer type, consisting of 4, 64-bit limbs.

I512

512-bit signed integer type, consisting of 8, 64-bit limbs.

Selector

Solidity contract functions are addressed using the first four bytes of the Keccak-256 hash of their signature.

StorageKey

An account storage key.

StorageValue

An account storage value.

TxHash

A transaction hash is a keccak hash of an RLP encoded signed transaction.

TxIndex

The index of transaction in a block.

TxNonce

The nonce of a transaction.

TxNumber

The sequence number of all existing transactions.

U0

Uint for 0 bits. Always zero. Similar to ().

U1

Uint for 1 bit. Similar to bool.

U8

8-bit unsigned integer type, consisting of 1, 64-bit limbs.

U16

16-bit unsigned integer type, consisting of 1, 64-bit limbs.

U24

24-bit unsigned integer type, consisting of 1, 64-bit limbs.

U32

32-bit unsigned integer type, consisting of 1, 64-bit limbs.

U40

40-bit unsigned integer type, consisting of 1, 64-bit limbs.

U48

48-bit unsigned integer type, consisting of 1, 64-bit limbs.

U56

56-bit unsigned integer type, consisting of 1, 64-bit limbs.

U64

64-bit unsigned integer type, consisting of 1, 64-bit limbs.

U72

72-bit unsigned integer type, consisting of 2, 64-bit limbs.

U80

80-bit unsigned integer type, consisting of 2, 64-bit limbs.

U88

88-bit unsigned integer type, consisting of 2, 64-bit limbs.

U96

96-bit unsigned integer type, consisting of 2, 64-bit limbs.

U104

104-bit unsigned integer type, consisting of 2, 64-bit limbs.

U112

112-bit unsigned integer type, consisting of 2, 64-bit limbs.

U120

120-bit unsigned integer type, consisting of 2, 64-bit limbs.

U128

128-bit unsigned integer type, consisting of 2, 64-bit limbs.

U136

136-bit unsigned integer type, consisting of 3, 64-bit limbs.

U144

144-bit unsigned integer type, consisting of 3, 64-bit limbs.

U152

152-bit unsigned integer type, consisting of 3, 64-bit limbs.

U160

160-bit unsigned integer type, consisting of 3, 64-bit limbs.

U168

168-bit unsigned integer type, consisting of 3, 64-bit limbs.

U176

176-bit unsigned integer type, consisting of 3, 64-bit limbs.

U184

184-bit unsigned integer type, consisting of 3, 64-bit limbs.

U192

192-bit unsigned integer type, consisting of 3, 64-bit limbs.

U200

200-bit unsigned integer type, consisting of 4, 64-bit limbs.

U208

208-bit unsigned integer type, consisting of 4, 64-bit limbs.

U216

216-bit unsigned integer type, consisting of 4, 64-bit limbs.

U224

224-bit unsigned integer type, consisting of 4, 64-bit limbs.

U232

232-bit unsigned integer type, consisting of 4, 64-bit limbs.

U240

240-bit unsigned integer type, consisting of 4, 64-bit limbs.

U248

248-bit unsigned integer type, consisting of 4, 64-bit limbs.

U256

256-bit unsigned integer type, consisting of 4, 64-bit limbs.

U320

Uint for 320 bits.

U384

Uint for 384 bits.

U448

Uint for 448 bits.

U512

512-bit unsigned integer type, consisting of 8, 64-bit limbs.

U1024

Uint for 1024 bits.

U2048

Uint for 2048 bits.

U4096

Uint for 4096 bits.