yet_another_parachain_runtime::pallet_revive::precompiles::alloy::primitives

Struct Signed

pub struct Signed<const BITS: usize, const LIMBS: usize>(/* private fields */);

Signed integer wrapping a ruint::Uint.

This signed integer implementation is fully abstract across the number of bits. It wraps a ruint::Uint, and co-opts the most significant bit to represent the sign. The number is represented in two’s complement, using the underlying Uint’s u64 limbs. The limbs can be accessed via the Signed::as_limbs() method, and are least-significant first.

Aliases

We provide aliases for every bit-width divisble by 8, from 8 to 256. These are located in crate::aliases and are named I256, I248 etc. Most users will want crate::I256.

Usage

// Instantiate from a number
let a = I256::unchecked_from(1);
// Use `try_from` if you're not sure it'll fit
let b = I256::try_from(200000382).unwrap();

// Or parse from a string :)
let c = "100".parse::<I256>().unwrap();
let d = "-0x138f".parse::<I256>().unwrap();

// Preceding plus is allowed but not recommended
let e = "+0xdeadbeef".parse::<I256>().unwrap();

// Underscores are ignored
let f = "1_000_000".parse::<I256>().unwrap();

// But invalid chars are not
assert!("^31".parse::<I256>().is_err());

// Math works great :)
let g = a * b + c - d;

// And so do comparisons!
assert!(e > a);

// We have some useful constants too
assert_eq!(I256::ZERO, I256::unchecked_from(0));
assert_eq!(I256::ONE, I256::unchecked_from(1));
assert_eq!(I256::MINUS_ONE, I256::unchecked_from(-1));

Implementations

impl<const BITS: usize, const LIMBS: usize> Signed<BITS, LIMBS>

pub const fn low_u8(&self) -> u8

Low word.

pub fn as_u8(&self) -> u8

Conversion to u8 with overflow checking.

Panics

Panics if the number is outside the u8 valid range.

pub const fn low_i8(&self) -> i8

Low word.

pub fn as_i8(&self) -> i8

Conversion to i8 with overflow checking.

Panics

Panics if the number is outside the i8 valid range.

pub const fn low_u16(&self) -> u16

Low word.

pub fn as_u16(&self) -> u16

Conversion to u16 with overflow checking.

Panics

Panics if the number is outside the u16 valid range.

pub const fn low_i16(&self) -> i16

Low word.

pub fn as_i16(&self) -> i16

Conversion to i16 with overflow checking.

Panics

Panics if the number is outside the i16 valid range.

pub const fn low_u32(&self) -> u32

Low word.

pub fn as_u32(&self) -> u32

Conversion to u32 with overflow checking.

Panics

Panics if the number is outside the u32 valid range.

pub const fn low_i32(&self) -> i32

Low word.

pub fn as_i32(&self) -> i32

Conversion to i32 with overflow checking.

Panics

Panics if the number is outside the i32 valid range.

pub const fn low_u64(&self) -> u64

Low word.

pub fn as_u64(&self) -> u64

Conversion to u64 with overflow checking.

Panics

Panics if the number is outside the u64 valid range.

pub const fn low_i64(&self) -> i64

Low word.

pub fn as_i64(&self) -> i64

Conversion to i64 with overflow checking.

Panics

Panics if the number is outside the i64 valid range.

pub const fn low_usize(&self) -> usize

Low word.

pub fn as_usize(&self) -> usize

Conversion to usize with overflow checking.

Panics

Panics if the number is outside the usize valid range.

pub const fn low_isize(&self) -> isize

Low word.

pub fn as_isize(&self) -> isize

Conversion to isize with overflow checking.

Panics

Panics if the number is outside the isize valid range.

impl<const BITS: usize, const LIMBS: usize> Signed<BITS, LIMBS>

pub const BITS: usize = BITS

Number of bits.

pub const BYTES: usize = Uint<BITS, LIMBS>::BYTES

The size of this integer type in bytes. Note that some bits may be forced zero if BITS is not cleanly divisible by eight.

pub const MIN: Signed<BITS, LIMBS>

The minimum value.

pub const MAX: Signed<BITS, LIMBS>

The maximum value.

pub const ZERO: Signed<BITS, LIMBS>

Zero (additive identity) of this type.

pub const ONE: Signed<BITS, LIMBS>

One (multiplicative identity) of this type.

pub const MINUS_ONE: Signed<BITS, LIMBS>

Minus one (multiplicative inverse) of this type.

pub const fn from_raw(val: Uint<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Coerces an unsigned integer into a signed one. If the unsigned integer is greater than or equal to 1 << 255, then the result will overflow into a negative value.

pub fn unchecked_from<T>(val: T) -> Signed<BITS, LIMBS>
where T: TryInto<Signed<BITS, LIMBS>>, <T as TryInto<Signed<BITS, LIMBS>>>::Error: Debug,

Shortcut for val.try_into().unwrap().

Panics

Panics if the conversion fails.

pub fn from<T>(value: T) -> Signed<BITS, LIMBS>
where Signed<BITS, LIMBS>: UintTryFrom<T>,

Construct a new Signed from the value.

Panics

Panics if the conversion fails, for example if the value is too large for the bit-size of the Signed. The panic will be attributed to the call site.

pub fn to<T>(&self) -> T
where Signed<BITS, LIMBS>: UintTryTo<T>, T: Debug,

Panics

Panics if the conversion fails, for example if the value is too large for the bit-size of the target type.

pub fn unchecked_into<T>(self) -> T
where Signed<BITS, LIMBS>: TryInto<T>, <Signed<BITS, LIMBS> as TryInto<T>>::Error: Debug,

Shortcut for self.try_into().unwrap().

Panics

Panics if the conversion fails.

pub const fn into_raw(self) -> Uint<BITS, LIMBS>

Returns the signed integer as a unsigned integer. If the value of self negative, then the two’s complement of its absolute value will be returned.

pub const fn sign(&self) -> Sign

Returns the sign of self.

pub const fn is_odd(&self) -> bool

Determines if the integer is odd.

pub const fn const_eq(&self, other: &Signed<BITS, LIMBS>) -> bool

Compile-time equality. NOT constant-time equality.

pub const fn is_zero(&self) -> bool

Returns true if self is zero and false if the number is negative or positive.

pub const fn is_positive(&self) -> bool

Returns true if self is positive and false if the number is zero or negative.

pub const fn is_negative(&self) -> bool

Returns true if self is negative and false if the number is zero or positive.

pub const fn count_ones(&self) -> usize

Returns the number of ones in the binary representation of self.

pub const fn count_zeros(&self) -> usize

Returns the number of zeros in the binary representation of self.

pub const fn leading_zeros(&self) -> usize

Returns the number of leading zeros in the binary representation of self.

pub fn trailing_zeros(&self) -> usize

Returns the number of leading zeros in the binary representation of self.

pub fn trailing_ones(&self) -> usize

Returns the number of leading ones in the binary representation of self.

pub const fn bit(&self, index: usize) -> bool

Returns whether a specific bit is set.

Returns false if index exceeds the bit width of the number.

pub const fn byte(&self, index: usize) -> u8

Returns a specific byte. The byte at index 0 is the least significant byte (little endian).

Panics

Panics if index exceeds the byte width of the number.

pub fn bits(&self) -> u32

Return the least number of bits needed to represent the number.

pub fn overflowing_from_sign_and_abs( sign: Sign, abs: Uint<BITS, LIMBS>, ) -> (Signed<BITS, LIMBS>, bool)

Creates a Signed from a sign and an absolute value. Returns the value and a bool that is true if the conversion caused an overflow.

pub fn checked_from_sign_and_abs( sign: Sign, abs: Uint<BITS, LIMBS>, ) -> Option<Signed<BITS, LIMBS>>

Creates a Signed from an absolute value and a negative flag. Returns None if it would overflow as Signed.

pub fn from_dec_str( value: &str, ) -> Result<Signed<BITS, LIMBS>, ParseSignedError>

Convert from a decimal string.

pub fn to_dec_string(&self) -> String

Convert to a decimal string.

pub fn from_hex_str( value: &str, ) -> Result<Signed<BITS, LIMBS>, ParseSignedError>

Convert from a hex string.

pub fn to_hex_string(&self) -> String

Convert to a hex string.

pub fn into_sign_and_abs(&self) -> (Sign, Uint<BITS, LIMBS>)

Splits a Signed into its absolute value and negative flag.

pub const fn to_be_bytes<const BYTES: usize>(&self) -> [u8; BYTES]

Converts self to a big-endian byte array of size exactly Self::BYTES.

Panics

Panics if the generic parameter BYTES is not exactly Self::BYTES. Ideally this would be a compile time error, but this is blocked by Rust issue #60551.

pub const fn to_le_bytes<const BYTES: usize>(&self) -> [u8; BYTES]

Converts self to a little-endian byte array of size exactly Self::BYTES.

Panics

Panics if the generic parameter BYTES is not exactly Self::BYTES. Ideally this would be a compile time error, but this is blocked by Rust issue #60551.

pub const fn from_be_bytes<const BYTES: usize>( bytes: [u8; BYTES], ) -> Signed<BITS, LIMBS>

Converts a big-endian byte array of size exactly Self::BYTES.

Panics

Panics if the generic parameter BYTES is not exactly Self::BYTES. Ideally this would be a compile time error, but this is blocked by Rust issue #60551.

Panics if the value is too large for the bit-size of the Uint.

pub const fn from_le_bytes<const BYTES: usize>( bytes: [u8; BYTES], ) -> Signed<BITS, LIMBS>

Convert from an array in LE format

Panics

Panics if the given array is not the correct length.

pub fn try_from_be_slice(slice: &[u8]) -> Option<Signed<BITS, LIMBS>>

Creates a new integer from a big endian slice of bytes.

The slice is interpreted as a big endian number. Leading zeros are ignored. The slice can be any length.

Returns None if the value is larger than fits the Uint.

pub fn try_from_le_slice(slice: &[u8]) -> Option<Signed<BITS, LIMBS>>

Creates a new integer from a little endian slice of bytes.

The slice is interpreted as a big endian number. Leading zeros are ignored. The slice can be any length.

Returns None if the value is larger than fits the Uint.

pub const fn as_limbs(&self) -> &[u64; LIMBS]

View the array of limbs.

pub const fn into_limbs(self) -> [u64; LIMBS]

Convert to a array of limbs.

Limbs are least significant first.

pub const fn from_limbs(limbs: [u64; LIMBS]) -> Signed<BITS, LIMBS>

Construct a new integer from little-endian a array of limbs.

Panics

Panics if LIMBS is not equal to nlimbs(BITS).

Panics if the value is to large for the bit-size of the Uint.

pub fn from_base_be<I>( base: u64, digits: I, ) -> Result<Signed<BITS, LIMBS>, BaseConvertError>
where I: IntoIterator<Item = u64>,

Constructs the Signed from digits in the base base in big-endian. Wrapper around ruint’s from_base_be

Errors

BaseConvertError::InvalidBase if the base is less than 2.

BaseConvertError::InvalidDigit if a digit is out of range.

BaseConvertError::Overflow if the number is too large to fit.

impl<const BITS: usize, const LIMBS: usize> Signed<BITS, LIMBS>

pub fn abs(self) -> Signed<BITS, LIMBS>

Computes the absolute value of self.

Overflow behavior

The absolute value of Self::MIN cannot be represented as Self and attempting to calculate it will cause an overflow. This means that code in debug mode will trigger a panic on this case and optimized code will return Self::MIN without a panic.

pub fn overflowing_abs(self) -> (Signed<BITS, LIMBS>, bool)

Computes the absolute value of self.

Returns a tuple of the absolute version of self along with a boolean indicating whether an overflow happened. If self is the minimum value then the minimum value will be returned again and true will be returned for an overflow happening.

pub fn checked_abs(self) -> Option<Signed<BITS, LIMBS>>

Checked absolute value. Computes self.abs(), returning None if self == MIN.

pub fn saturating_abs(self) -> Signed<BITS, LIMBS>

Saturating absolute value. Computes self.abs(), returning MAX if self == MIN instead of overflowing.

pub fn wrapping_abs(self) -> Signed<BITS, LIMBS>

Wrapping absolute value. Computes self.abs(), wrapping around at the boundary of the type.

pub fn unsigned_abs(self) -> Uint<BITS, LIMBS>

Computes the absolute value of self without any wrapping or panicking.

pub fn overflowing_neg(self) -> (Signed<BITS, LIMBS>, bool)

Negates self, overflowing if this is equal to the minimum value.

Returns a tuple of the negated version of self along with a boolean indicating whether an overflow happened. If self is the minimum value, then the minimum value will be returned again and true will be returned for an overflow happening.

pub fn checked_neg(self) -> Option<Signed<BITS, LIMBS>>

Checked negation. Computes -self, returning None if self == MIN.

pub fn saturating_neg(self) -> Signed<BITS, LIMBS>

Saturating negation. Computes -self, returning MAX if self == MIN instead of overflowing.

pub fn wrapping_neg(self) -> Signed<BITS, LIMBS>

Wrapping (modular) negation. Computes -self, wrapping around at the boundary of the type.

The only case where such wrapping can occur is when one negates MIN on a signed type (where MIN is the negative minimal value for the type); this is a positive value that is too large to represent in the type. In such a case, this function returns MIN itself.

pub const fn overflowing_add( self, rhs: Signed<BITS, LIMBS>, ) -> (Signed<BITS, LIMBS>, bool)

Calculates self + rhs

Returns a tuple of the addition along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub const fn checked_add( self, rhs: Signed<BITS, LIMBS>, ) -> Option<Signed<BITS, LIMBS>>

Checked integer addition. Computes self + rhs, returning None if overflow occurred.

pub const fn saturating_add( self, rhs: Signed<BITS, LIMBS>, ) -> Signed<BITS, LIMBS>

Saturating integer addition. Computes self + rhs, saturating at the numeric bounds instead of overflowing.

pub const fn wrapping_add(self, rhs: Signed<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Wrapping (modular) addition. Computes self + rhs, wrapping around at the boundary of the type.

pub const fn overflowing_sub( self, rhs: Signed<BITS, LIMBS>, ) -> (Signed<BITS, LIMBS>, bool)

Calculates self - rhs

Returns a tuple of the subtraction along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub const fn checked_sub( self, rhs: Signed<BITS, LIMBS>, ) -> Option<Signed<BITS, LIMBS>>

Checked integer subtraction. Computes self - rhs, returning None if overflow occurred.

pub const fn saturating_sub( self, rhs: Signed<BITS, LIMBS>, ) -> Signed<BITS, LIMBS>

Saturating integer subtraction. Computes self - rhs, saturating at the numeric bounds instead of overflowing.

pub const fn wrapping_sub(self, rhs: Signed<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Wrapping (modular) subtraction. Computes self - rhs, wrapping around at the boundary of the type.

pub fn overflowing_mul( self, rhs: Signed<BITS, LIMBS>, ) -> (Signed<BITS, LIMBS>, bool)

Calculates self * rhs

Returns a tuple of the multiplication along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub fn checked_mul( self, rhs: Signed<BITS, LIMBS>, ) -> Option<Signed<BITS, LIMBS>>

Checked integer multiplication. Computes self * rhs, returning None if overflow occurred.

pub fn saturating_mul(self, rhs: Signed<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Saturating integer multiplication. Computes self * rhs, saturating at the numeric bounds instead of overflowing.

pub fn wrapping_mul(self, rhs: Signed<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Wrapping (modular) multiplication. Computes self * rhs, wrapping around at the boundary of the type.

pub fn overflowing_div( self, rhs: Signed<BITS, LIMBS>, ) -> (Signed<BITS, LIMBS>, bool)

Calculates self / rhs

Returns a tuple of the divisor along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then self is returned.

Panics

If rhs is 0.

pub fn checked_div( self, rhs: Signed<BITS, LIMBS>, ) -> Option<Signed<BITS, LIMBS>>

Checked integer division. Computes self / rhs, returning None if rhs == 0 or the division results in overflow.

pub fn saturating_div(self, rhs: Signed<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Saturating integer division. Computes self / rhs, saturating at the numeric bounds instead of overflowing.

Panics

If rhs is 0.

pub fn wrapping_div(self, rhs: Signed<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Wrapping (modular) division. Computes self / rhs, wrapping around at the boundary of the type.

The only case where such wrapping can occur is when one divides MIN / -1 on a signed type (where MIN is the negative minimal value for the type); this is equivalent to -MIN, a positive value that is too large to represent in the type. In such a case, this function returns MIN itself.

Panics

If rhs is 0.

pub fn overflowing_rem( self, rhs: Signed<BITS, LIMBS>, ) -> (Signed<BITS, LIMBS>, bool)

Calculates self % rhs

Returns a tuple of the remainder after dividing along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then 0 is returned.

Panics

If rhs is 0.

pub fn checked_rem( self, rhs: Signed<BITS, LIMBS>, ) -> Option<Signed<BITS, LIMBS>>

Checked integer remainder. Computes self % rhs, returning None if rhs == 0 or the division results in overflow.

pub fn wrapping_rem(self, rhs: Signed<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Wrapping (modular) remainder. Computes self % rhs, wrapping around at the boundary of the type.

Such wrap-around never actually occurs mathematically; implementation artifacts make x % y invalid for MIN / -1 on a signed type (where MIN is the negative minimal value). In such a case, this function returns 0.

Panics

If rhs is 0.

pub fn div_euclid(self, rhs: Signed<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Calculates the quotient of Euclidean division of self by rhs.

This computes the integer q such that self = q * rhs + r, with r = self.rem_euclid(rhs) and 0 <= r < abs(rhs).

In other words, the result is self / rhs rounded to the integer q such that self >= q * rhs. If self > 0, this is equal to round towards zero (the default in Rust); if self < 0, this is equal to round towards +/- infinity.

Panics

If rhs is 0 or the division results in overflow.

pub fn overflowing_div_euclid( self, rhs: Signed<BITS, LIMBS>, ) -> (Signed<BITS, LIMBS>, bool)

Calculates the quotient of Euclidean division self.div_euclid(rhs).

Returns a tuple of the divisor along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then self is returned.

Panics

If rhs is 0.

pub fn checked_div_euclid( self, rhs: Signed<BITS, LIMBS>, ) -> Option<Signed<BITS, LIMBS>>

Checked Euclidean division. Computes self.div_euclid(rhs), returning None if rhs == 0 or the division results in overflow.

pub fn wrapping_div_euclid( self, rhs: Signed<BITS, LIMBS>, ) -> Signed<BITS, LIMBS>

Wrapping Euclidean division. Computes self.div_euclid(rhs), wrapping around at the boundary of the type.

Wrapping will only occur in MIN / -1 on a signed type (where MIN is the negative minimal value for the type). This is equivalent to -MIN, a positive value that is too large to represent in the type. In this case, this method returns MIN itself.

Panics

If rhs is 0.

pub fn rem_euclid(self, rhs: Signed<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Calculates the least nonnegative remainder of self (mod rhs).

This is done as if by the Euclidean division algorithm – given r = self.rem_euclid(rhs), self = rhs * self.div_euclid(rhs) + r, and 0 <= r < abs(rhs).

Panics

If rhs is 0 or the division results in overflow.

pub fn overflowing_rem_euclid( self, rhs: Signed<BITS, LIMBS>, ) -> (Signed<BITS, LIMBS>, bool)

Overflowing Euclidean remainder. Calculates self.rem_euclid(rhs).

Returns a tuple of the remainder after dividing along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then 0 is returned.

Panics

If rhs is 0.

pub fn wrapping_rem_euclid( self, rhs: Signed<BITS, LIMBS>, ) -> Signed<BITS, LIMBS>

Wrapping Euclidean remainder. Computes self.rem_euclid(rhs), wrapping around at the boundary of the type.

Wrapping will only occur in MIN % -1 on a signed type (where MIN is the negative minimal value for the type). In this case, this method returns 0.

Panics

If rhs is 0.

pub fn checked_rem_euclid( self, rhs: Signed<BITS, LIMBS>, ) -> Option<Signed<BITS, LIMBS>>

Checked Euclidean remainder. Computes self.rem_euclid(rhs), returning None if rhs == 0 or the division results in overflow.

pub fn exp10(n: usize) -> Signed<BITS, LIMBS>

Create 10**n as this type.

Panics

If the result overflows the type.

pub fn pow(self, exp: Uint<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Raises self to the power of exp, using exponentiation by squaring.

Panics

If the result overflows the type in debug mode.

pub fn overflowing_pow( self, exp: Uint<BITS, LIMBS>, ) -> (Signed<BITS, LIMBS>, bool)

Raises self to the power of exp, using exponentiation by squaring.

Returns a tuple of the exponentiation along with a bool indicating whether an overflow happened.

pub fn checked_pow(self, exp: Uint<BITS, LIMBS>) -> Option<Signed<BITS, LIMBS>>

Checked exponentiation. Computes self.pow(exp), returning None if overflow occurred.

pub fn saturating_pow(self, exp: Uint<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Saturating integer exponentiation. Computes self.pow(exp), saturating at the numeric bounds instead of overflowing.

pub fn wrapping_pow(self, exp: Uint<BITS, LIMBS>) -> Signed<BITS, LIMBS>

Raises self to the power of exp, wrapping around at the boundary of the type.

pub fn overflowing_shl(self, rhs: usize) -> (Signed<BITS, LIMBS>, bool)

Shifts self left by rhs bits.

Returns a tuple of the shifted version of self along with a boolean indicating whether the shift value was larger than or equal to the number of bits.

pub fn checked_shl(self, rhs: usize) -> Option<Signed<BITS, LIMBS>>

Checked shift left. Computes self << rhs, returning None if rhs is larger than or equal to the number of bits in self.

pub fn wrapping_shl(self, rhs: usize) -> Signed<BITS, LIMBS>

Wrapping shift left. Computes self << rhs, returning 0 if larger than or equal to the number of bits in self.

pub fn overflowing_shr(self, rhs: usize) -> (Signed<BITS, LIMBS>, bool)

Shifts self right by rhs bits.

Returns a tuple of the shifted version of self along with a boolean indicating whether the shift value was larger than or equal to the number of bits.

pub fn checked_shr(self, rhs: usize) -> Option<Signed<BITS, LIMBS>>

Checked shift right. Computes self >> rhs, returning None if rhs is larger than or equal to the number of bits in self.

pub fn wrapping_shr(self, rhs: usize) -> Signed<BITS, LIMBS>

Wrapping shift right. Computes self >> rhs, returning 0 if larger than or equal to the number of bits in self.

pub fn asr(self, rhs: usize) -> Signed<BITS, LIMBS>

Arithmetic shift right operation. Computes self >> rhs maintaining the original sign. If the number is positive this is the same as logic shift right.

pub fn asl(self, rhs: usize) -> Option<Signed<BITS, LIMBS>>

Arithmetic shift left operation. Computes self << rhs, checking for overflow on the final result.

Returns None if the operation overflowed (most significant bit changes).

pub fn twos_complement(self) -> Uint<BITS, LIMBS>

Compute the two’s complement of this number.