yet_another_parachain_runtime::pallet_revive::precompiles::alloy::sol_types::abi

Trait Token

pub trait Token<'de>: Sized + Sealed {
    const DYNAMIC: bool;

    // Required methods
    fn decode_from(dec: &mut Decoder<'de>) -> Result<Self, Error>;
    fn head_words(&self) -> usize;
    fn tail_words(&self) -> usize;
    fn head_append(&self, enc: &mut Encoder);
    fn tail_append(&self, enc: &mut Encoder);

    // Provided method
    fn total_words(&self) -> usize { ... }
}

Ethereum ABI tokens.

Tokens are an intermediate state between ABI-encoded blobs, and Rust types.

ABI encoding uses 5 types:

WordToken: Single EVM words (a 32-byte string)

FixedSeqToken: Sequences with a fixed length T[M]

DynSeqToken: Sequences with a dynamic length T[]

PackedSeqToken: Dynamic-length byte arrays bytes or string

Tuples (T, U, V, ...) (implemented for arity 0..=24)

A token with a lifetime borrows its data from elsewhere. During decoding, it borrows its data from the decoder. During encoding, it borrows its data from the Rust value being encoded.

This trait allows us to encode and decode data with minimal copying. It may also be used to enable zero-copy decoding of data, or fast transformation of encoded blobs without full decoding.

This trait is sealed and cannot be implemented for types outside of this crate. It is implemented only for the types listed above.

Required Associated Constants

const DYNAMIC: bool

True if the token represents a dynamically-sized type.

Required Methods

fn decode_from(dec: &mut Decoder<'de>) -> Result<Self, Error>

Decode a token from a decoder.

fn head_words(&self) -> usize

Calculate the number of head words.

fn tail_words(&self) -> usize

Calculate the number of tail words.

fn head_append(&self, enc: &mut Encoder)

Append head words to the encoder.

fn tail_append(&self, enc: &mut Encoder)

Append tail words to the encoder.

Provided Methods

fn total_words(&self) -> usize

Calculate the total number of head and tail words.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.