yet_another_parachain_runtime::pallet_revive::precompiles::alloy::sol_types::abi

Trait TokenSeq

pub trait TokenSeq<'a>: Token<'a> {
    const IS_TUPLE: bool = false;

    // Required methods
    fn encode_sequence(&self, enc: &mut Encoder);
    fn decode_sequence(dec: &mut Decoder<'a>) -> Result<Self, Error>;
}

A token composed of a sequence of other tokens.

This functions is an extension trait for Token, and is only implemented by FixedSeqToken, DynSeqToken, PackedSeqToken, and tuples of Tokens (including WordToken).

Provided Associated Constants

const IS_TUPLE: bool = false

True for tuples only.

Required Methods

fn encode_sequence(&self, enc: &mut Encoder)

ABI-encode the token sequence into the encoder.

fn decode_sequence(dec: &mut Decoder<'a>) -> Result<Self, Error>

ABI-decode the token sequence from the encoder.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.