yet_another_parachain_runtime::pallet_revive::precompiles::alloy::sol_types::sol_data

Trait SupportedInt

pub trait SupportedInt: Sealed {
    type Int: Copy + PartialOrd + Ord + Eq + Hash + Not + BitAnd + BitOr + BitXor + Add + Sub + Mul + Div + Rem + AddAssign + SubAssign + MulAssign + DivAssign + RemAssign + Debug + Display + LowerHex + UpperHex + Octal + Binary;
    type Uint: Copy + PartialOrd + Ord + Eq + Hash + Not + BitAnd + BitOr + BitXor + Add + Sub + Mul + Div + Rem + AddAssign + SubAssign + MulAssign + DivAssign + RemAssign + Debug + Display + LowerHex + UpperHex + Octal + Binary;

    const INT_NAME: &'static str;
    const UINT_NAME: &'static str;
    const BITS: usize;
    const SKIP_BYTES: usize;
    const BYTES: usize = _;
    const WORD_MSB: usize = _;

    // Required methods
    fn tokenize_int(int: Self::Int) -> WordToken;
    fn detokenize_int(token: WordToken) -> Self::Int;
    fn encode_packed_to_int(int: Self::Int, out: &mut Vec<u8>);
    fn tokenize_uint(uint: Self::Uint) -> WordToken;
    fn detokenize_uint(token: WordToken) -> Self::Uint;
    fn encode_packed_to_uint(uint: Self::Uint, out: &mut Vec<u8>);
}

Statically guarantees that a Int or Uint bit count is marked as supported.

This trait is sealed: the list of implementors below is total.

Users do not have the ability to mark additional IntBitCount<N> values as supported. Only Int and Uint with supported byte counts are constructable.

Required Associated Constants

const INT_NAME: &'static str

The name of the Int type: int<N>

const UINT_NAME: &'static str

The name of the Uint type: uint<N>

const BITS: usize

The number of bits in the integer: BITS

Note that this is not equal to Self::Int::BITS.

const SKIP_BYTES: usize

The difference between the representation’s and this integer’s bytes: (Self::Int::BITS - Self::BITS) / 8

E.g.: word[Self::WORD_MSB - Self::SKIP_BYTES..] == int.to_be_bytes()

Provided Associated Constants

const BYTES: usize = _

The number of bytes in the integer: BITS / 8

const WORD_MSB: usize = _

The index of the most significant byte in the Word type.

E.g.: word[Self::WORD_MSB..] == int.to_be_bytes()[Self::SKIP_BYTES..]

Required Associated Types

type Int: Copy + PartialOrd + Ord + Eq + Hash + Not + BitAnd + BitOr + BitXor + Add + Sub + Mul + Div + Rem + AddAssign + SubAssign + MulAssign + DivAssign + RemAssign + Debug + Display + LowerHex + UpperHex + Octal + Binary

The signed integer Rust representation.

type Uint: Copy + PartialOrd + Ord + Eq + Hash + Not + BitAnd + BitOr + BitXor + Add + Sub + Mul + Div + Rem + AddAssign + SubAssign + MulAssign + DivAssign + RemAssign + Debug + Display + LowerHex + UpperHex + Octal + Binary

The unsigned integer Rust representation.

Required Methods

fn tokenize_int(int: Self::Int) -> WordToken

Tokenizes a signed integer.

fn detokenize_int(token: WordToken) -> Self::Int

Detokenizes a signed integer.

fn encode_packed_to_int(int: Self::Int, out: &mut Vec<u8>)

ABI-encode a signed integer in packed mode.

fn tokenize_uint(uint: Self::Uint) -> WordToken

Tokenizes an unsigned integer.

fn detokenize_uint(token: WordToken) -> Self::Uint

Detokenizes an unsigned integer.

fn encode_packed_to_uint(uint: Self::Uint, out: &mut Vec<u8>)

ABI-encode an unsigned integer in packed mode.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.