yet_another_parachain_runtime::pallet_revive::precompiles::alloy::sol_types

Trait SolValue

pub trait SolValue: SolTypeValue<Self::SolType> {
    type SolType: SolType;

    // Provided methods
    fn sol_name(&self) -> &'static str { ... }
    fn tokenize(&self) -> <Self::SolType as SolType>::Token<'_> { ... }
    fn detokenize(token: <Self::SolType as SolType>::Token<'_>) -> Self
       where Self: From<<Self::SolType as SolType>::RustType> { ... }
    fn abi_encoded_size(&self) -> usize { ... }
    fn eip712_data_word(&self) -> FixedBytes<32> { ... }
    fn abi_encode_packed_to(&self, out: &mut Vec<u8>) { ... }
    fn abi_encode_packed(&self) -> Vec<u8>  { ... }
    fn abi_encode(&self) -> Vec<u8>  { ... }
    fn abi_encode_sequence(&self) -> Vec<u8> 
       where <Self::SolType as SolType>::Token<'a>: for<'a> TokenSeq<'a> { ... }
    fn abi_encode_params(&self) -> Vec<u8> 
       where <Self::SolType as SolType>::Token<'a>: for<'a> TokenSeq<'a> { ... }
    fn abi_decode(data: &[u8]) -> Result<Self, Error>
       where Self: From<<Self::SolType as SolType>::RustType> { ... }
    fn abi_decode_validate(data: &[u8]) -> Result<Self, Error>
       where Self: From<<Self::SolType as SolType>::RustType> { ... }
    fn abi_decode_params<'de>(data: &'de [u8]) -> Result<Self, Error>
       where Self: From<<Self::SolType as SolType>::RustType>,
             <Self::SolType as SolType>::Token<'de>: TokenSeq<'de> { ... }
    fn abi_decode_params_validate<'de>(data: &'de [u8]) -> Result<Self, Error>
       where Self: From<<Self::SolType as SolType>::RustType>,
             <Self::SolType as SolType>::Token<'de>: TokenSeq<'de> { ... }
    fn abi_decode_sequence<'de>(data: &'de [u8]) -> Result<Self, Error>
       where Self: From<<Self::SolType as SolType>::RustType>,
             <Self::SolType as SolType>::Token<'de>: TokenSeq<'de> { ... }
    fn abi_decode_sequence_validate<'de>(data: &'de [u8]) -> Result<Self, Error>
       where Self: From<<Self::SolType as SolType>::RustType>,
             <Self::SolType as SolType>::Token<'de>: TokenSeq<'de> { ... }

}

A Solidity value.

This is a convenience trait that re-exports the logic in SolType with less generic implementations so that they can be used as methods with self receivers.

See SolType for more information.

Implementer’s Guide

It should not be necessary to implement this trait manually. Instead, use the sol! procedural macro to parse Solidity syntax into types that implement this trait.

Examples

use alloy_sol_types::SolValue;

let my_values = ("hello", 0xdeadbeef_u32, true, [0x42_u8; 24]);
let _ = my_values.abi_encode();
let _ = my_values.abi_encode_packed();
assert_eq!(my_values.sol_name(), "(string,uint32,bool,bytes24)");

Required Associated Types

type SolType: SolType

The Solidity type that this type corresponds to.

Provided Methods

fn sol_name(&self) -> &'static str

The name of the associated Solidity type.

See SolType::SOL_NAME for more information.

fn tokenize(&self) -> <Self::SolType as SolType>::Token<'_>

Tokenizes the given value into this type’s token.

See SolType::tokenize for more information.

fn detokenize(token: <Self::SolType as SolType>::Token<'_>) -> Self
where Self: From<<Self::SolType as SolType>::RustType>,

Detokenize a value from the given token.

See SolType::detokenize for more information.

fn abi_encoded_size(&self) -> usize

Calculate the ABI-encoded size of the data.

See SolType::abi_encoded_size for more information.

fn eip712_data_word(&self) -> FixedBytes<32>

Encode this data according to EIP-712 encodeData rules, and hash it if necessary.

See SolType::eip712_data_word for more information.

fn abi_encode_packed_to(&self, out: &mut Vec<u8>)

Non-standard Packed Mode ABI encoding.

See SolType::abi_encode_packed_to for more information.

fn abi_encode_packed(&self) -> Vec<u8>

Non-standard Packed Mode ABI encoding.

See SolType::abi_encode_packed for more information.

fn abi_encode(&self) -> Vec<u8>

ABI-encodes the value.

See SolType::abi_encode for more information.

fn abi_encode_sequence(&self) -> Vec<u8>
where <Self::SolType as SolType>::Token<'a>: for<'a> TokenSeq<'a>,

Encodes an ABI sequence.

See SolType::abi_encode_sequence for more information.

fn abi_encode_params(&self) -> Vec<u8>
where <Self::SolType as SolType>::Token<'a>: for<'a> TokenSeq<'a>,

Encodes an ABI sequence suitable for function parameters.

See SolType::abi_encode_params for more information.

fn abi_decode(data: &[u8]) -> Result<Self, Error>
where Self: From<<Self::SolType as SolType>::RustType>,

ABI-decode this type from the given data.

See SolType::abi_decode for more information.

fn abi_decode_validate(data: &[u8]) -> Result<Self, Error>
where Self: From<<Self::SolType as SolType>::RustType>,

ABI-decode this type from the given data, with validation.

See SolType::abi_decode_validate for more information.

fn abi_decode_params<'de>(data: &'de [u8]) -> Result<Self, Error>
where Self: From<<Self::SolType as SolType>::RustType>, <Self::SolType as SolType>::Token<'de>: TokenSeq<'de>,

ABI-decode this type from the given data.

See SolType::abi_decode_params for more information.

fn abi_decode_params_validate<'de>(data: &'de [u8]) -> Result<Self, Error>
where Self: From<<Self::SolType as SolType>::RustType>, <Self::SolType as SolType>::Token<'de>: TokenSeq<'de>,

ABI-decode this type from the given data, with validation.

See SolType::abi_decode_params_validate for more information.

fn abi_decode_sequence<'de>(data: &'de [u8]) -> Result<Self, Error>
where Self: From<<Self::SolType as SolType>::RustType>, <Self::SolType as SolType>::Token<'de>: TokenSeq<'de>,

ABI-decode this type from the given data.

See SolType::abi_decode_sequence for more information.

fn abi_decode_sequence_validate<'de>(data: &'de [u8]) -> Result<Self, Error>
where Self: From<<Self::SolType as SolType>::RustType>, <Self::SolType as SolType>::Token<'de>: TokenSeq<'de>,

ABI-decode this type from the given data, with validation.

See SolType::abi_decode_sequence_validate for more information.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.