yet_another_parachain_runtime::pallet_revive

Struct Pallet

pub struct Pallet<T>(/* private fields */);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T> Pallet<T>
where T: Config,

pub fn checking_account() -> <T as Config>::AccountId

impl<T> Pallet<T>
where T: Config,

pub fn eth_transact( origin: <T as Config>::RuntimeOrigin, payload: Vec<u8>, ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

A raw EVM transaction, typically dispatched by an Ethereum JSON-RPC server.

Parameters

payload: The encoded crate::evm::TransactionSigned.

Note

This call cannot be dispatched directly; attempting to do so will result in a failed transaction. It serves as a wrapper for an Ethereum transaction. When submitted, the runtime converts it into a sp_runtime::generic::CheckedExtrinsic by recovering the signer and validating the transaction.

pub fn call( origin: <T as Config>::RuntimeOrigin, dest: H160, value: <T as Config>::Balance, weight_limit: Weight, storage_deposit_limit: <T as Config>::Balance, data: Vec<u8>, ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Makes a call to an account, optionally transferring some balance.

Parameters

dest: Address of the contract to call.

value: The balance to transfer from the origin to dest.

weight_limit: The weight limit enforced when executing the constructor.

storage_deposit_limit: The maximum amount of balance that can be charged from the caller to pay for the storage consumed.

data: The input data to pass to the contract.

If the account is a smart-contract account, the associated code will be executed and any value will be transferred.

If the account is a regular account, any value will be transferred.

If no account exists and the call value is not less than existential_deposit, a regular account will be created and any value will be transferred.

pub fn instantiate( origin: <T as Config>::RuntimeOrigin, value: <T as Config>::Balance, weight_limit: Weight, storage_deposit_limit: <T as Config>::Balance, code_hash: H256, data: Vec<u8>, salt: Option<[u8; 32]>, ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Instantiates a contract from a previously deployed vm binary.

This function is identical to Self::instantiate_with_code but without the code deployment step. Instead, the code_hash of an on-chain deployed vm binary must be supplied.

pub fn instantiate_with_code( origin: <T as Config>::RuntimeOrigin, value: <T as Config>::Balance, weight_limit: Weight, storage_deposit_limit: <T as Config>::Balance, code: Vec<u8>, data: Vec<u8>, salt: Option<[u8; 32]>, ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Instantiates a new contract from the supplied code optionally transferring some balance.

This dispatchable has the same effect as calling Self::upload_code + Self::instantiate. Bundling them together provides efficiency gains. Please also check the documentation of Self::upload_code.

Parameters

value: The balance to transfer from the origin to the newly created contract.

weight_limit: The weight limit enforced when executing the constructor.

storage_deposit_limit: The maximum amount of balance that can be charged/reserved from the caller to pay for the storage consumed.

code: The contract code to deploy in raw bytes.

data: The input data to pass to the contract constructor.

salt: Used for the address derivation. If Some is supplied then CREATE2 semantics are used. If None then CRATE1 is used.

Instantiation is executed as follows:

The supplied code is deployed, and a code_hash is created for that code.

If the code_hash already exists on the chain the underlying code will be shared.

The destination address is computed based on the sender, code_hash and the salt.

The smart-contract account is created at the computed address.

The value is transferred to the new account.

The deploy function is executed in the context of the newly-created account.

pub fn eth_instantiate_with_code( origin: <T as Config>::RuntimeOrigin, value: U256, weight_limit: Weight, eth_gas_limit: U256, code: Vec<u8>, data: Vec<u8>, transaction_encoded: Vec<u8>, effective_gas_price: U256, encoded_len: u32, ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Same as Self::instantiate_with_code, but intended to be dispatched only by an EVM transaction through the EVM compatibility layer.

Parameters

value: The balance to transfer from the origin to the newly created contract.

weight_limit: The gas limit used to derive the transaction weight for transaction payment

eth_gas_limit: The Ethereum gas limit governing the resource usage of the execution

code: The contract code to deploy in raw bytes.

data: The input data to pass to the contract constructor.

transaction_encoded: The RLP encoding of the signed Ethereum transaction, represented as crate::evm::TransactionSigned, provided by the Ethereum wallet. This is used for building the Ethereum transaction root.

effective_gas_price: the price of a unit of gas

encoded len: the byte code size of the eth_transact extrinsic

Calling this dispatchable ensures that the origin’s nonce is bumped only once, via the CheckNonce transaction extension. In contrast, Self::instantiate_with_code also bumps the nonce after contract instantiation, since it may be invoked multiple times within a batch call transaction.

pub fn eth_call( origin: <T as Config>::RuntimeOrigin, dest: H160, value: U256, weight_limit: Weight, eth_gas_limit: U256, data: Vec<u8>, transaction_encoded: Vec<u8>, effective_gas_price: U256, encoded_len: u32, ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Same as Self::call, but intended to be dispatched only by an EVM transaction through the EVM compatibility layer.

Parameters

dest: The Ethereum address of the account to be called

value: The balance to transfer from the origin to the newly created contract.

weight_limit: The gas limit used to derive the transaction weight for transaction payment

eth_gas_limit: The Ethereum gas limit governing the resource usage of the execution

data: The input data to pass to the contract constructor.

transaction_encoded: The RLP encoding of the signed Ethereum transaction, represented as crate::evm::TransactionSigned, provided by the Ethereum wallet. This is used for building the Ethereum transaction root.

effective_gas_price: the price of a unit of gas

encoded len: the byte code size of the eth_transact extrinsic

pub fn eth_substrate_call( origin: <T as Config>::RuntimeOrigin, call: Box<<T as Config>::RuntimeCall>, transaction_encoded: Vec<u8>, ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Executes a Substrate runtime call from an Ethereum transaction.

This dispatchable is intended to be called only through the EVM compatibility layer. The provided call will be dispatched using RawOrigin::Signed.

Parameters

origin: Must be an Origin::EthTransaction origin.

call: The Substrate runtime call to execute.

transaction_encoded: The RLP encoding of the Ethereum transaction,

pub fn upload_code( origin: <T as Config>::RuntimeOrigin, code: Vec<u8>, storage_deposit_limit: <T as Config>::Balance, ) -> Result<(), DispatchError>

Upload new code without instantiating a contract from it.

If the code does not already exist a deposit is reserved from the caller The size of the reserve depends on the size of the supplied code.

Note

Anyone can instantiate a contract from any uploaded code and thus prevent its removal. To avoid this situation a constructor could employ access control so that it can only be instantiated by permissioned entities. The same is true when uploading through Self::instantiate_with_code.

If the refcount of the code reaches zero after terminating the last contract that references this code, the code will be removed automatically.

pub fn remove_code( origin: <T as Config>::RuntimeOrigin, code_hash: H256, ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Remove the code stored under code_hash and refund the deposit to its owner.

A code can only be removed by its original uploader (its owner) and only if it is not used by any contract.

pub fn set_code( origin: <T as Config>::RuntimeOrigin, dest: H160, code_hash: H256, ) -> Result<(), DispatchError>

Privileged function that changes the code of an existing contract.

This takes care of updating refcounts and all other necessary operations. Returns an error if either the code_hash or dest do not exist.

Note

This does not change the address of the contract in question. This means that the contract address is no longer derived from its code hash after calling this dispatchable.

pub fn map_account( origin: <T as Config>::RuntimeOrigin, ) -> Result<(), DispatchError>

Register the callers account id so that it can be used in contract interactions.

This will error if the origin is already mapped or is a eth native Address20. It will take a deposit that can be released by calling Self::unmap_account.

pub fn unmap_account( origin: <T as Config>::RuntimeOrigin, ) -> Result<(), DispatchError>

Unregister the callers account id in order to free the deposit.

There is no reason to ever call this function other than freeing up the deposit. This is only useful when the account should no longer be used.

pub fn dispatch_as_fallback_account( origin: <T as Config>::RuntimeOrigin, call: Box<<T as Config>::RuntimeCall>, ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Dispatch an call with the origin set to the callers fallback address.

Every AccountId32 can control its corresponding fallback account. The fallback account is the AccountId20 with the last 12 bytes set to 0xEE. This is essentially a recovery function in case an AccountId20 was used without creating a mapping first.

impl<T> Pallet<T>
where T: Config,

pub fn bare_call( origin: <T as Config>::RuntimeOrigin, dest: H160, evm_value: U256, transaction_limits: TransactionLimits<T>, data: Vec<u8>, exec_config: ExecConfig<T>, ) -> ContractResult<ExecReturnValue, <T as Config>::Balance>

A generalized version of Self::call.

Identical to Self::call but tailored towards being called by other code within the runtime as opposed to from an extrinsic. It returns more information and allows the enablement of features that are not suitable for an extrinsic (debugging, event collection).

pub fn prepare_dry_run(account: &<T as Config>::AccountId)

Prepare a dry run for the given account.

This function is public because it is called by the runtime API implementation (see impl_runtime_apis_plus_revive).

pub fn bare_instantiate( origin: <T as Config>::RuntimeOrigin, evm_value: U256, transaction_limits: TransactionLimits<T>, code: Code, data: Vec<u8>, salt: Option<[u8; 32]>, exec_config: ExecConfig<T>, ) -> ContractResult<InstantiateReturnValue, <T as Config>::Balance>

A generalized version of Self::instantiate or Self::instantiate_with_code.

Identical to Self::instantiate or Self::instantiate_with_code but tailored towards being called by other code within the runtime as opposed to from an extrinsic. It returns more information to the caller useful to estimate the cost of the operation.

pub fn dry_run_eth_transact( tx: GenericTransaction, dry_run_config: DryRunConfig<<<T as Config>::Time as Time>::Moment>, ) -> Result<EthTransactInfo<<T as Config>::Balance>, EthTransactError>
where <T as Config>::Nonce: Into<U256>, <T as Config>::RuntimeCall: SetWeightLimit,

Dry-run Ethereum calls.

Parameters

tx: The Ethereum transaction to simulate.

pub fn evm_balance(address: &H160) -> U256

Get the balance with EVM decimals of the given address.

Returns the spendable balance excluding the existential deposit.

pub fn eth_block() -> Block

Get the current Ethereum block from storage.

pub fn eth_block_hash_from_number(number: U256) -> Option<H256>

Convert the Ethereum block number into the Ethereum block hash.

Note

The Ethereum block number is identical to the Substrate block number. If the provided block number is outside of the pruning None is returned.

pub fn eth_receipt_data() -> Vec<ReceiptGasInfo>

The details needed to reconstruct the receipt information offchain.

pub fn set_evm_balance(address: &H160, evm_value: U256) -> Result<(), Error<T>>

Set the EVM balance of an account.

The account’s total balance becomes the EVM value plus the existential deposit, consistent with evm_balance which returns the spendable balance excluding the existential deposit.

pub fn new_balance_with_dust( evm_value: U256, ) -> Result<(<T as Config>::Balance, u32), BalanceConversionError>

Construct native balance from EVM balance.

Adds the existential deposit and returns the native balance plus the dust.

pub fn evm_nonce(address: &H160) -> u32
where <T as Config>::Nonce: Into<u32>,

Get the nonce for the given address.

pub fn evm_block_gas_limit() -> U256

Get the block gas limit.

pub fn evm_max_extrinsic_weight() -> Weight

The maximum weight an eth_transact is allowed to consume.

pub fn evm_base_fee() -> U256

Get the base gas price.

pub fn evm_tracer(tracer_type: TracerType) -> Tracer<T>
where <T as Config>::Nonce: Into<u32>,

Build an EVM tracer from the given tracer type.

pub fn bare_upload_code( origin: <T as Config>::RuntimeOrigin, code: Vec<u8>, storage_deposit_limit: <T as Config>::Balance, ) -> Result<CodeUploadReturnValue<<T as Config>::Balance>, DispatchError>

A generalized version of Self::upload_code.

It is identical to Self::upload_code and only differs in the information it returns.

pub fn get_storage( address: H160, key: [u8; 32], ) -> Result<Option<Vec<u8>>, ContractAccessError>

Query storage of a specified contract under a specified key.

pub fn get_immutables( address: H160, ) -> Option<BoundedVec<u8, ConstU32<pallet_revive::::ImmutableData::{constant#0}>>>

Get the immutable data of a specified contract.

Returns None if the contract does not exist or has no immutable data.

pub fn set_immutables( address: H160, data: BoundedVec<u8, ConstU32<pallet_revive::::ImmutableData::{constant#0}>>, ) -> Result<(), ContractAccessError>

Sets immutable data of a contract

Returns an error if the contract does not exist.

Warning

Does not collect any storage deposit. Not safe to be called by user controlled code.

pub fn get_storage_var_key( address: H160, key: Vec<u8>, ) -> Result<Option<Vec<u8>>, ContractAccessError>

Query storage of a specified contract under a specified variable-sized key.

pub fn convert_native_to_evm( value: impl Into<BalanceWithDust<<T as Config>::Balance>>, ) -> U256

Convert a native balance to EVM balance.

pub fn set_storage( address: H160, key: [u8; 32], value: Option<Vec<u8>>, ) -> Result<WriteOutcome, ContractAccessError>

Set storage of a specified contract under a specified key.

If the value is None, the storage entry is deleted.

Returns an error if the contract does not exist or if the write operation fails.

Warning

Does not collect any storage deposit. Not safe to be called by user controlled code.

pub fn set_storage_var_key( address: H160, key: Vec<u8>, value: Option<Vec<u8>>, ) -> Result<WriteOutcome, ContractAccessError>

Set the storage of a specified contract under a specified variable-sized key.

If the value is None, the storage entry is deleted.

Returns an error if the contract does not exist, if the key decoding fails, or if the write operation fails.

Warning

Does not collect any storage deposit. Not safe to be called by user controlled code.

pub fn account_id() -> <T as Config>::AccountId

Pallet account, used to hold funds for contracts upload deposit.

pub fn block_author() -> H160

The address of the validator that produced the current block.

pub fn code(address: &H160) -> Vec<u8>

Returns the code at address.

This takes pre-compiles into account.

pub fn try_upload_code( origin: <T as Config>::AccountId, code: Vec<u8>, code_type: BytecodeType, meter: &mut ResourceMeter<T, Root>, exec_config: &ExecConfig<T>, ) -> Result<ContractBlob<T>, DispatchError>

Uploads new code and returns the Vm binary contract blob and deposit amount collected.