yet_another_parachain_runtime::pallet_revive

Trait Executable

pub trait Executable<T>: Sized
where
    T: Config,
{
    // Required methods
    fn from_storage<S>(
        code_hash: H256,
        meter: &mut ResourceMeter<T, S>,
    ) -> Result<Self, DispatchError>
       where S: State;
    fn from_evm_init_code(
        code: Vec<u8>,
        owner: <T as Config>::AccountId,
    ) -> Result<Self, DispatchError>;
    fn execute<E>(
        self,
        ext: &mut E,
        function: ExportedFunction,
        input_data: Vec<u8>,
    ) -> Result<ExecReturnValue, ExecError>
       where E: Ext<T = T>;
    fn code_info(&self) -> &CodeInfo<T>;
    fn code(&self) -> &[u8] ;
    fn code_hash(&self) -> &H256;
}

A trait that represents something that can be executed.

In the on-chain environment this would be represented by a vm binary module. This trait exists in order to be able to mock the vm logic for testing.

Required Methods

fn from_storage<S>( code_hash: H256, meter: &mut ResourceMeter<T, S>, ) -> Result<Self, DispatchError>
where S: State,

Load the executable from storage.

Note

Charges size base load weight from the weight meter.

fn from_evm_init_code( code: Vec<u8>, owner: <T as Config>::AccountId, ) -> Result<Self, DispatchError>

Load the executable from EVM bytecode

fn execute<E>( self, ext: &mut E, function: ExportedFunction, input_data: Vec<u8>, ) -> Result<ExecReturnValue, ExecError>
where E: Ext<T = T>,

Execute the specified exported function and return the result.

When the specified function is Constructor the executable is stored and its refcount incremented.

Note

This functions expects to be executed in a storage transaction that rolls back all of its emitted storage changes.

fn code_info(&self) -> &CodeInfo<T>

The code info of the executable.

fn code(&self) -> &[u8]

The raw code of the executable.

fn code_hash(&self) -> &H256

The code hash of the executable.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.