yet_another_parachain_runtime::pallet_salary

Trait Config

pub trait Config<I = ()>: Config + Config
where
    Self::RuntimeEvent: From<Event<Self, I>>,
    I: 'static,
{
    type WeightInfo: WeightInfo;
    type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>;
    type Paymaster: Pay<Beneficiary = Self::AccountId, AssetKind = ()>;
    type Members: RankedMembers<AccountId = Self::AccountId>;
    type Salary: GetSalary<<Self::Members as RankedMembers>::Rank, Self::AccountId, <Self::Paymaster as Pay>::Balance>;
    type RegistrationPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type PayoutPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type Budget: Get<<Self::Paymaster as Pay>::Balance>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>

The runtime event type.

type Paymaster: Pay<Beneficiary = Self::AccountId, AssetKind = ()>

Means by which we can make payments to accounts. This also defines the currency and the balance which we use to denote that currency.

type Members: RankedMembers<AccountId = Self::AccountId>

The current membership of payees.

type Salary: GetSalary<<Self::Members as RankedMembers>::Rank, Self::AccountId, <Self::Paymaster as Pay>::Balance>

The maximum payout to be made for a single period to an active member of the given rank.

The benchmarks require that this be non-zero for some rank at most 255.

type RegistrationPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The number of blocks within a cycle which accounts have to register their intent to claim.

The number of blocks between sequential payout cycles is the sum of this and PayoutPeriod.

type PayoutPeriod: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The number of blocks within a cycle which accounts have to claim the payout.

The number of blocks between sequential payout cycles is the sum of this and RegistrationPeriod.

type Budget: Get<<Self::Paymaster as Pay>::Balance>

The total budget per cycle.

This may change over the course of a cycle without any problem.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.