yet_another_parachain_runtime

Crate pallet_session

Session Pallet

The Session pallet allows validators to manage their session keys, provides a function for changing the session length, and handles session rotation.

Config

Call

Pallet

Overview

Terminology

Session: A session is a period of time that has a constant set of validators. Validators can only join or exit the validator set at a session change. It is measured in block numbers. The block where a session is ended is determined by the ShouldEndSession trait. When the session is ending, a new validator set can be chosen by OnSessionEnding implementations.

Session key: A session key is actually several keys kept together that provide the various signing functions required by network authorities/validators in pursuit of their duties.

Validator ID: Every account has an associated validator ID. For some simple staking systems, this may just be the same as the account ID. For staking systems using a stash/controller model, the validator ID would be the stash account ID of the controller.

Session key configuration process: Session keys are set using set_keys for use not in the next session, but the session after next. They are stored in NextKeys, a mapping between the caller’s ValidatorId and the session keys provided. set_keys allows users to set their session key prior to being selected as validator. It is a public call since it uses ensure_signed, which checks that the origin is a signed account. As such, the account ID of the origin stored in NextKeys may not necessarily be associated with a block author or a validator. The session keys of accounts are removed once their account balance is zero.

Session length: This pallet does not assume anything about the length of each session. Rather, it relies on an implementation of ShouldEndSession to dictate a new session’s start. This pallet provides the PeriodicSessions struct for simple periodic sessions.

Session rotation configuration: Configure as either a ‘normal’ (rewardable session where rewards are applied) or ‘exceptional’ (slashable) session rotation.

Session rotation process: At the beginning of each block, the on_initialize function queries the provided implementation of ShouldEndSession. If the session is to end the newly activated validator IDs and session keys are taken from storage and passed to the SessionHandler. The validator set supplied by SessionManager::new_session and the corresponding session keys, which may have been registered via set_keys during the previous session, are written to storage where they will wait one session before being passed to the SessionHandler themselves.

Goals

The Session pallet is designed to make the following possible:

Set session keys of the validator set for upcoming sessions.

Control the length of sessions.

Configure and switch between either normal or exceptional session rotations.

Interface

Dispatchable Functions

set_keys - Set a validator’s session keys for upcoming sessions.

Public Functions

rotate_session - Change to the next session. Register the new authority set. Queue changes for next session rotation.

disable_index - Disable a validator by index.

disable - Disable a validator by Validator ID

Usage

Example from the FRAME

The Staking pallet uses the Session pallet to get the validator set.

use pallet_session as session;

fn validators<T: pallet_session::Config>() -> Vec<<T as pallet_session::Config>::ValidatorId> {
	pallet_session::Validators::<T>::get()
}

Related Pallets

Staking

Modules

disabling

historical

An opt-in utility for tracking historical sessions in FRAME-session.

migrations

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_session

Macros

log

Structs

FindAccountFromAuthorIndex

Wraps the author-scraping logic for consensus engines that can recover the canonical index of an author. This then transforms it into the registering account-ID of that session key index.

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

PeriodicSessions

Ends the session after a fixed period of blocks.

TestSessionHandler

SessionHandler for tests that use UintAuthorityId as Keys.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

Error for the session pallet.

Event

The Event enum of this pallet

HoldReason

A reason for the pallet placing a hold on funds.

Traits

Config

Configuration trait of this pallet.

SessionHandler

Handler for session life cycle events.

SessionManager

A trait for managing creation of new validator set.

ShouldEndSession

Decides whether the session should be ended.

WeightInfo

Weight functions needed for pallet_session.

Type Aliases

CurrentIndex

Current index of the session.

DisabledValidators

Indices of disabled validators.

KeyOwner

The owner of a key. The key is the KeyTypeId + the encoded key.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextKeys

The next session keys for a validator.

QueuedChanged

True if the underlying economic identities or weighting behind the validators has changed in the queued validator set.

QueuedKeys

The queued keys for the next session. When the next session begins, these keys will be used to determine the validator’s session keys.

Validators

The current set of validators.