yet_another_parachain_runtime::pallet_session::pallet

Trait Config

pub trait Config: Config + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{
    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type ValidatorId: Member + Parameter + MaybeSerializeDeserialize + MaxEncodedLen + TryFrom<Self::AccountId>;
    type ValidatorIdOf: Convert<Self::AccountId, Option<Self::ValidatorId>>;
    type ShouldEndSession: ShouldEndSession<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type NextSessionRotation: EstimateNextSessionRotation<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type SessionManager: SessionManager<Self::ValidatorId>;
    type SessionHandler: SessionHandler<Self::ValidatorId>;
    type Keys: OpaqueKeys + Member + Parameter + MaybeSerializeDeserialize;
    type DisablingStrategy: DisablingStrategy<Self>;
    type WeightInfo: WeightInfo;
    type Currency: Mutate<Self::AccountId> + Mutate<Self::AccountId>
       where <Self::Currency as Inspect<Self::AccountId>>::Reason: From<HoldReason>;
    type KeyDeposit: Get<<Self::Currency as Inspect<Self::AccountId>>::Balance>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

The overarching event type.

type ValidatorId: Member + Parameter + MaybeSerializeDeserialize + MaxEncodedLen + TryFrom<Self::AccountId>

A stable ID for a validator.

type ValidatorIdOf: Convert<Self::AccountId, Option<Self::ValidatorId>>

A conversion from account ID to validator ID.

It is also a means to check that an account id is eligible to set session keys, through being associated with a validator id. To disable this check, use sp_runtime::traits::ConvertInto.

Its cost must be at most one storage read.

type ShouldEndSession: ShouldEndSession<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Indicator for when to end the session.

type NextSessionRotation: EstimateNextSessionRotation<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

Something that can predict the next session rotation. This should typically come from the same logical unit that provides ShouldEndSession, yet, it gives a best effort estimate. It is helpful to implement EstimateNextNewSession.

type SessionManager: SessionManager<Self::ValidatorId>

Handler for managing new session.

type SessionHandler: SessionHandler<Self::ValidatorId>

Handler when a session has changed.

type Keys: OpaqueKeys + Member + Parameter + MaybeSerializeDeserialize

The keys.

type DisablingStrategy: DisablingStrategy<Self>

DisablingStragegy controls how validators are disabled

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type Currency: Mutate<Self::AccountId> + Mutate<Self::AccountId> where <Self::Currency as Inspect<Self::AccountId>>::Reason: From<HoldReason>

The currency type for placing holds when setting keys.

type KeyDeposit: Get<<Self::Currency as Inspect<Self::AccountId>>::Balance>

The amount to be held when setting keys.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.