yet_another_parachain_runtime::pallet_society

Trait Config

pub trait Config<I = ()>: Config + Config
where
    Self::RuntimeEvent: From<Event<Self, I>>,
    I: 'static,
{

    type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>;
    type PalletId: Get<PalletId>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type Randomness: Randomness<Self::Hash, <Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type GraceStrikes: Get<u32>;
    type PeriodSpend: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type VotingPeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type ClaimPeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type MaxLockDuration: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type FounderSetOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type ChallengePeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type MaxPayouts: Get<u32>;
    type MaxBids: Get<u32>;
    type WeightInfo: WeightInfo;
    type BlockNumberProvider: BlockNumberProvider;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>

The overarching event type.

type PalletId: Get<PalletId>

The societies’s pallet id

type Currency: ReservableCurrency<Self::AccountId>

The currency type used for bidding.

type Randomness: Randomness<Self::Hash, <Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

Something that provides randomness in the runtime.

type GraceStrikes: Get<u32>

The maximum number of strikes before a member gets funds slashed.

type PeriodSpend: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount of incentive paid within each period. Doesn’t include VoterTip.

type VotingPeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

The number of Config::BlockNumberProvider blocks on which new candidates should be voted on. Together with ClaimPeriod, this sums to the number of blocks between candidate intake periods.

type ClaimPeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

The number of Config::BlockNumberProvider blocks on which new candidates can claim their membership and be the named head.

type MaxLockDuration: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

The maximum duration of the payout lock.

type FounderSetOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin that is allowed to call found.

type ChallengePeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

The number of Config::BlockNumberProvider blocks between membership challenges.

type MaxPayouts: Get<u32>

The maximum number of payouts a member may have waiting unclaimed.

type MaxBids: Get<u32>

The maximum number of bids at once.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type BlockNumberProvider: BlockNumberProvider

Provider for the block number. Normally this is the frame_system pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.