yet_another_parachain_runtime::pallet_staking::benchmarking

Trait Benchmarking

pub trait Benchmarking {
    // Required methods
    fn benchmarks(extra: bool) -> Vec<BenchmarkMetadata>;
    fn run_benchmark(
        name: &[u8],
        selected_components: &[(BenchmarkParameter, u32)],
        whitelist: &[TrackedStorageKey],
        verify: bool,
        internal_repeats: u32,
    ) -> Result<Vec<BenchmarkResult>, BenchmarkError>;
}

The pallet benchmarking trait.

Required Methods

fn benchmarks(extra: bool) -> Vec<BenchmarkMetadata>

Get the benchmarks available for this pallet. Generally there is one benchmark per extrinsic, so these are sometimes just called “extrinsics”.

Parameters

extra: Also return benchmarks marked “extra” which would otherwise not be needed for weight calculation.

fn run_benchmark( name: &[u8], selected_components: &[(BenchmarkParameter, u32)], whitelist: &[TrackedStorageKey], verify: bool, internal_repeats: u32, ) -> Result<Vec<BenchmarkResult>, BenchmarkError>

Run the benchmarks for this pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.