yet_another_parachain_runtime::pallet_staking_async

Trait Config

pub trait Config: Config + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{

    type OldCurrency: InspectLockableCurrency<Self::AccountId, Moment = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, Balance = Self::CurrencyBalance>;
    type Currency: Mutate<Self::AccountId, Reason = Self::RuntimeHoldReason, Balance = Self::CurrencyBalance, Balance = Self::CurrencyBalance, Balance = Self::CurrencyBalance> + Mutate<Self::AccountId> + Balanced<Self::AccountId>;
    type RuntimeHoldReason: From<HoldReason>;
    type CurrencyBalance: AtLeast32BitUnsigned + FullCodec + DecodeWithMemTracking + HasCompact + Copy + MaybeSerializeDeserialize + Debug + Default + From<u64> + TypeInfo + Send + Sync + MaxEncodedLen
       where <Self::CurrencyBalance as HasCompact>::Type: DecodeWithMemTracking;
    type CurrencyToVote: CurrencyToVote<Self::CurrencyBalance>;
    type ElectionProvider: ElectionProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, DataProvider = Pallet<Self>>;
    type NominationsQuota: NominationsQuota<Self::CurrencyBalance>;
    type HistoryDepth: Get<u32>;
    type RewardRemainder: OnUnbalanced<Imbalance<<Self::Currency as Inspect<Self::AccountId>>::Balance, <Self::Currency as Balanced<Self::AccountId>>::OnDropCredit, <Self::Currency as Balanced<Self::AccountId>>::OnDropDebt>>;
    type Slash: OnUnbalanced<Imbalance<<Self::Currency as Inspect<Self::AccountId>>::Balance, <Self::Currency as Balanced<Self::AccountId>>::OnDropCredit, <Self::Currency as Balanced<Self::AccountId>>::OnDropDebt>>;
    type Reward: OnUnbalanced<Imbalance<<Self::Currency as Inspect<Self::AccountId>>::Balance, <Self::Currency as Balanced<Self::AccountId>>::OnDropDebt, <Self::Currency as Balanced<Self::AccountId>>::OnDropCredit>>;
    type SessionsPerEra: Get<u32>;
    type PlanningEraOffset: Get<u32>;
    type BondingDuration: Get<u32>;
    type SlashDeferDuration: Get<u32>;
    type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type EraPayout: EraPayout<Self::CurrencyBalance>;
    type MaxExposurePageSize: Get<u32>;
    type MaxValidatorSet: Get<u32>;
    type VoterList: SortedListProvider<Self::AccountId, Score = u64>;
    type TargetList: SortedListProvider<Self::AccountId, Score = Self::CurrencyBalance>;
    type MaxUnlockingChunks: Get<u32>;
    type MaxControllersInDeprecationBatch: Get<u32>;
    type EventListeners: OnStakingUpdate<Self::AccountId, Self::CurrencyBalance>;
    type MaxInvulnerables: Get<u32>;
    type MaxEraDuration: Get<u64>;
    type MaxPruningItems: Get<u32>;
    type RcClientInterface: RcClientInterface<AccountId = Self::AccountId>;
    type Filter: Contains<Self::AccountId>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type OldCurrency: InspectLockableCurrency<Self::AccountId, Moment = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, Balance = Self::CurrencyBalance>

The old trait for staking balance. Deprecated and only used for migrating old ledgers.

type Currency: Mutate<Self::AccountId, Reason = Self::RuntimeHoldReason, Balance = Self::CurrencyBalance, Balance = Self::CurrencyBalance, Balance = Self::CurrencyBalance> + Mutate<Self::AccountId> + Balanced<Self::AccountId>

The staking balance.

type RuntimeHoldReason: From<HoldReason>

Overarching hold reason.

type CurrencyBalance: AtLeast32BitUnsigned + FullCodec + DecodeWithMemTracking + HasCompact + Copy + MaybeSerializeDeserialize + Debug + Default + From<u64> + TypeInfo + Send + Sync + MaxEncodedLen where <Self::CurrencyBalance as HasCompact>::Type: DecodeWithMemTracking

Just the Currency::Balance type; we have this item to allow us to constrain it to From<u64>.

type CurrencyToVote: CurrencyToVote<Self::CurrencyBalance>

Convert a balance into a number used for election calculation. This must fit into a u64 but is allowed to be sensibly lossy. The u64 is used to communicate with the frame_election_provider_support crate which accepts u64 numbers and does operations in 128. Consequently, the backward convert is used convert the u128s from sp-elections back to a BalanceOf.

type ElectionProvider: ElectionProvider<AccountId = Self::AccountId, BlockNumber = <<Self::Block as HeaderProvider>::HeaderT as Header>::Number, DataProvider = Pallet<Self>>

Something that provides the election functionality.

type NominationsQuota: NominationsQuota<Self::CurrencyBalance>

Something that defines the maximum number of nominations per nominator.

type HistoryDepth: Get<u32>

Number of eras to keep in history.

Following information is kept for eras in [current_era - HistoryDepth, current_era]: ErasValidatorPrefs, ErasValidatorReward, ErasRewardPoints, ErasTotalStake, ClaimedRewards, ErasStakersPaged, ErasStakersOverview.

Must be more than the number of eras delayed by session. I.e. active era must always be in history. I.e. active_era > current_era - history_depth must be guaranteed.

If migrating an existing pallet from storage value to config value, this should be set to same value or greater as in storage.

type RewardRemainder: OnUnbalanced<Imbalance<<Self::Currency as Inspect<Self::AccountId>>::Balance, <Self::Currency as Balanced<Self::AccountId>>::OnDropCredit, <Self::Currency as Balanced<Self::AccountId>>::OnDropDebt>>

Tokens have been minted and are unused for validator-reward. See Era payout.

type Slash: OnUnbalanced<Imbalance<<Self::Currency as Inspect<Self::AccountId>>::Balance, <Self::Currency as Balanced<Self::AccountId>>::OnDropCredit, <Self::Currency as Balanced<Self::AccountId>>::OnDropDebt>>

Handler for the unbalanced reduction when slashing a staker.

type Reward: OnUnbalanced<Imbalance<<Self::Currency as Inspect<Self::AccountId>>::Balance, <Self::Currency as Balanced<Self::AccountId>>::OnDropDebt, <Self::Currency as Balanced<Self::AccountId>>::OnDropCredit>>

Handler for the unbalanced increment when rewarding a staker. NOTE: in most cases, the implementation of OnUnbalanced should modify the total issuance.

type SessionsPerEra: Get<u32>

Number of sessions per era, as per the preferences of the relay chain.

type PlanningEraOffset: Get<u32>

Number of sessions before the end of an era when the election for the next era will start.

This determines how many sessions before the last session of the era the staking election process should begin.

The value is bounded between 1 (election starts at the beginning of the last session) and SessionsPerEra (election starts at the beginning of the first session of the era).

Example:

If SessionsPerEra = 6 and PlanningEraOffset = 1, the election starts at the beginning of session 6 - 1 = 5.

If PlanningEraOffset = 6, the election starts at the beginning of session 6 - 6 = 0, meaning it starts at the very beginning of the era.

type BondingDuration: Get<u32>

Number of eras that staked funds must remain bonded for.

type SlashDeferDuration: Get<u32>

Number of eras that slashes are deferred by, after computation.

This should be less than the bonding duration. Set to 0 if slashes should be applied immediately, without opportunity for intervention.

type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which can manage less critical staking parameters that does not require root.

Supported actions: (1) cancel deferred slash, (2) set minimum commission.

type EraPayout: EraPayout<Self::CurrencyBalance>

The payout for validators and the system for the current era. See Era payout.

type MaxExposurePageSize: Get<u32>

The maximum size of each T::ExposurePage.

An ExposurePage is weakly bounded to a maximum of MaxExposurePageSize nominators.

For older non-paged exposure, a reward payout was restricted to the top MaxExposurePageSize nominators. This is to limit the i/o cost for the nominator payout.

Note: MaxExposurePageSize is used to bound ClaimedRewards and is unsafe to reduce without handling it in a migration.

type MaxValidatorSet: Get<u32>

The absolute maximum of winner validators this pallet should return.

As this pallet supports multi-block election, the set of winner validators per election is bounded by this type.

type VoterList: SortedListProvider<Self::AccountId, Score = u64>

Something that provides a best-effort sorted list of voters aka electing nominators, used for NPoS election.

The changes to nominators are reported to this. Moreover, each validatorโs self-vote is also reported as one independent vote.

To keep the load off the chain as much as possible, changes made to the staked amount via rewards and slashes are not reported and thus need to be manually fixed by the staker. In case of bags-list, this always means using rebag and putInFrontOf.

Invariant: what comes out of this list will always be a nominator.

type TargetList: SortedListProvider<Self::AccountId, Score = Self::CurrencyBalance>

WIP: This is a noop as of now, the actual business logic thatโs described below is going to be introduced in a follow-up PR.

Something that provides a best-effort sorted list of targets aka electable validators, used for NPoS election.

The changes to the approval stake of each validator are reported to this. This means any change to:

The stake of any validator or nominator.

The targets of any nominator

The role of any staker (e.g. validator -> chilled, nominator -> validator, etc)

Unlike VoterList, the values in this list are always kept up to date with reward and slash as well, and thus represent the accurate approval stake of all account being nominated by nominators.

Note that while at the time of nomination, all targets are checked to be real validators, they can chill at any point, and their approval stakes will still be recorded. This implies that what comes out of iterating this list MIGHT NOT BE AN ACTIVE VALIDATOR.

type MaxUnlockingChunks: Get<u32>

The maximum number of unlocking chunks a StakingLedger can have. Effectively determines how many unique eras a staker may be unbonding in.

Note: MaxUnlockingChunks is used as the upper bound for the BoundedVec item StakingLedger.unlocking. Setting this value lower than the existing value can lead to inconsistencies in the StakingLedger and will need to be handled properly in a runtime migration. The test reducing_max_unlocking_chunks_abrupt shows this effect.

type MaxControllersInDeprecationBatch: Get<u32>

The maximum amount of controller accounts that can be deprecated in one call.

type EventListeners: OnStakingUpdate<Self::AccountId, Self::CurrencyBalance>

Something that listens to staking updates and performs actions based on the data it receives.

WARNING: this only reports slashing and withdraw events for the time being.

type MaxInvulnerables: Get<u32>

Maximum number of invulnerable validators.

type MaxEraDuration: Get<u64>

Maximum allowed era duration in milliseconds.

This provides a defensive upper bound to cap the effective era duration, preventing excessively long eras from causing runaway inflation (e.g., due to bugs). If the actual era duration exceeds this value, it will be clamped to this maximum.

Example: For an ideal era duration of 24 hours (86,400,000 ms), this can be set to 604,800,000 ms (7 days).

type MaxPruningItems: Get<u32>

Maximum number of storage items that can be pruned in a single call.

This controls how many storage items can be deleted in each call to prune_era_step. This should be set to a conservative value (e.g., 100-500 items) to ensure pruning doesnโt consume too much block space. The actual weight is determined by benchmarks.

type RcClientInterface: RcClientInterface<AccountId = Self::AccountId>

Interface to talk to the RC-Client pallet, possibly sending election results to the relay chain.

type Filter: Contains<Self::AccountId>

Filter some accounts from participating in staking.

This is useful for example to blacklist an account that is participating in staking in another way (such as pools).

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.