yet_another_parachain_runtime

Crate pallet_staking_async_ah_client

The client for AssetHub, intended to be used in the relay chain.

The counter-part for this pallet is pallet-staking-async-rc-client on AssetHub.

This documentation is divided into the following sections:

Incoming messages: the messages that we receive from the relay chian.

Outgoing messages: the messaged that we sent to the relay chain.

Local interfaces: the interfaces that we expose to other pallets in the runtime.

Incoming Messages

All incoming messages are handled via Call. They are all gated to be dispatched only by Config::AssetHubOrigin. The only one is:

Call::validator_set: A new validator set for a planning session index.

Outgoing Messages

All outgoing messages are handled by a single trait pallet_staking_async_rc_client::SendToAssetHub. They match the incoming messages of the rc-client pallet.

Local Interfaces:

Living on the relay chain, this pallet must:

Implement pallet_session::SessionManager (and historical variant thereof) to give information to the session pallet.

Implements SessionInterface to receive information from the session pallet

Implement sp_staking::offence::OnOffenceHandler.

Implement reward related APIs (frame_support::traits::RewardsReporter).

Future Plans

Governance functions to force set validators.

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Macros

log

Structs

DefaultExposureOf

See pallet_staking::DefaultExposureOf. This type is the same, except it is duplicated here so that an rc-runtime can use it after pallet-staking is fully removed as a dependency.

GenesisConfig

Can be used to configure the genesis state of this pallet.

OffenceSendQueue

Wrapper struct for storing offences, and getting them back page by page.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

OperatingMode

Represents the operating mode of the pallet.

UnexpectedKind

Represents unexpected or invariant-breaking conditions encountered during execution.

Traits

Config

Configuration trait of this pallet.

SendToAssetHub

The interface to communicate to asset hub.

SessionInterface

Interface to talk to the local session pallet.

Type Aliases

BalanceOf

The balance type seen from this pallet’s PoV.

IncompleteValidatorSetReport

An incomplete validator set report.

Mode

Indicates the current operating mode of the pallet.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextSessionChangesValidators

A storage value that is set when a new_session gives a new validator set to the session pallet, and is cleared on the next call.

OffenceDetailsOf

Type alias for offence details

OutgoingSessionReport

A session report that is outgoing, and should be sent.

QueuedOffenceOf

A single buffered offence in OffenceSendQueue.

QueuedOffencePageOf

A page of buffered offences in OffenceSendQueue.

ValidatorPoints

All of the points of the validators.

ValidatorSet

The queued validator sets for a given planning session index.

ValidatorSetAppliedAt

The session index at which the latest elected validator set was applied.