yet_another_parachain_runtime::pallet_state_trie_migration

Trait Config

pub trait Config: Config + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{
    type ControlOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type SignedFilter: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>;
    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type Currency: Inspect<Self::AccountId, Reason = Self::RuntimeHoldReason> + Mutate<Self::AccountId> + Mutate<Self::AccountId> + Balanced<Self::AccountId>;
    type RuntimeHoldReason: From<HoldReason>;
    type MaxKeyLen: Get<u32>;
    type SignedDepositPerItem: Get<<Self::Currency as Inspect<Self::AccountId>>::Balance>;
    type SignedDepositBase: Get<<Self::Currency as Inspect<Self::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configurations of this pallet.

Required Associated Types

type ControlOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that can control the configurations of this pallet.

type SignedFilter: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>

Filter on which origin that trigger the manual migrations.

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

The overarching event type.

type Currency: Inspect<Self::AccountId, Reason = Self::RuntimeHoldReason> + Mutate<Self::AccountId> + Mutate<Self::AccountId> + Balanced<Self::AccountId>

The currency provider type.

type RuntimeHoldReason: From<HoldReason>

The overarching runtime hold reason.

type MaxKeyLen: Get<u32>

Maximal number of bytes that a key can have.

FRAME itself does not limit the key length. The concrete value must therefore depend on your storage usage. A frame_support::storage::StorageNMap for example can have an arbitrary number of keys which are then hashed and concatenated, resulting in arbitrarily long keys.

Use the state migration RPC to retrieve the length of the longest key in your storage: https://github.com/paritytech/substrate/issues/11642

The migration will halt with a Halted event if this value is too small. Since there is no real penalty from over-estimating, it is advised to use a large value. The default is 512 byte.

Some key lengths for reference:

frame_support::storage::StorageValue: 32 byte

frame_support::storage::StorageMap: 64 byte

frame_support::storage::StorageDoubleMap: 96 byte

For more info see https://www.shawntabrizi.com/blog/substrate/querying-substrate-storage-via-rpc/

type SignedDepositPerItem: Get<<Self::Currency as Inspect<Self::AccountId>>::Balance>

The amount of deposit collected per item in advance, for signed migrations.

This should reflect the average storage value size in the worse case.

type SignedDepositBase: Get<<Self::Currency as Inspect<Self::AccountId>>::Balance>

The base value of Config::SignedDepositPerItem.

Final deposit is items * SignedDepositPerItem + SignedDepositBase.

type WeightInfo: WeightInfo

The weight information of this pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.