yet_another_parachain_runtime::pallet_tips

Trait Config

pub trait Config<I = ()>:
    Config
    + Config<I>
    + Config
where
    Self::RuntimeEvent: From<Event<Self, I>>,
    I: 'static,
{
    type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>;
    type MaximumReasonLength: Get<u32>;
    type DataDepositPerByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type TipCountdown: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type TipFindersFee: Get<Percent>;
    type TipReportDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaxTipAmount: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type Tippers: SortedMembers<Self::AccountId> + ContainsLengthBound;
    type OnSlash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>

The overarching event type.

type MaximumReasonLength: Get<u32>

Maximum acceptable reason length.

Benchmarks depend on this value, be sure to update weights file when changing this value

type DataDepositPerByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit per byte within the tip report reason or bounty description.

type TipCountdown: Get<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The period for which a tip remains open after is has achieved threshold tippers.

type TipFindersFee: Get<Percent>

The percent of the final tip which goes to the original reporter of the tip.

type TipReportDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The non-zero amount held on deposit for placing a tip report.

type MaxTipAmount: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The maximum amount for a single tip.

type Tippers: SortedMembers<Self::AccountId> + ContainsLengthBound

Origin from which tippers must come.

ContainsLengthBound::max_len must be cost free (i.e. no storage read or heavy operation). Benchmarks depend on the value of ContainsLengthBound::max_len be sure to update weights file when altering this method.

type OnSlash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for the unbalanced decrease when slashing for a removed tip.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.