yet_another_parachain_runtime

Crate pallet_transaction_payment

Transaction Payment Pallet

This pallet provides the basic logic needed to pay the absolute minimum amount needed for a transaction to be included. This includes:

base fee: This is the minimum amount a user pays for a transaction. It is declared as a base weight in the runtime and converted to a fee using WeightToFee.

weight fee: A fee proportional to amount of weight a transaction consumes.

length fee: A fee proportional to the encoded length of the transaction.

tip: An optional tip. Tip increases the priority of the transaction, giving it a higher chance to be included by the transaction queue.

The base fee and adjusted weight and length fees constitute the inclusion fee, which is the minimum fee for a transaction to be included in a block.

The formula of final fee:

inclusion_fee = base_fee + length_fee + [targeted_fee_adjustment * weight_fee];
final_fee = inclusion_fee + tip;

targeted_fee_adjustment: This is a multiplier that can tune the final fee based on the congestion of the network.

Additionally, this pallet allows one to configure:

The mapping between one unit of weight to one unit of fee via Config::WeightToFee.

A means of updating the fee for the next block, via defining a multiplier, based on the final state of the chain at the end of the previous block. This can be configured via Config::FeeMultiplierUpdate

How the fees are paid via Config::OnChargeTransaction.

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_transaction_payment

Structs

ChargeTransactionPayment

Require the transactor pay for themselves and maybe include a tip to gain additional priority in the queue.

ConstFeeMultiplier

A struct to make the fee multiplier a constant

CurrencyAdapterDeprecated

Implements the transaction payment for a pallet implementing the Currency trait (eg. the pallet_balances) using an unbalance handler (implementing OnUnbalanced).

FeeDetails

The FeeDetails is composed of:

FungibleAdapter

Implements transaction payment for a pallet implementing the frame_support::traits::fungible trait (eg. pallet_balances) using an unbalance handler (implementing OnUnbalanced).

GenesisConfig

Can be used to configure the genesis state of this pallet.

InclusionFee

The base fee and adjusted weight and length fees constitute the inclusion fee.

NextFeeMultiplierOnEmpty

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

RuntimeDispatchInfo

Information related to a dispatchable’s class, weight, and fee that can be queried from the runtime.

TargetedFeeAdjustment

A struct to update the weight multiplier per block. It implements Convert<Multiplier, Multiplier>, meaning that it can convert the previous multiplier to the next one. This should be called on on_finalize of a block, prior to potentially cleaning the weight data from the system pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Event

The Event enum of this pallet

Pre

The info passed between the prepare and post-dispatch steps for the ChargeAssetTxPayment extension.

Releases

Storage releases of the pallet.

Val

The info passed between the validate and prepare steps for the ChargeAssetTxPayment extension.

Traits

Config

Configuration trait of this pallet.

DefaultConfig

Based on Config. Auto-generated by #[pallet::config(with_default)]. Can be used in tandem with #[register_default_config] and #[derive_impl] to derive test config traits based on existing pallet config traits in a safe and developer-friendly way.

MultiplierUpdate

Something that can convert the current multiplier to the next one.

OnChargeTransaction

Handle withdrawing, refunding and depositing of transaction fees.

TxCreditHold

Needs to be implemented for every OnChargeTransaction.

WeightInfo

Weight functions needed for pallet_transaction_payment.

Functions

__type_value_for_next_fee_multiplier_on_empty

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

Multiplier

Fee multiplier.

NextFeeMultiplier

Storage type is StorageValue with value type Multiplier.

StorageVersion

Storage type is StorageValue with value type Releases.