yet_another_parachain_runtime::pallet_treasury

Trait Config

pub trait Config<I = ()>: Config + Config
where
    Self::RuntimeEvent: From<Event<Self, I>>,
    I: 'static,
{

    type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type RejectOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>;
    type SpendPeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type Burn: Get<Permill>;
    type PalletId: Get<PalletId>;
    type BurnDestination: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type WeightInfo: WeightInfo;
    type SpendFunds: SpendFunds<Self, I>;
    type MaxApprovals: Get<u32>;
    type SpendOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = <Self::Currency as Currency<Self::AccountId>>::Balance>;
    type AssetKind: Parameter + MaxEncodedLen;
    type Beneficiary: Parameter + MaxEncodedLen;
    type BeneficiaryLookup: StaticLookup<Target = Self::Beneficiary>;
    type Paymaster: Pay<Beneficiary = Self::Beneficiary, AssetKind = Self::AssetKind>;
    type BalanceConverter: ConversionFromAssetBalance<<Self::Paymaster as Pay>::Balance, Self::AssetKind, <Self::Currency as Currency<Self::AccountId>>::Balance>;
    type PayoutPeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>;
    type BenchmarkHelper: ArgumentsFactory<Self::AssetKind, Self::Beneficiary>;
    type BlockNumberProvider: BlockNumberProvider;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

The staking balance.

type RejectOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which rejections must come.

type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>

The overarching event type.

type SpendPeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

Period between successive spends.

type Burn: Get<Permill>

Percentage of spare funds (if any) that are burnt per spend period.

type PalletId: Get<PalletId>

The treasury’s pallet id, used for deriving its sovereign account ID.

type BurnDestination: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for the unbalanced decrease when treasury funds are burned.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type SpendFunds: SpendFunds<Self, I>

Runtime hooks to external pallet using treasury to compute spend funds.

type MaxApprovals: Get<u32>

DEPRECATED: associated with spend_local call and will be removed in May 2025. Refer to https://github.com/paritytech/polkadot-sdk/pull/5961 for migration to spend.

The maximum number of approvals that can wait in the spending queue.

NOTE: This parameter is also used within the Bounties Pallet extension if enabled.

type SpendOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = <Self::Currency as Currency<Self::AccountId>>::Balance>

The origin required for approving spends from the treasury outside of the proposal process. The Success value is the maximum amount in a native asset that this origin is allowed to spend at a time.

type AssetKind: Parameter + MaxEncodedLen

Type parameter representing the asset kinds to be spent from the treasury.

type Beneficiary: Parameter + MaxEncodedLen

Type parameter used to identify the beneficiaries eligible to receive treasury spends.

type BeneficiaryLookup: StaticLookup<Target = Self::Beneficiary>

Converting trait to take a source type and convert to Self::Beneficiary.

type Paymaster: Pay<Beneficiary = Self::Beneficiary, AssetKind = Self::AssetKind>

Type for processing spends of Self::AssetKind in favor of Self::Beneficiary.

type BalanceConverter: ConversionFromAssetBalance<<Self::Paymaster as Pay>::Balance, Self::AssetKind, <Self::Currency as Currency<Self::AccountId>>::Balance>

Type for converting the balance of an Self::AssetKind to the balance of the native asset, solely for the purpose of asserting the result against the maximum allowed spend amount of the Self::SpendOrigin.

type PayoutPeriod: Get<<Self::BlockNumberProvider as BlockNumberProvider>::BlockNumber>

The period during which an approved treasury spend has to be claimed.

type BenchmarkHelper: ArgumentsFactory<Self::AssetKind, Self::Beneficiary>

Helper type for benchmarks.

type BlockNumberProvider: BlockNumberProvider

Provider for the block number. Normally this is the frame_system pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.