yet_another_parachain_runtime

Crate pallet_vesting

Vesting Pallet

Config

Call

Overview

A simple pallet providing a means of placing a linear curve on an account’s locked balance. This pallet ensures that there is a lock in place preventing the balance to drop below the unvested amount for any reason other than the ones specified in UnvestedFundsAllowedWithdrawReasons configuration value.

As the amount vested increases over time, the amount unvested reduces. However, locks remain in place and explicit action is needed on behalf of the user to ensure that the amount locked is equivalent to the amount remaining to be vested. This is done through a dispatchable function, either vest (in typical case where the sender is calling on their own behalf) or vest_other in case the sender is calling on another account’s behalf.

Interface

This pallet implements the VestingSchedule trait.

Dispatchable Functions

vest - Update the lock, reducing it in line with the amount “vested” so far.

vest_other - Update the lock of another account, reducing it in line with the amount “vested” so far.

Modules

migrations

Storage migrations for the vesting pallet.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_vesting

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

MaxVestingSchedulesGet

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

VestingInfo

Struct to encode the vesting schedule of an individual account.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

Error for the vesting pallet.

Event

The Event enum of this pallet

Releases

Traits

Config

Configuration trait of this pallet.

WeightInfo

Weight functions needed for pallet_vesting.

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

StorageVersion

Storage version of the pallet.

Vesting

Information regarding the vesting of a given account.