yet_another_parachain_runtime::pallet_xcm_bridge_hub::pallet

Trait Config

pub trait Config<I = ()>: Config<Self::BridgeMessagesPalletInstance> + Config
where
    Self::RuntimeEvent: From<Event<Self, I>>,
    I: 'static,
{

    type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>;
    type UniversalLocation: Get<Junctions>;
    type BridgedNetwork: Get<Location>;
    type BridgeMessagesPalletInstance: 'static;
    type MessageExportPrice: Get<Assets>;
    type DestinationVersion: GetVersion;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type OpenBridgeOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Location>;
    type BridgeOriginAccountIdConverter: ConvertLocation<<Self::ThisChain as Chain>::AccountId>;
    type BridgeDeposit: Get<<Self::ThisChain as Chain>::Balance>;
    type Currency: Mutate<<Self::ThisChain as Chain>::AccountId, Balance = <Self::ThisChain as Chain>::Balance, Reason = Self::RuntimeHoldReason>;
    type RuntimeHoldReason: From<HoldReason<I>>;
    type AllowWithoutBridgeDeposit: Contains<Location>;
    type LocalXcmChannelManager: LocalXcmChannelManager;
    type BlobDispatcher: DispatchBlob;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<Self::RuntimeEvent>

The overarching event type.

type UniversalLocation: Get<Junctions>

Runtime’s universal location.

type BridgedNetwork: Get<Location>

Bridged network as relative location of bridged GlobalConsensus.

type BridgeMessagesPalletInstance: 'static

Associated messages pallet instance that bridges us with the BridgedNetworkId consensus.

type MessageExportPrice: Get<Assets>

Price of single message export to the bridged consensus (Self::BridgedNetwork).

type DestinationVersion: GetVersion

Checks the XCM version for the destination.

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin that is allowed to call privileged operations on the pallet, e.g. open/close bridge for locations.

type OpenBridgeOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Location>

A set of XCM locations within local consensus system that are allowed to open bridges with remote destinations.

type BridgeOriginAccountIdConverter: ConvertLocation<<Self::ThisChain as Chain>::AccountId>

A converter between a location and a sovereign account.

type BridgeDeposit: Get<<Self::ThisChain as Chain>::Balance>

Amount of this chain native tokens that is reserved on the sibling parachain account when bridge open request is registered.

type Currency: Mutate<<Self::ThisChain as Chain>::AccountId, Balance = <Self::ThisChain as Chain>::Balance, Reason = Self::RuntimeHoldReason>

Currency used to pay for bridge registration.

type RuntimeHoldReason: From<HoldReason<I>>

The overarching runtime hold reason.

type AllowWithoutBridgeDeposit: Contains<Location>

Do not hold Self::BridgeDeposit for the location of Self::OpenBridgeOrigin. For example, it is possible to make an exception for a system parachain or relay.

type LocalXcmChannelManager: LocalXcmChannelManager

Local XCM channel manager.

type BlobDispatcher: DispatchBlob

XCM-level dispatcher for inbound bridge messages.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.