yet_another_parachain_runtime::pallet_xcm_bridge_hub

Type Alias BridgeOf

pub type BridgeOf<T, I> = Bridge<<T as Config<<T as Config<I>>::BridgeMessagesPalletInstance>>::ThisChain, <T as Config<<T as Config<I>>::BridgeMessagesPalletInstance>>::LaneId>;

An alias for the bridge metadata.

Aliased Type

struct BridgeOf<T, I> {
    pub bridge_origin_relative_location: Box<VersionedLocation>,
    pub bridge_origin_universal_location: Box<VersionedInteriorLocation>,
    pub bridge_destination_universal_location: Box<VersionedInteriorLocation>,
    pub state: BridgeState,
    pub bridge_owner_account: <<T as Config<<T as Config<I>>::BridgeMessagesPalletInstance>>::ThisChain as Chain>::AccountId,
    pub deposit: <<T as Config<<T as Config<I>>::BridgeMessagesPalletInstance>>::ThisChain as Chain>::Balance,
    pub lane_id: <T as Config<<T as Config<I>>::BridgeMessagesPalletInstance>>::LaneId,
}

Fields
bridge_origin_relative_location: Box<VersionedLocation>

Relative location of the bridge origin chain. This is expected to be convertible to the latest XCM, so the check and migration needs to be ensured.

bridge_origin_universal_location: Box<VersionedInteriorLocation>

See BridgeLocations::bridge_origin_universal_location. Stored for BridgeId sanity check.

bridge_destination_universal_location: Box<VersionedInteriorLocation>

See BridgeLocations::bridge_destination_universal_location. Stored for BridgeId sanity check.

state: BridgeState

Current bridge state.

bridge_owner_account: <<T as Config<<T as Config<I>>::BridgeMessagesPalletInstance>>::ThisChain as Chain>::AccountId

Account with the reserved funds. Derived from self.bridge_origin_relative_location.

deposit: <<T as Config<<T as Config<I>>::BridgeMessagesPalletInstance>>::ThisChain as Chain>::Balance

Reserved amount on the sovereign account of the sibling bridge origin.

lane_id: <T as Config<<T as Config<I>>::BridgeMessagesPalletInstance>>::LaneId

Mapping to the unique LaneId.