yet_another_parachain_runtime::polkadot_cli::service::rococo_runtime

Type Alias Migrations

pub type Migrations = (VersionedMigration<0, 2, VersionUncheckedMigrateToV2<Runtime, (), ()>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, MigrateToV7<Runtime>, VersionedMigration<0, 1, VersionUncheckedMigrateToV1<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, VersionedMigration<1, 2, UncheckedMigrateToV2<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, MigrateToV8<Runtime>, MigrateToV9<Runtime>, VersionedMigration<0, 1, VersionUncheckedMigrateToV1<Runtime, ()>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, MigrateV0ToV1<Runtime>, MigrateV0ToV1<Runtime, Instance2>, VersionedMigration<0, 1, MigrateToV1Impl<Runtime, BalanceTransferAllowDeath>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, UnlockAndUnreserveAllFunds<UnlockConfig>, UnlockAndUnreserveAllFunds<UnlockConfig>, UnreserveDeposits<UnlockConfig, ()>, Migration<Runtime, (), BalanceUnreserveWeight>, RemovePallet<DemocracyPalletName, <Runtime as Config>::DbWeight>, RemovePallet<CouncilPalletName, <Runtime as Config>::DbWeight>, RemovePallet<TechnicalCommitteePalletName, <Runtime as Config>::DbWeight>, RemovePallet<PhragmenElectionPalletName, <Runtime as Config>::DbWeight>, RemovePallet<TechnicalMembershipPalletName, <Runtime as Config>::DbWeight>, RemovePallet<TipsPalletName, <Runtime as Config>::DbWeight>, VersionedMigration<4, 5, UncheckedMigrateImpl<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, VersionedMigration<9, 10, VersionUncheckedMigrateToV10<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, VersionedMigration<0, 1, VersionUncheckedMigrateV0ToV1<Runtime, rococo_runtime::::migrations::Unreleased::{constant#0}>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, VersionedMigration<10, 11, UncheckedMigrateToV11<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, MigrateToCoretime<Runtime, WithUniqueTopic<ChildParachainRouter<Runtime, Pallet<Runtime>, ExponentialPrice<FeeAssetId, BaseDeliveryFee, TransactionByteFee, Pallet<Runtime>>>>, GetLegacyLeaseImpl, rococo_runtime::::migrations::Unreleased::{constant#1}>, VersionedMigration<11, 12, UncheckedMigrateToV12<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, VersionedMigration<0, 1, UncheckedMigrateToV1<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, VersionedMigration<0, 1, VersionUncheckedMigrateV0ToV1<Runtime, InitOffenceSeverity<Runtime>>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, MigrateToLatestXcmVersion<Runtime>, VersionedMigration<0, 1, VersionUncheckedMigrateToV1<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, VersionedMigration<0, 1, VersionUncheckedMigrateToV1<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>, VersionedMigration<2, 3, UncheckedMigrateToV3<Runtime>, Pallet<Runtime>, <Runtime as Config>::DbWeight>);

All migrations that will run on the next runtime upgrade.

This contains the combined migrations of the last 10 releases. It allows to skip runtime upgrades in case governance decides to do so. THE ORDER IS IMPORTANT.