yet_another_parachain_runtime::polkadot_node_subsystem::overseer::gen::futures::prelude::stream

Struct FlatMapUnordered

pub struct FlatMapUnordered<St, U, F>
where
    St: Stream,
    U: Stream + Unpin,
    F: FnMut(<St as Stream>::Item) -> U,
{ /* private fields */ }

Stream for the flat_map_unordered method.

Implementations

impl<St, U, F> FlatMapUnordered<St, U, F>
where St: Stream, U: Stream + Unpin, F: FnMut(<St as Stream>::Item) -> U,

pub fn get_ref(&self) -> &St

Acquires a reference to the underlying sink or stream that this combinator is pulling from.

pub fn get_mut(&mut self) -> &mut St

Acquires a mutable reference to the underlying sink or stream that this combinator is pulling from.

Note that care must be taken to avoid tampering with the state of the sink or stream which may otherwise confuse this combinator.

pub fn get_pin_mut(self: Pin<&mut FlatMapUnordered<St, U, F>>) -> Pin<&mut St>

Acquires a pinned mutable reference to the underlying sink or stream that this combinator is pulling from.

Note that care must be taken to avoid tampering with the state of the sink or stream which may otherwise confuse this combinator.

pub fn into_inner(self) -> St

Consumes this combinator, returning the underlying sink or stream.

Note that this may discard intermediate state of this combinator, so care should be taken to avoid losing resources when this is called.