yet_another_parachain_runtime::polkadot_node_subsystem::overseer

Struct OverseerSubsystemContext

pub struct OverseerSubsystemContext<M>
where
    M: AssociateOutgoing + Send + 'static,
{ /* private fields */ }

A context type that is given to the Subsystem upon spawning. It can be used by Subsystem to communicate with other Subsystems or to spawn it’s SubsystemJobs.