yet_another_parachain_runtime::polkadot_node_subsystem_util

Module inclusion_emulator

An emulator for node-side code to predict the results of on-chain parachain inclusion and predict future constraints.

Structs

ConstraintModifications

Modifications to constraints as a result of prospective candidates.

Constraints

Constraints on the actions that can be taken by a new parachain block. These limitations are implicitly associated with some particular parachain, which should be apparent from usage.

Fragment

A parachain fragment, representing another prospective parachain block.

InboundHrmpLimitations

Constraints on inbound HRMP channels.

OutboundHrmpChannelLimitations

Constraints on outbound HRMP channels.

OutboundHrmpChannelModification

An update to outbound HRMP channels.

ProspectiveCandidate

The prospective candidate.

RelayChainBlockInfo

Information about a relay-chain block.

Enums

FragmentValidityError

Kinds of errors with the validity of a fragment.

HrmpWatermarkUpdate

An update to the HRMP Watermark.

ModificationError

Kinds of errors that can occur when modifying constraints.

Traits

HypotheticalOrConcreteCandidate

Trait for a hypothetical or concrete candidate, as needed when assessing the validity of a potential candidate.

Functions

validate_commitments

Validates if the candidate commitments are obeying the constraints.