yet_another_parachain_runtime::polkadot_node_subsystem_util

Module runtime

Convenient and efficient runtime info access. Convenient interface to runtime information.

Structs

ClaimQueueSnapshot

A snapshot of the runtime claim queue at an arbitrary relay chain block.

Config

Configuration for construction a RuntimeInfo.

ExtendedSessionInfo

SessionInfo with additional useful data for validator nodes.

RuntimeInfo

Caching of session info.

ValidatorInfo

Information about ourselves, in case we are an Authority.

Enums

Error

FatalError

JfyiError

Functions

check_signature

Convenience function for quickly checking the signature on signed data.

fetch_claim_queue

Fetch the claim queue and wrap it into a helpful ClaimQueueSnapshot

fetch_scheduling_lookahead

Returns the lookahead from the scheduler params if the runtime supports it, or default value if scheduling lookahead API is not supported by runtime.

fetch_validation_code_bomb_limit

Fetch the validation code bomb limit from the runtime.

get_availability_cores

Request availability cores from the runtime.

get_candidate_events

Get CandidateEvents for the given relay_parent.

get_group_rotation_info

Get group rotation info based on the given relay_parent.

get_occupied_cores

Variant of request_availability_cores that only returns occupied ones.

get_on_chain_votes

Get on chain votes.

get_unapplied_slashes

Fetch a list of PendingSlashes from the runtime. Will fallback to unapplied_slashes if the runtime does not support unapplied_slashes_v2.

get_validation_code_by_hash

Fetch ValidationCode by hash from the runtime.

key_ownership_proof

Generate validator key ownership proof.

recv_runtime

Receive a response from a runtime request and convert errors.

submit_report_dispute_lost

Submit a past-session dispute slashing report.