yet_another_parachain_runtime::polkadot_node_subsystem_util

Trait Subsystem

pub trait Subsystem<Ctx, E>
where
    Ctx: SubsystemContext,
    E: Error + Send + Sync + 'static + From<OrchestraError>,
{
    // Required method
    fn start(self, ctx: Ctx) -> SpawnedSubsystem<E>;
}

A trait that describes the Subsystems that can run on the Orchestra.

It is generic over the message type circulating in the system. The idea that we want some type containing persistent state that can spawn actually running subsystems when asked.

Required Methods

fn start(self, ctx: Ctx) -> SpawnedSubsystem<E>

Start this Subsystem and return SpawnedSubsystem.