yet_another_parachain_runtime::polkadot_runtime_common::identity_migrator::pallet

Trait Config

pub trait Config:
    Config
    + Config
    + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{
    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type Reaper: EnsureOrigin<Self::RuntimeOrigin>;
    type ReapIdentityHandler: OnReapIdentity<Self::AccountId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

Overarching event type.

type Reaper: EnsureOrigin<Self::RuntimeOrigin>

The origin that can reap identities. Expected to be EnsureSigned<AccountId> on the source chain such that anyone can all this function.

type ReapIdentityHandler: OnReapIdentity<Self::AccountId>

A handler for what to do when an identity is reaped.

type WeightInfo: WeightInfo

Weight information for the extrinsics in the pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.