yet_another_parachain_runtime::polkadot_runtime_parachains::disputes

Trait SlashingHandler

pub trait SlashingHandler<BlockNumber> {
    // Required methods
    fn punish_for_invalid(
        session: u32,
        candidate_hash: CandidateHash,
        losers: impl IntoIterator<Item = ValidatorIndex>,
        backers: impl IntoIterator<Item = ValidatorIndex>,
    );
    fn punish_against_valid(
        session: u32,
        candidate_hash: CandidateHash,
        losers: impl IntoIterator<Item = ValidatorIndex>,
        backers: impl IntoIterator<Item = ValidatorIndex>,
    );
    fn initializer_initialize(now: BlockNumber) -> Weight;
    fn initializer_finalize();
    fn initializer_on_new_session(session_index: u32);
}

Punishment hooks for disputes.

Required Methods

fn punish_for_invalid( session: u32, candidate_hash: CandidateHash, losers: impl IntoIterator<Item = ValidatorIndex>, backers: impl IntoIterator<Item = ValidatorIndex>, )

Punish a series of validators who were for an invalid parablock. This is expected to trigger a large punishment for backers and a medium punishment for other approvers.

fn punish_against_valid( session: u32, candidate_hash: CandidateHash, losers: impl IntoIterator<Item = ValidatorIndex>, backers: impl IntoIterator<Item = ValidatorIndex>, )

Punish a series of validators who were against a valid parablock. This is expected to be a minor punishment.

fn initializer_initialize(now: BlockNumber) -> Weight

Called by the initializer to initialize the slashing pallet.

fn initializer_finalize()

Called by the initializer to finalize the slashing pallet.

fn initializer_on_new_session(session_index: u32)

Called by the initializer to note that a new session has started.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.