yet_another_parachain_runtime::polkadot_runtime_parachains::paras

Trait Config

pub trait Config:
    Config
    + Config
    + Config
    + CreateBare<Call<Self>>
    + Config
where
    Self::RuntimeEvent: From<Event<Self>>,
{
    type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>;
    type UnsignedPriority: Get<u64>;
    type NextSessionRotation: EstimateNextSessionRotation<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type QueueFootprinter: QueueFootprinter<Origin = UmpQueueId>;
    type OnNewHead: OnNewHead;
    type WeightInfo: WeightInfo;
    type AssignCoretime: AssignCoretime;
    type Fungible: Mutate<Self::AccountId>
       where <Self::Fungible as Inspect<Self::AccountId>>::Balance: From<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>;
    type CooldownRemovalMultiplier: Get<<Self::Fungible as Inspect<Self::AccountId>>::Balance>;
    type AuthorizeCurrentCodeOrigin: EnsureOriginWithArg<Self::RuntimeOrigin, Id>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<Self::RuntimeEvent>

type UnsignedPriority: Get<u64>

type NextSessionRotation: EstimateNextSessionRotation<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

type QueueFootprinter: QueueFootprinter<Origin = UmpQueueId>

Retrieve how many UMP messages are enqueued for this para-chain.

This is used to judge whether or not a para-chain can offboard. Per default this should be set to the ParaInclusion pallet.

type OnNewHead: OnNewHead

Runtime hook for when a parachain head is updated.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type AssignCoretime: AssignCoretime

Runtime hook for assigning coretime for a given parachain.

This is only used at genesis or by root.

TODO: Remove once coretime is the standard across all chains.

type Fungible: Mutate<Self::AccountId> where <Self::Fungible as Inspect<Self::AccountId>>::Balance: From<<<Self::Block as HeaderProvider>::HeaderT as Header>::Number>

The fungible instance used by the runtime.

type CooldownRemovalMultiplier: Get<<Self::Fungible as Inspect<Self::AccountId>>::Balance>

Multiplier to determine the cost of removing upgrade cooldown.

After a parachain upgrades their runtime, an upgrade cooldown is applied (configuration::HostConfiguration::validation_upgrade_cooldown). This cooldown exists to prevent spamming the relay chain with runtime upgrades. But as life is going on, mistakes can happen and a consequent may be required. The cooldown period can be removed by using Pallet::remove_upgrade_cooldown. This dispatchable will use this multiplier to determine the cost for removing the upgrade cooldown. Time left for the cooldown multiplied with this multiplier determines the cost.

type AuthorizeCurrentCodeOrigin: EnsureOriginWithArg<Self::RuntimeOrigin, Id>

The origin that can authorize Pallet::authorize_force_set_current_code_hash.

In the end this allows Pallet::apply_authorized_force_set_current_code to force set the current code without paying any fee. So, the origin should be chosen with care.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.