yet_another_parachain_runtime::polkadot_sdk_frame

Module pallet_macros

Macros used within the main pallet macro.

Note: All of these macros are “stubs” and not really usable outside #[pallet] mod pallet { .. }. They are mainly provided for documentation and IDE support.

To view a list of all the macros and their documentation, follow the links in the ‘Re-exports’ section below:

Attribute Macros

authorize

Allows to authorize some general transactions with specific dispatchable functions (dispatchable functions a.k.a. calls).

call

call_index

Each dispatchable may also be annotated with the #[pallet::call_index($idx)] attribute, which explicitly defines the codec index for the dispatchable function in the Call enum.

compact

composite_enum

config

constant

derive_impl

This attribute can be used to derive a full implementation of a trait based on a local partial impl and an external impl containing defaults that can be overridden in the local impl.

disable_frame_system_supertrait_check

disable_try_decode_storage

error

event

extra_constants

feeless_if

generate_deposit

genesis_build

genesis_config

getter

hooks

import_section

include_metadata

inherent

no_default

no_default_bounds

origin

pallet

pallet_section

storage

storage_prefix

storage_version

task_condition

Allows defining conditions for a task to run.

task_index

Allows defining an index for a task.

task_list

Allows defining an iterator over available work items for a task.

task_weight

Allows defining the weight of a task.

tasks_experimental

Allows you to define some service work that can be recognized by a script or an off-chain worker.

type_value

unbounded

validate_unsigned

view_functions

weight

weight_of_authorize

Allows to define the weight of the authorize function.

whitelist_storage