yet_another_parachain_runtime::polkadot_sdk_frame::runtime

Module prelude

The main prelude of FRAME for building runtimes.

A runtime typically starts with:

use polkadot_sdk_frame::runtime::prelude::*;

This automatically brings in polkadot_sdk_frame::prelude::*.

Macros

assert_eq_error_rate

Copied from sp-runtime and documented there.

construct_runtime

Construct a runtime, with the given name and the given pallets.

create_runtime_strDeprecated

Deprecated Cow::Borrowed() wrapper.

defensive

Generic function to mark an execution path as ONLY defensive.

defensive_assert

Trigger a defensive failure if a condition is not met.

ensure

Evaluate $x:expr and if not true return Err($y:expr).

impl_runtime_apis

impl_tx_ext_default

Helper macro to be used in a impl TransactionExtension block to add default implementations of weight, validate, prepare or any combinations of the them.

ord_parameter_types

Macro for easily creating a new implementation of both the Get and Contains traits. Use exactly as with parameter_types, only the type must be Ord.

parameter_types

Create new implementations of the Get trait.

Structs

Account

Details of an account for which a private key is contained in the keystore.

AsEnsureOriginWithArg

Blake2_128

Hash storage keys with blake2 128

Blake2_256

Hash storage keys with blake2 256

Blake2_128Concat

Hash storage keys with concat(blake2_128(key), key)

BlakeTwo256

Blake2-256 Hash implementation.

BoundedBTreeMap

A bounded map based on a B-Tree.

BoundedBTreeSet

A bounded set based on a B-Tree.

BoundedSlice

A bounded slice.

BoundedVec

A bounded vector.

CallMetadata

The function and pallet name of the Call.

CheckInherentsResult

The result of checking inherents.

ConstBool

Const getter for a basic type.

ConstI8

Const getter for a basic type.

ConstI16

Const getter for a basic type.

ConstI32

Const getter for a basic type.

ConstI64

Const getter for a basic type.

ConstI128

Const getter for a basic type.

ConstU8

Const getter for a basic type.

ConstU16

Const getter for a basic type.

ConstU32

Const getter for a basic type.

ConstU64

Const getter for a basic type.

ConstU128

Const getter for a basic type.

ConstUint

Const getter for unsigned integers.

CountedStorageMap

A wrapper around a StorageMap and a StorageValue (with the value being u32) to keep track of how many items are in a map, without needing to iterate all the values.

CountedStorageNMap

A wrapper around a StorageNMap and a StorageValue (with the value being u32) to keep track of how many items are in a map, without needing to iterate all the values.

DispatchErrorWithPostInfo

Result of a Dispatchable which contains the DispatchResult and additional information about the Dispatchable that is only known post dispatch.

EitherOf

“OR gate” implementation of EnsureOrigin, Success type for both L and R must be equal.

EitherOfDiverse

“OR gate” implementation of EnsureOrigin allowing for different Success types for L and R, with them combined using an Either type.

EnsureNever

Always fail.

EnsureNone

Ensure the origin is None. i.e. unsigned transaction.

EnsureRoot

Ensure the origin is Root.

EnsureRootWithSuccess

Ensure the origin is Root and return the provided Success value.

EnsureSigned

Ensure the origin is any Signed origin.

EnsureSignedBy

Ensure the origin is Signed origin from the given AccountId.

Executive

Main entry point for certain runtime actions as e.g. execute_block.

FixedFee

Implementor of WeightToFee such that it maps any unit of weight to a fixed fee.

FixedI64

A fixed point number representation in the range. Fixed Point 64 bits signed, range = [-9223372036.854775808, 9223372036.854775807]

FixedI128

A fixed point number representation in the range. Fixed Point 128 bits signed, range = [-170141183460469231731.687303715884105728, 170141183460469231731.687303715884105727]

FixedU64

A fixed point number representation in the range. Fixed Point 64 bits unsigned, range = [0.000000000, 18446744073.709551615]

FixedU128

A fixed point number representation in the range. Fixed Point 128 bits unsigned, range = [0.000000000000000000, 340282366920938463463.374607431768211455]

ForAll

Marker struct used to flag using all supported keys to sign a payload.

ForAny

Marker struct used to flag using any of the supported keys to sign a payload.

GetDefault

Implement Get by returning Default for any type that implements Default.

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H512

Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

Identity

Store the key directly.

IdentityLookup

A lookup implementation returning the input value.

InherentData

Inherent data to include in a block.

InsideBoth

A Contains implementation which contains all members of These which are also members of Those.

Keccak256

Keccak-256 Hash implementation.

MakeFatalError

Auxiliary to make any given error resolve to is_fatal_error() == true for IsFatalError.

MapSuccess

A derivative EnsureOrigin implementation. It mutates the Success result of an Original implementation with a given Mutator.

NMapKey

A type used exclusively by storage maps as their key type.

NativeVersion

The version of the native runtime.

NoOpPoll

NoOp polling is required if pallet-referenda functionality not needed.

OpaqueMetadata

Stores the encoded RuntimeMetadata for the native side as opaque type.

OptionQuery

Implements QueryKindTrait with Query type being Option<_>.

PalletId

A pallet identifier. These are per pallet and should be stored in a registry somewhere.

PerU16

A fixed point representation of a number in the range [0, 1].

Perbill

A fixed point representation of a number in the range [0, 1].

Percent

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

Perquintill

A fixed point representation of a number in the range [0, 1].

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

PostDispatchInfo

Weight information that is only available post dispatch. NOTE: This can only be used to reduce the weight or fee, not increase it.

Rational128

A wrapper for any rational number with a 128 bit numerator and denominator.

RationalInfinite

A wrapper for any rational number with infinitely large numerator and denominator.

ReduceBy

Mutator which reduces a scalar by a particular amount.

ReplaceWithDefault

Morpher to disregard the source value and replace with the default of V.

ResultQuery

Implements QueryKindTrait with Query type being Result<Value, PalletError>.

RuntimeVersion

Runtime version. This should not be thought of as classic Semver (major/minor/tiny). This triplet have different semantics and mis-interpretation could cause problems. In particular: bug fixes should result in an increment of spec_version and possibly authoring_version, absolutely not impl_version since they change the semantics of the runtime.

Signer

Provides an implementation for signing transaction payloads.

StorageDoubleMap

A type representing a double map in storage. This structure associates a pair of keys with a value of a specified type stored on-chain.

StorageMap

A type representing a map in storage. A storage map is a mapping of keys to values of a given type stored on-chain.

StorageNMap

A type representing an NMap in storage. This structure associates an arbitrary number of keys with a value of a specified type stored on-chain.

StorageValue

A type representing a value in storage. A storage value is a single value of a given type stored on-chain.

StorageVersion

The storage version of a pallet.

SubmitTransaction

Provides the ability to directly submit signed and unsigned transaction onchain.

TrailingZeroInput

Input that adds infinite number of zero after wrapped input.

Twox64Concat

Hash storage keys with concat(twox64(key), key)

Twox128

Hash storage keys with twox 128

Twox256

Hash storage keys with twox 256

U256

Little-endian large integer type 256-bit unsigned integer.

U512

Little-endian large integer type 512-bits unsigned integer.

ValidTransaction

Information concerning a valid transaction.

ValueQuery

Implements QueryKindTrait with Query type being Value.

VariantCountOf

Adapter for Get<u32> to access VARIANT_COUNT from trait pub trait VariantCount {.

WeakBoundedVec

A weakly bounded vector.

Weight

WithdrawReasons

Reasons for moving funds out of an account.

Enums

ArithmeticError

Arithmetic errors.

BalanceStatus

Status of funds.

DepositConsequence

One of a number of consequences of withdrawing a fungible from an account.

DispatchClass

A generalized group of dispatch types.

DispatchError

Reason why a dispatch call failed.

Everything

A Contains implementation that contains every value.

ExecutiveError

ExistenceRequirement

Simple boolean for whether an account needs to be kept in existence.

ExtrinsicInclusionMode

Confines the kind of extrinsics that can be included in a block.

Fortitude

The privilege with which a withdraw operation is conducted.

InvalidTransaction

An invalid transaction validity.

Pays

Explicit enum to denote if a transaction pays fee or not.

Precision

The precision required of an operation generally involving some aspect of quantitative fund withdrawal or transfer.

Preservation

The mode by which we describe whether an operation should keep an account alive.

Provenance

The origin of funds to be used for a deposit operation.

Rounding

The rounding method to use for unsigned quantities.

SignedRounding

The rounding method to use.

Sr25519Keyring

Set of test accounts.

TokenError

Description of what went wrong when trying to complete an operation on a token.

TransactionSource

The source of the transaction.

TransactionValidityError

Errors that can occur while checking the validity of a transaction.

UnknownTransaction

An unknown transaction validity.

WithdrawConsequence

One of a number of consequences of withdrawing a fungible from an account.

Constants

DEV_RUNTIME_PRESET

The default development preset used to communicate with the runtime via GenesisBuilder interface.

LOCAL_TESTNET_RUNTIME_PRESET

The default local_testnet preset used to communicate with the runtime via GenesisBuilder interface.

MAX_MODULE_ERROR_ENCODED_SIZE

The number of bytes of the module-specific error field defined in ModuleError. In FRAME, this is the maximum encoded size of a pallet error type.

Traits

AccountIdConversion

This type can be converted into and possibly from an AccountId (which itself is generic).

AppCrypto

A type binding runtime-level Public/Signature pair with crypto wrapped by RuntimeAppPublic.

AsTransactionAuthorizedOrigin

Interface to differentiate between Runtime Origins authorized to include a transaction into the block and dispatch it, and those who aren’t.

AssetId

Simple amalgamation trait to collect together properties for an AssetId under one roof.

AtLeast8Bit

A meta trait for arithmetic.

AtLeast8BitUnsigned

A meta trait for arithmetic. Same as AtLeast8Bit, but also bounded to be unsigned.

AtLeast16Bit

A meta trait for arithmetic.

AtLeast16BitUnsigned

A meta trait for arithmetic. Same as AtLeast16Bit, but also bounded to be unsigned.

AtLeast32Bit

A meta trait for arithmetic.

AtLeast32BitUnsigned

A meta trait for arithmetic. Same as AtLeast32Bit, but also bounded to be unsigned.

Authorize

A trait to allow calls to authorize themselves from the origin None.

BaseArithmetic

A meta trait for arithmetic type operations, regardless of any limitation on size.

BlockNumberProvider

Get current block number

BoundToRuntimeAppPublic

Something that is bound to a fixed RuntimeAppPublic.

Bounded

Numbers which have upper and lower bounds

BuildGenesisConfig

A trait to define the build function of a genesis config for both runtime and pallets.

ChangeMembers

Trait for type that can handle incremental changes to a set of account IDs.

CheckedAdd

Performs addition that returns None instead of wrapping around on overflow.

CheckedConversion

Convenience type to work around the highly unergonomic syntax needed to invoke the functions of overloaded generic traits, in this case TryFrom and TryInto.

CheckedDiv

Performs division that returns None instead of panicking on division by zero and instead of wrapping around on underflow and overflow.

CheckedMul

Performs multiplication that returns None instead of wrapping around on underflow or overflow.

CheckedNeg

Performs negation that returns None if the result can’t be represented.

CheckedRem

Performs an integral remainder that returns None instead of panicking on division by zero and instead of wrapping around on underflow and overflow.

CheckedShl

Performs a left shift that returns None on shifts larger than or equal to the type width.

CheckedShr

Performs a right shift that returns None on shifts larger than or equal to the type width.

CheckedSub

Performs subtraction that returns None instead of wrapping around on underflow.

Contains

A trait for querying whether a type can be said to “contain” a value.

Convert

Infallible conversion trait. Generic over both source and destination types.

ConvertBack

Reversing infallible conversion trait. Generic over both source and destination types.

CreateAuthorizedTransaction

Interface for creating a transaction for a call that will be authorized.

CreateBare

Interface for creating a bare extrinsic.

CreateInherent

Interface for creating a bare extrinsic.

CreateSignedTransaction

Interface for creating an old-school signed transaction.

CreateTransaction

Interface for creating a transaction.

CreateTransactionBase

Common interface for the CreateTransaction trait family to unify the Call type.

Debug

? formatting.

Decode

Trait that allows zero-copy read of value-references from slices in LE format.

DecodeWithMemTracking

Marker trait used for identifying types that call the [Input::on_before_alloc_mem] hook while decoding.

Defensive

A trait to handle errors and options when you are really sure that a condition must hold, but not brave enough to expect on it, or a default fallback value makes more sense.

DefensiveSaturating

A variant of Defensive with the same rationale, for the arithmetic operations where in case an infallible operation fails, it saturates.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

DispatchTransaction

Single-function utility trait with a blanket impl over TransactionExtension in order to provide transaction dispatching functionality. We avoid implementing this directly on the trait since we never want it to be overriden by the trait implementation.

Dispatchable

A lazy call (module function and argument values) that can be executed via its dispatch method.

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

Ensure

EnsureAdd

Performs addition that returns ArithmeticError instead of wrapping around on overflow.

EnsureAddAssign

Performs self addition that returns ArithmeticError instead of wrapping around on overflow.

EnsureDiv

Performs division that returns ArithmeticError instead of wrapping around on overflow.

EnsureDivAssign

Performs self division that returns ArithmeticError instead of wrapping around on overflow.

EnsureFixedPointNumber

Extends FixedPointNumber with the Ensure family functions.

EnsureFrom

Similar to TryFrom but returning an ArithmeticError error.

EnsureInto

Similar to TryInto but returning an ArithmeticError error.

EnsureMul

Performs multiplication that returns ArithmeticError instead of wrapping around on overflow.

EnsureMulAssign

Performs self multiplication that returns ArithmeticError instead of wrapping around on overflow.

EnsureOp

Meta trait that supports all immutable arithmetic Ensure* operations

EnsureOpAssign

Meta trait that supports all assigned arithmetic Ensure* operations

EnsureOrigin

Some sort of check on the origin is performed by this object.

EnsureSub

Performs subtraction that returns ArithmeticError instead of wrapping around on underflow.

EnsureSubAssign

Performs self subtraction that returns ArithmeticError instead of wrapping around on underflow.

EstimateNextSessionRotation

Something that can estimate at which block the next session rotation will happen (i.e. a new session starts).

FixedPointNumber

Something that implements a decimal fixed point number.

FixedPointOperand

Integer types that can be used to interact with FixedPointNumber implementations.

Get

A trait for querying a single value from a type.

GetCallMetadata

Gets the metadata for the Call - function name and pallet name.

GetDispatchInfo

A Dispatchable function (aka transaction) that can carry some static information along with it, using the #[pallet::weight] attribute.

GetStorageVersion

Provides information about a pallet’s storage versions.

Hash

Abstraction around hashing

Hooks

The pallet hooks trait. This is merely an umbrella trait for:

IdentifyAccount

Some type that is able to be collapsed into an account ID. It is not possible to recreate the original value from the account ID.

InitializeMembers

Trait for type that can handle the initialization of account IDs at genesis.

InstanceFilter

Simple trait for providing a filter over a reference to some type, given an instance of itself.

IntegerSquareRoot

A trait implementing integer square root.

IsSubType

Something that can be checked to be a of sub type T.

IsType

Trait to be used when types are exactly same.

MaxEncodedLen

Items implementing MaxEncodedLen have a statically known maximum encoded size.

MaybeSerializeDeserialize

A type that implements Serialize, DeserializeOwned and Debug when in std environment or serde feature is activated.

Member

A type that can be used in runtime structures.

MultiplyArg

MultiplyRational

Normalizable

A collection-like object that is made of values of type T and can normalize its individual values around a centric point.

OnInitializeWithWeightRegistration

Extension trait for OnInitialize.

OnRuntimeUpgrade

See Hooks::on_runtime_upgrade.

OnUnbalanced

Handler for when some currency “account” decreased in balance for some reason.

One

Defines a multiplicative identity element for Self.

OneSessionHandler

A session handler for specific key type.

OriginTrait

Methods available on frame_system::Config::RuntimeOrigin.

PalletInfoAccess

Provides information about the pallet itself and its setup in the runtime.

Parameter

A type that can be used as a parameter in a dispatchable function.

Pay

Can be implemented by PayFromAccount using a fungible impl, but can also be implemented with XCM/Asset and made generic over assets.

PerThing

Something that implements a fixed point ration with an arbitrary granularity X, as parts per X.

ProvideInherent

A pallet that provides or verifies an inherent extrinsic will implement this trait.

RankedMembers

Ranked membership data structure.

RankedMembersSwapHandler

Handler that can deal with the swap of two members.

RationalArg

ReciprocalArg

SaturatedConversion

Convenience type to work around the highly unergonomic syntax needed to invoke the functions of overloaded generic traits, in this case SaturatedFrom and SaturatedInto.

Saturating

Saturating arithmetic operations, returning maximum or minimum values instead of overflowing.

SendSignedTransaction

Submit a signed transaction to the transaction pool.

SendUnsignedTransaction

Submit an unsigned transaction onchain with a signed payload

Serialize

A data structure that can be serialized into any data format supported by Serde.

SignMessage

A message signer.

Signed

Useful functions for signed numbers (i.e. numbers that can be negative).

SignedPayload

Utility trait to be implemented on payloads that can be signed.

SigningTypes

A wrapper around the types which are used for signing.

StaticLookup

Means of changing one type into another in a manner dependent on the source type. This variant is different to Lookup in that it doesn’t (can cannot) require any context.

StorageInfoTrait

A trait to give information about storage.

StorageList

A non-continuous container type.

Task

A general-purpose trait which defines a type of service work (i.e., work to performed by an off-chain worker) including methods for enumerating, validating, indexing, and running tasks of this type.

ThresholdOrd

Trait for comparing two numbers with an threshold.

TransactionExtension

Means by which a transaction may be extended. This type embodies both the data and the logic that should be additionally associated with the transaction. It should be plain old data.

TypeInfo

Implementors return their meta type information.

TypedGet

A trait for querying a single value from a type defined in the trait.

UniqueSaturatedFrom

Just like From except that if the source value is too big to fit into the destination type then it’ll saturate the destination.

UniqueSaturatedInto

Just like Into except that if the source value is too big to fit into the destination type then it’ll saturate the destination.

Unsigned

A trait for values which cannot be negative

ValidateUnsigned

Provide validation for unsigned extrinsics.

VariantCount

Trait to get the number of variants in any enum.

Zero

Defines an additive identity element for Self.

Functions

blake2_64

Do a Blake2 64-bit hash and return result.

blake2_128

Do a Blake2 128-bit hash and return result.

blake2_256

Do a Blake2 256-bit hash and return result.

blake2_512

Do a Blake2 512-bit hash and return result.

blake2_512_into

Do a Blake2 512-bit hash and place result in dest.

build_state

Build GenesisConfig from a JSON blob not using any defaults and store it in the storage. For more info refer to sp_genesis_builder::GenesisBuilder::build_state.

checked_pow

Raises a value to the power of exp, returning None if an overflow occurred.

ensure_authorized

Ensure that the origin o represents an extrinsic with authorized call. Returns Ok or an Err otherwise.

ensure_none

Ensure that the origin o represents an unsigned extrinsic. Returns Ok or an Err otherwise.

ensure_pow

Raises a value to the power of exp, returning ArithmeticError if an overflow occurred.

ensure_root

Ensure that the origin o represents the root. Returns Ok or an Err otherwise.

ensure_signed

Ensure that the origin o represents a signed extrinsic (i.e. transaction). Returns Ok with the account that signed the extrinsic or an Err otherwise.

ensure_signed_or_root

Ensure that the origin o represents either a signed extrinsic (i.e. transaction) or the root. Returns Ok with the account that signed the extrinsic, None if it was root, or an Err otherwise.

get_preset

Get the default GenesisConfig as a JSON blob if name is None.

keccak_256

Do a keccak 256-bit hash and return result.

keccak_512

Do a keccak 512-bit hash and return result.

normalize

Normalize input so that the sum of all elements reaches targeted_sum.

sha2_256

Do a sha2 256-bit hash and return result.

twox_64

Do a XX 64-bit hash and return result.

twox_64_into

Do a XX 64-bit hash and place result in dest.

twox_128

Do a XX 128-bit hash and return result.

twox_256

Do a XX 256-bit hash and return result.

twox_128_into

Do a XX 128-bit hash and place result in dest.

twox_256_into

Do a XX 256-bit hash and place result in dest.

Type Aliases

AccountIdFor

Type alias for the AccountId associated type of system config.

ApplyExtrinsicResult

The result of applying of an extrinsic.

BlockNumberFor

Type alias for the BlockNumber associated type of system config.

CallOf

CheckedOf

DispatchInfoOf

Shortcut to reference the Info type of a Dispatchable.

DispatchResult

Un-augmented version of DispatchResultWithPostInfo that can be returned from dispatchable functions and is automatically converted to the augmented type. Should be used whenever the PostDispatchInfo does not need to be overwritten. As this should be the common case it is the implicit return type when none is specified.

DispatchResultWithInfo

Return type of a Dispatchable which contains the DispatchResult and additional information about the Dispatchable that is only known post dispatch.

DispatchResultWithPostInfo

The return type of a Dispatchable in frame. When returned explicitly from a dispatchable function it allows overriding the default PostDispatchInfo returned from a dispatch.

ExtensionVersion

Type to represent the version of the Extension used in this extrinsic.

ExtrinsicFor

Type alias for the Extrinsic associated type of system config.

GenesisBuilderResult

The result type alias, used in build methods. Err contains formatted error message.

HeaderFor

Type alias for the Header.

InherentIdentifier

An identifier for an inherent.

InnerOf

Get the inner type of a PerThing.

NoFee

An implementation of WeightToFee that collects no fee.

OriginFor

Type alias for the Origin associated type of system config.

OriginOf

PresetId

The type representing preset ID.

RuntimeCallFor

Type alias for the RuntimeCall associated type of system config.

TransactionLongevity

Minimum number of blocks a transaction will remain valid for. TransactionLongevity::max_value() means “forever”.

TransactionPriority

Priority for a transaction. Additive. Higher is better.

TransactionTag

Tag for a transaction. No two transactions with the same tag should be placed on-chain.

TransactionValidity

Information on a transaction’s validity and, if valid, on how it relates to other transactions.

TransactionValidityWithRefund

Information on a transaction’s validity and, if valid, on how it relates to other transactions and some refund for the operation.

UpperOf

Get the upper type of a PerThing.

ValidateResult

Shortcut for the result value of the validate function.

Attribute Macros

authorize

Allows to authorize some general transactions with specific dispatchable functions (dispatchable functions a.k.a. calls).

call

call_index

Each dispatchable may also be annotated with the #[pallet::call_index($idx)] attribute, which explicitly defines the codec index for the dispatchable function in the Call enum.

compact

composite_enum

config

constant

derive_impl

This attribute can be used to derive a full implementation of a trait based on a local partial impl and an external impl containing defaults that can be overridden in the local impl.

disable_frame_system_supertrait_check

disable_try_decode_storage

error

event

extra_constants

feeless_if

frame_construct_runtime

Construct a runtime, with the given name and the given pallets.

generate_deposit

genesis_build

genesis_config

getter

hooks

import_section

include_metadata

inherent

inject_runtime_type

The optional attribute #[inject_runtime_type] can be attached to RuntimeCall, RuntimeEvent, RuntimeOrigin or PalletInfo in an impl statement that has #[register_default_impl] attached to indicate that this item is generated by construct_runtime.

no_default

no_default_bounds

origin

pallet_section

register_default_impl

Attach this attribute to an impl statement that you want to use with #[derive_impl(..)].

runtime_version

storage

storage_prefix

storage_version

task_condition

Allows defining conditions for a task to run.

task_index

Allows defining an index for a task.

task_list

Allows defining an iterator over available work items for a task.

task_weight

Allows defining the weight of a task.

tasks_experimental

Allows you to define some service work that can be recognized by a script or an off-chain worker.

type_value

unbounded

validate_unsigned

view_functions

weight

weight_of_authorize

Allows to define the weight of the authorize function.

whitelist_storage

Derive Macros

CloneNoBound

Derive [Clone] but do not bound any generic.

Debug

Derive macro generating an impl of the trait Debug.

DebugNoBound

Derive [Debug] but do not bound any generics.

Decode

Derive parity_scale_codec::Decode for struct and enum.

DecodeWithMemTracking

Derive parity_scale_codec::DecodeWithMemTracking for struct and enum.

DefaultNoBound

derive Default but do no bound any generic. Docs are at frame_support::DefaultNoBound.

Deserialize

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.

EqNoBound

DeriveEq but do no bound any generic.

MaxEncodedLen

Derive parity_scale_codec::MaxEncodedLen for struct and enum.

OrdNoBound

Derive [Ord] but do no bound any generic. Docs are at frame_support::OrdNoBound.

PartialEqNoBound

Derive [PartialEq] but do not bound any generic.

PartialOrdNoBound

Derive [PartialOrd] but do not bound any generic. Docs are at frame_support::PartialOrdNoBound.

RuntimeDebug

RuntimeDebugNoBound

Derive [Debug], if std is enabled it uses frame_support::DebugNoBound, if std is not enabled it just returns "<wasm:stripped>". This behaviour is useful to prevent bloating the runtime WASM blob from unneeded code.

Serialize

TypeInfo