yet_another_parachain_runtime::polkadot_sdk_frame::traits

Trait InstanceFilter

pub trait InstanceFilter<T>:
    Sized
    + Send
    + Sync {
    // Required method
    fn filter(&self, _: &T) -> bool;

    // Provided method
    fn is_superset(&self, _o: &Self) -> bool { ... }
}

Simple trait for providing a filter over a reference to some type, given an instance of itself.

Required Methods

fn filter(&self, _: &T) -> bool

Determine if a given value should be allowed through the filter (returns true) or not.

Provided Methods

fn is_superset(&self, _o: &Self) -> bool

Determines whether self matches at least everything that _o does.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.