yet_another_parachain_runtime::polkadot_sdk_frame::traits

Trait TryDecodeEntireStorage

pub trait TryDecodeEntireStorage {
    // Required method
    fn try_decode_entire_state(    ) -> Result<usize, Vec<TryDecodeEntireStorageError>>;
}

Decode the entire data under the given storage type.

For values, this is trivial. For all kinds of maps, it should decode all the values associated with all keys existing in the map.

Tuple implementations are provided and simply decode each type in the tuple, summing up the decoded bytes if Ok(_).

Required Methods

fn try_decode_entire_state() -> Result<usize, Vec<TryDecodeEntireStorageError>>

Decode the entire data under the given storage, returning Ok(bytes_decoded) if success.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.