yet_another_parachain_runtime::sc_authority_discovery

Trait AuthorityDiscovery

pub trait AuthorityDiscovery<Block>
where
    Block: Block,
{
    // Required methods
    fn authorities<'life0, 'async_trait>(
        &'life0 self,
        at: <Block as Block>::Hash,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Public>, ApiError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn best_hash<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<<Block as Block>::Hash, Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
}

Wrapper for AuthorityDiscoveryApi. Can be be implemented by any struct without dependency on the runtime.

Required Methods

fn authorities<'life0, 'async_trait>( &'life0 self, at: <Block as Block>::Hash, ) -> Pin<Box<dyn Future<Output = Result<Vec<Public>, ApiError>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Retrieve authority identifiers of the current and next authority set.

fn best_hash<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<<Block as Block>::Hash, Error>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Retrieve best block hash