yet_another_parachain_runtime

Crate sc_cli

Substrate CLI library.

To see a full list of commands available, see commands.

Modules

arg_enums

Definitions of ValueEnum types.

clap

Command Line Argument Parser for Rust

commands

Various subcommands that can be included in a substrate-based chain’s CLI.

Macros

with_crypto_scheme

Allows for calling $method with appropriate crypto impl.

Structs

BlockNumberOrHash

Wrapper type that is either a Hash or the number of a Block.

BuildSpecCmd

The build-spec command used to build a specification.

ChainInfoCmd

The chain-info subcommand used to output db meta columns information.

CheckBlockCmd

The check-block command used to validate blocks.

CryptoSchemeFlag

Optional flag for specifying crypto algorithm

DatabaseParams

Parameters for database

ExecutionStrategiesParams

Execution strategies parameters.

ExportBlocksCmd

The export-blocks command used to export blocks.

ExportChainSpecCmd

Export a chain-spec to a JSON file in plain or in raw storage format.

ExportStateCmd

The export-state command used to export the state of a given block into a chain spec.

GenerateCmd

The generate command

GenerateKeyCmdCommon

Common arguments accross all generate key commands, subkey and node.

GenericNumber

Wrapper type of String that holds an unsigned integer of arbitrary size, formatted as a decimal.

ImportBlocksCmd

The import-blocks command used to import blocks.

ImportParams

Parameters for block import.

InsertKeyCmd

The insert command

InspectKeyCmd

The inspect command

InspectNodeKeyCmd

The inspect-node-key command

KeystoreParams

Parameters of the keystore

LoggerBuilder

A builder that is used to initialize the global logger.

MessageParams

Params to configure how a message should be passed into a command.

MixnetParams

Parameters used to create the mixnet configuration.

NetworkParams

Parameters used to create the network configuration.

NetworkSchemeFlag

Optional flag for specifying network scheme

NodeKeyParams

Parameters used to create the NodeKeyConfig, which determines the keypair used for libp2p networking.

OffchainWorkerParams

Offchain worker related parameters.

OutputTypeFlag

Optional flag for specifying output type

PrometheusParams

Parameters used to config prometheus.

PruningParams

Parameters to define the pruning mode

PurgeChainCmd

The purge-chain command used to remove the whole chain.

RevertCmd

The revert command used revert the chain to a previous state.

RpcEndpoint

Represent a single RPC endpoint with its configuration.

RpcParams

Parameters of RPC.

RunCmd

The run command used to run a node.

Runner

A Substrate CLI runtime that can be used to run a node or a command

RuntimeParams

Parameters used to config runtime.

RuntimeVersion

Runtime version. This should not be thought of as classic Semver (major/minor/tiny). This triplet have different semantics and mis-interpretation could cause problems. In particular: bug fixes should result in an increment of spec_version and possibly authoring_version, absolutely not impl_version since they change the semantics of the runtime.

SharedParams

Shared parameters used by all CoreParams.

SignCmd

The sign command

Signals

Abstraction over OS signals to handle the shutdown of the node smoothly.

TelemetryParams

Parameters used to config telemetry.

TransactionPoolParams

Parameters used to create the pool configuration.

VanityCmd

The vanity command

VerifyCmd

The verify command

Enums

Cors

CORS setting

CryptoScheme

The crypto scheme to use.

Database

Database backend

DatabasePruningMode

Specifies the pruning mode of the database.

Error

Error type for the CLI.

ExecutionStrategy

How to execute blocks

KeySubcommand

Key utilities for the cli.

NetworkBackendType

Network backend type.

NodeKeyType

The type of the node key.

OffchainWorkerEnabled

Whether off-chain workers are enabled.

OutputType

The type of the output format.

Role

Role of the local node.

RpcMethods

Available RPC methods.

SyncMode

Syncing mode.

TracingReceiver

TransactionPoolType

Type of transaction pool to be used

TrieCacheWarmUpStrategy

Warmup strategy for the trie cache.

WasmExecutionMethod

How to execute Wasm runtime code.

WasmtimeInstantiationStrategy

The instantiation strategy to use in compiled mode.

Constants

DEFAULT_WASMTIME_INSTANTIATION_STRATEGY

The default WasmtimeInstantiationStrategy.

DEFAULT_WASM_EXECUTION_METHOD

The default WasmExecutionMethod.

RPC_DEFAULT_MAX_CONNECTIONS

The default concurrent connection limit.

RPC_DEFAULT_MAX_REQUEST_SIZE_MB

The default max request size in MB.

RPC_DEFAULT_MAX_RESPONSE_SIZE_MB

The default max response size in MB.

RPC_DEFAULT_MAX_SUBS_PER_CONN

The default max number of subscriptions per connection.

RPC_DEFAULT_MESSAGE_CAPACITY_PER_CONN

The default number of messages the RPC server is allowed to keep in memory per connection.

RPC_DEFAULT_PORT

The default port.

Traits

ChainSpec

Common interface of a chain specification.

CliConfiguration

A trait that allows converting an object to a Configuration

DefaultConfigurationValues

Default configuration values used by Substrate

SubstrateCli

Substrate client CLI

Functions

build_runtime

Build a tokio runtime with all features.

execution_method_from_cli

Converts the execution method and instantiation strategy command line arguments into an execution method which can be used internally.

generate_node_name

Generate a valid random name for the node

parse_ss58_address_format

Parse Ss58AddressFormat

print_node_infos

Log information about the node itself.

secret_string_from_str

Parse a secret string, returning a displayable error.

Type Aliases

Result

Result type alias for the CLI.