yet_another_parachain_runtime::sc_cli::clap

Module builder

Define Command line arguments

Modules

styling

Terminal Styles for help and error output

Structs

Arg

The abstract representation of a command line argument. Used to set all the options and relationships that define a valid argument for the program.

ArgGroup

Family of related arguments.

BoolValueParser

Implementation for ValueParser::bool

BoolishValueParser

Parse bool-like string values, everything else is true

Command

Build a command-line interface.

EnumValueParser

Parse an ValueEnum value.

FalseyValueParser

Parse false-like string values, everything else is true

MapValueParser

Adapt a TypedValueParser from one value to another

NonEmptyStringValueParser

Parse non-empty string values

OsStr

A UTF-8-encoded fixed string

OsStringValueParser

Implementation for ValueParser::os_string

PathBufValueParser

Implementation for ValueParser::path_buf

PossibleValue

A possible value of an argument.

PossibleValuesParser

Verify the value is from an enumerated set of PossibleValue.

RangedI64ValueParser

Parse number that fall within a range of values

RangedU64ValueParser

Parse number that fall within a range of values

Str

A UTF-8-encoded fixed string

StringValueParser

Implementation for ValueParser::string

StyledStr

Terminal-styling container

Styles

Terminal styling definitions

TryMapValueParser

Adapt a TypedValueParser from one value to another

UnknownArgumentValueParser

When encountered, report ErrorKind::UnknownArgument

ValueParser

Parse/validate argument values

ValueRange

Values per occurrence for an argument

Enums

ArgAction

Behavior of arguments when they are encountered while parsing

ArgPredicate

Operations to perform on argument values

Resettable

Clearable builder value

ValueHint

Provide shell with hint on how to complete an argument.

Traits

IntoResettable

Convert to the intended resettable type

TypedValueParser

Parse/validate argument values

ValueParserFactory

Register a type with value_parser!