yet_another_parachain_runtime::sc_cli

Module commands

Various subcommands that can be included in a substrate-based chain’s CLI.

Modules

utils

subcommand utilities

Structs

BuildSpecCmd

The build-spec command used to build a specification.

ChainInfoCmd

The chain-info subcommand used to output db meta columns information.

CheckBlockCmd

The check-block command used to validate blocks.

ExportBlocksCmd

The export-blocks command used to export blocks.

ExportChainSpecCmd

Export a chain-spec to a JSON file in plain or in raw storage format.

ExportStateCmd

The export-state command used to export the state of a given block into a chain spec.

GenerateCmd

The generate command

GenerateKeyCmdCommon

Common arguments accross all generate key commands, subkey and node.

ImportBlocksCmd

The import-blocks command used to import blocks.

InsertKeyCmd

The insert command

InspectKeyCmd

The inspect command

InspectNodeKeyCmd

The inspect-node-key command

PurgeChainCmd

The purge-chain command used to remove the whole chain.

RevertCmd

The revert command used revert the chain to a previous state.

RunCmd

The run command used to run a node.

SignCmd

The sign command

VanityCmd

The vanity command

VerifyCmd

The verify command

Enums

KeySubcommand

Key utilities for the cli.