yet_another_parachain_runtime::sc_cli

Trait CliConfiguration

pub trait CliConfiguration<DCV = ()>: Sized
where
    DCV: DefaultConfigurationValues,
{

    // Required method
    fn shared_params(&self) -> &SharedParams;

    // Provided methods
    fn import_params(&self) -> Option<&ImportParams> { ... }
    fn pruning_params(&self) -> Option<&PruningParams> { ... }
    fn keystore_params(&self) -> Option<&KeystoreParams> { ... }
    fn network_params(&self) -> Option<&NetworkParams> { ... }
    fn offchain_worker_params(&self) -> Option<&OffchainWorkerParams> { ... }
    fn node_key_params(&self) -> Option<&NodeKeyParams> { ... }
    fn database_params(&self) -> Option<&DatabaseParams> { ... }
    fn base_path(&self) -> Result<Option<BasePath>, Error> { ... }
    fn is_dev(&self) -> Result<bool, Error> { ... }
    fn role(&self, _is_dev: bool) -> Result<Role, Error> { ... }
    fn transaction_pool(
        &self,
        _is_dev: bool,
    ) -> Result<TransactionPoolOptions, Error> { ... }
    fn network_config(
        &self,
        chain_spec: &Box<dyn ChainSpec>,
        is_dev: bool,
        is_validator: bool,
        net_config_dir: PathBuf,
        client_id: &str,
        node_name: &str,
        node_key: NodeKeyConfig,
        default_listen_port: u16,
    ) -> Result<NetworkConfiguration, Error> { ... }
    fn keystore_config(
        &self,
        config_dir: &PathBuf,
    ) -> Result<KeystoreConfig, Error> { ... }
    fn database_cache_size(&self) -> Result<Option<usize>, Error> { ... }
    fn database(&self) -> Result<Option<Database>, Error> { ... }
    fn database_config(
        &self,
        base_path: &PathBuf,
        cache_size: usize,
        database: Database,
    ) -> Result<DatabaseSource, Error> { ... }
    fn trie_cache_maximum_size(&self) -> Result<Option<usize>, Error> { ... }
    fn warm_up_trie_cache(
        &self,
    ) -> Result<Option<TrieCacheWarmUpStrategy>, Error> { ... }
    fn state_pruning(&self) -> Result<Option<PruningMode>, Error> { ... }
    fn blocks_pruning(&self) -> Result<BlocksPruning, Error> { ... }
    fn chain_id(&self, is_dev: bool) -> Result<String, Error> { ... }
    fn node_name(&self) -> Result<String, Error> { ... }
    fn wasm_method(&self) -> Result<WasmExecutionMethod, Error> { ... }
    fn wasm_runtime_overrides(&self) -> Option<PathBuf> { ... }
    fn rpc_addr(
        &self,
        _default_listen_port: u16,
    ) -> Result<Option<Vec<RpcEndpoint>>, Error> { ... }
    fn rpc_methods(&self) -> Result<RpcMethods, Error> { ... }
    fn rpc_max_connections(&self) -> Result<u32, Error> { ... }
    fn rpc_cors(&self, _is_dev: bool) -> Result<Option<Vec<String>>, Error> { ... }
    fn rpc_max_request_size(&self) -> Result<u32, Error> { ... }
    fn rpc_max_response_size(&self) -> Result<u32, Error> { ... }
    fn rpc_max_subscriptions_per_connection(&self) -> Result<u32, Error> { ... }
    fn rpc_buffer_capacity_per_connection(&self) -> Result<u32, Error> { ... }
    fn rpc_batch_config(&self) -> Result<BatchRequestConfig, Error> { ... }
    fn rpc_rate_limit(&self) -> Result<Option<NonZero<u32>>, Error> { ... }
    fn rpc_rate_limit_whitelisted_ips(&self) -> Result<Vec<IpNetwork>, Error> { ... }
    fn rpc_rate_limit_trust_proxy_headers(&self) -> Result<bool, Error> { ... }
    fn prometheus_config(
        &self,
        _default_listen_port: u16,
        _chain_spec: &Box<dyn ChainSpec>,
    ) -> Result<Option<PrometheusConfig>, Error> { ... }
    fn telemetry_endpoints(
        &self,
        chain_spec: &Box<dyn ChainSpec>,
    ) -> Result<Option<TelemetryEndpoints>, Error> { ... }
    fn default_heap_pages(&self) -> Result<Option<u64>, Error> { ... }
    fn offchain_worker(
        &self,
        role: &Role,
    ) -> Result<OffchainWorkerConfig, Error> { ... }
    fn force_authoring(&self) -> Result<bool, Error> { ... }
    fn disable_grandpa(&self) -> Result<bool, Error> { ... }
    fn dev_key_seed(&self, _is_dev: bool) -> Result<Option<String>, Error> { ... }
    fn tracing_targets(&self) -> Result<Option<String>, Error> { ... }
    fn tracing_receiver(&self) -> Result<TracingReceiver, Error> { ... }
    fn node_key(&self, net_config_dir: &PathBuf) -> Result<NodeKeyConfig, Error> { ... }
    fn max_runtime_instances(&self) -> Result<Option<usize>, Error> { ... }
    fn runtime_cache_size(&self) -> Result<u8, Error> { ... }
    fn announce_block(&self) -> Result<bool, Error> { ... }
    fn create_configuration<C>(
        &self,
        cli: &C,
        tokio_handle: Handle,
    ) -> Result<Configuration, Error>
       where C: SubstrateCli { ... }
    fn log_filters(&self) -> Result<String, Error> { ... }
    fn detailed_log_output(&self) -> Result<bool, Error> { ... }
    fn enable_log_reloading(&self) -> Result<bool, Error> { ... }
    fn disable_log_color(&self) -> Result<bool, Error> { ... }
    fn init<F>(
        &self,
        support_url: &String,
        impl_version: &String,
        logger_hook: F,
    ) -> Result<(), Error>
       where F: FnOnce(&mut LoggerBuilder) { ... }

}

A trait that allows converting an object to a Configuration

Required Methods

fn shared_params(&self) -> &SharedParams

Get the SharedParams for this object

Provided Methods

fn import_params(&self) -> Option<&ImportParams>

Get the ImportParams for this object

fn pruning_params(&self) -> Option<&PruningParams>

Get the PruningParams for this object

fn keystore_params(&self) -> Option<&KeystoreParams>

Get the KeystoreParams for this object

fn network_params(&self) -> Option<&NetworkParams>

Get the NetworkParams for this object

fn offchain_worker_params(&self) -> Option<&OffchainWorkerParams>

Get a reference to OffchainWorkerParams for this object.

fn node_key_params(&self) -> Option<&NodeKeyParams>

Get the NodeKeyParams for this object

fn database_params(&self) -> Option<&DatabaseParams>

Get the DatabaseParams for this object

fn base_path(&self) -> Result<Option<BasePath>, Error>

Get the base path of the configuration (if any)

By default this is retrieved from SharedParams.

fn is_dev(&self) -> Result<bool, Error>

Returns true if the node is for development or not

By default this is retrieved from SharedParams.

fn role(&self, _is_dev: bool) -> Result<Role, Error>

Gets the role

By default this is Role::Full.

fn transaction_pool( &self, _is_dev: bool, ) -> Result<TransactionPoolOptions, Error>

Get the transaction pool options

By default this is TransactionPoolOptions::default().

fn network_config( &self, chain_spec: &Box<dyn ChainSpec>, is_dev: bool, is_validator: bool, net_config_dir: PathBuf, client_id: &str, node_name: &str, node_key: NodeKeyConfig, default_listen_port: u16, ) -> Result<NetworkConfiguration, Error>

Get the network configuration

By default this is retrieved from NetworkParams if it is available otherwise it creates a default NetworkConfiguration based on node_name, client_id, node_key and net_config_dir.

fn keystore_config(&self, config_dir: &PathBuf) -> Result<KeystoreConfig, Error>

Get the keystore configuration.

By default this is retrieved from KeystoreParams if it is available. Otherwise it uses KeystoreConfig::InMemory.

fn database_cache_size(&self) -> Result<Option<usize>, Error>

Get the database cache size.

By default this is retrieved from DatabaseParams if it is available. Otherwise its None.

fn database(&self) -> Result<Option<Database>, Error>

Get the database backend variant.

By default this is retrieved from DatabaseParams if it is available. Otherwise its None.

fn database_config( &self, base_path: &PathBuf, cache_size: usize, database: Database, ) -> Result<DatabaseSource, Error>

Get the database configuration object for the parameters provided

fn trie_cache_maximum_size(&self) -> Result<Option<usize>, Error>

Get the trie cache maximum size.

By default this is retrieved from ImportParams if it is available. Otherwise its 0. If None is returned the trie cache is disabled.

fn warm_up_trie_cache(&self) -> Result<Option<TrieCacheWarmUpStrategy>, Error>

Get if we should warm up the trie cache.

By default this is retrieved from ImportParams if it is available. Otherwise its None.

fn state_pruning(&self) -> Result<Option<PruningMode>, Error>

Get the state pruning mode.

By default this is retrieved from PruningMode if it is available. Otherwise its PruningMode::default().

fn blocks_pruning(&self) -> Result<BlocksPruning, Error>

Get the block pruning mode.

By default this is retrieved from block_pruning if it is available. Otherwise its BlocksPruning::KeepFinalized.

fn chain_id(&self, is_dev: bool) -> Result<String, Error>

Get the chain ID (string).

By default this is retrieved from SharedParams.

fn node_name(&self) -> Result<String, Error>

Get the name of the node.

By default a random name is generated.

fn wasm_method(&self) -> Result<WasmExecutionMethod, Error>

Get the WASM execution method.

By default this is retrieved from ImportParams if it is available. Otherwise its WasmExecutionMethod::default().

fn wasm_runtime_overrides(&self) -> Option<PathBuf>

Get the path where WASM overrides live.

By default this is None.

fn rpc_addr( &self, _default_listen_port: u16, ) -> Result<Option<Vec<RpcEndpoint>>, Error>

Get the RPC address.

fn rpc_methods(&self) -> Result<RpcMethods, Error>

Returns the RPC method set to expose.

By default this is RpcMethods::Auto (unsafe RPCs are denied iff rpc_external returns true, respectively).

fn rpc_max_connections(&self) -> Result<u32, Error>

Get the maximum number of RPC server connections.

fn rpc_cors(&self, _is_dev: bool) -> Result<Option<Vec<String>>, Error>

Get the RPC cors (None if disabled)

By default this is Some(Vec::new()).

fn rpc_max_request_size(&self) -> Result<u32, Error>

Get maximum RPC request payload size.

fn rpc_max_response_size(&self) -> Result<u32, Error>

Get maximum RPC response payload size.

fn rpc_max_subscriptions_per_connection(&self) -> Result<u32, Error>

Get maximum number of subscriptions per connection.

fn rpc_buffer_capacity_per_connection(&self) -> Result<u32, Error>

The number of messages the RPC server is allowed to keep in memory per connection.

fn rpc_batch_config(&self) -> Result<BatchRequestConfig, Error>

RPC server batch request configuration.

fn rpc_rate_limit(&self) -> Result<Option<NonZero<u32>>, Error>

RPC rate limit configuration.

fn rpc_rate_limit_whitelisted_ips(&self) -> Result<Vec<IpNetwork>, Error>

RPC rate limit whitelisted ip addresses.

fn rpc_rate_limit_trust_proxy_headers(&self) -> Result<bool, Error>

RPC rate limit trust proxy headers.

fn prometheus_config( &self, _default_listen_port: u16, _chain_spec: &Box<dyn ChainSpec>, ) -> Result<Option<PrometheusConfig>, Error>

Get the prometheus configuration (None if disabled)

By default this is None.

fn telemetry_endpoints( &self, chain_spec: &Box<dyn ChainSpec>, ) -> Result<Option<TelemetryEndpoints>, Error>

Get the telemetry endpoints (if any)

By default this is retrieved from the chain spec loaded by load_spec.

fn default_heap_pages(&self) -> Result<Option<u64>, Error>

Get the default value for heap pages

By default this is None.

fn offchain_worker(&self, role: &Role) -> Result<OffchainWorkerConfig, Error>

Returns an offchain worker config wrapped in Ok(_)

By default offchain workers are disabled.

fn force_authoring(&self) -> Result<bool, Error>

Returns Ok(true) if authoring should be forced

By default this is false.

fn disable_grandpa(&self) -> Result<bool, Error>

Returns Ok(true) if grandpa should be disabled

By default this is false.

fn dev_key_seed(&self, _is_dev: bool) -> Result<Option<String>, Error>

Get the development key seed from the current object

By default this is None.

fn tracing_targets(&self) -> Result<Option<String>, Error>

Get the tracing targets from the current object (if any)

By default this is retrieved from SharedParams if it is available. Otherwise its None.

fn tracing_receiver(&self) -> Result<TracingReceiver, Error>

Get the TracingReceiver value from the current object

By default this is retrieved from SharedParams if it is available. Otherwise its TracingReceiver::default().

fn node_key(&self, net_config_dir: &PathBuf) -> Result<NodeKeyConfig, Error>

Get the node key from the current object

By default this is retrieved from NodeKeyParams if it is available. Otherwise its NodeKeyConfig::default().

fn max_runtime_instances(&self) -> Result<Option<usize>, Error>

Get maximum runtime instances

By default this is None.

fn runtime_cache_size(&self) -> Result<u8, Error>

Get maximum different runtimes in cache

By default this is 2.

fn announce_block(&self) -> Result<bool, Error>

Activate or not the automatic announcing of blocks after import

By default this is false.

fn create_configuration<C>( &self, cli: &C, tokio_handle: Handle, ) -> Result<Configuration, Error>
where C: SubstrateCli,

Create a Configuration object from the current object

fn log_filters(&self) -> Result<String, Error>

Get the filters for the logging.

This should be a list of comma-separated values. Example: foo=trace,bar=debug,baz=info

By default this is retrieved from SharedParams.

fn detailed_log_output(&self) -> Result<bool, Error>

Should the detailed log output be enabled.

fn enable_log_reloading(&self) -> Result<bool, Error>

Is log reloading enabled?

fn disable_log_color(&self) -> Result<bool, Error>

Should the log color output be disabled?

fn init<F>( &self, support_url: &String, impl_version: &String, logger_hook: F, ) -> Result<(), Error>
where F: FnOnce(&mut LoggerBuilder),

Initialize substrate. This must be done only once per process.

This method:

Sets the panic handler

Optionally customize logger/profiling

Initializes the logger

Raises the FD limit

The logger_hook closure is executed before the logger is constructed and initialized. It is useful for setting up a custom profiler.

Example:

use sc_tracing::{SpanDatum, TraceEvent};
struct TestProfiler;

impl sc_tracing::TraceHandler for TestProfiler {
 	fn handle_span(&self, sd: &SpanDatum) {}
		fn handle_event(&self, _event: &TraceEvent) {}
};

fn logger_hook() -> impl FnOnce(&mut sc_cli::LoggerBuilder, &sc_service::Configuration) -> () {
	|logger_builder, config| {
			logger_builder.with_custom_profiling(Box::new(TestProfiler{}));
	}
}

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.