yet_another_parachain_runtime

Crate sc_client_api

Substrate client interfaces.

Modules

backend

Substrate Client data backend

blockchain

Substrate blockchain traits and primitives.

call_executor

A method call executor interface.

client

A set of APIs supported by the client along with their primitives.

execution_extensions

Execution extensions for runtime calls.

in_mem

In memory client backend

leaves

Helper for managing the set of available leaves in the chain for DB implementations.

notifications

Storage notifications

proof_provider

Proof utilities

utils

Utility methods for the client.

Structs

BlockImportNotification

Summary of an imported block

ClientImportOperation

Import operation wrapper.

ClientInfo

Client info

CompactProof

Storage proof in compact form.

FinalityNotification

Summary of a finalized block.

FinalizeSummary

Finalization operation summary.

ImportSummary

Import operation summary.

IoInfo

I/O statistics for client instance.

KeyValueStates

Multiple key value state. States are ordered by root storage key.

KeysIter

An Iterator that iterates keys in a given block under a prefix.

MemoryInfo

Memory statistics for client instance.

MemorySize

A wrapper to store the size of some memory.

PairsIter

An Iterator that iterates keys and values in a given block under a prefix.

PrefixedStorageKey

Storage key of a child trie, it contains the prefix to the key.

StaleBlock

A stale block.

StorageChangeSet

Storage change set

StorageData

Storage data associated to a StorageKey.

StorageEventStream

Type that implements futures::Stream of storage change events.

StorageKey

Storage key.

StorageNotification

A type of a message delivered to the subscribers

StorageNotifications

Manages storage listeners.

StorageProof

A proof that some set of key-value pairs are included in the storage trie. The proof contains the storage values so that the partial storage backend can be reconstructed by a verifier that does not already have access to the key-value pairs.

UnpinHandle

Keeps a specific block pinned while the handle is alive. Once the last handle instance for a given block is dropped, the block is unpinned in the Backend.

UnpinHandleInner

Sends a message to the pinning-worker once dropped to unpin a block in the backend.

UsageInfo

Usage statistics for running client instance.

Enums

ChildInfo

Information related to a child state.

ImportNotificationAction

Describes which block import notification stream should be notified.

MerkleValue

Either the hash or value of a node depending on its size.

NewBlockState

State of a new block.

TrieCacheContext

Specify the desired trie cache context when calling Backend::state_at.

UnpinWorkerMessage

Message that signals notification-based pinning actions to the pinning-worker.

Traits

AuxStore

Provides access to an auxiliary database.

Backend

Client backend.

BlockBackend

Interface for fetching block data.

BlockImportOperation

Block insertion operation.

BlockOf

Figure out the block type for a given type (for now, just a Client).

BlockchainEvents

A source of blockchain events.

CallExecutor

Method call executor.

ExecutorProvider

Executor Provider

Finalizer

Finalize Facilities

HeaderBackend

Blockchain database header backend. Does not perform any validation.

LocalBackend

Mark for all Backend implementations, that are making use of state data, stored locally.

LockImportRun

Interface for performing operations on the backend.

PreCommitActions

Interface to perform auxiliary actions before committing a block import or finality operation.

ProofProvider

Interface for providing block proving utilities.

ProvideUncles

Provide a list of potential uncle headers for a given block.

StateBackend

A state backend is used to read state data and can have changes committed to it.

StorageProvider

Provides access to storage primitives

UsageProvider

Usage Information Provider interface

Functions

apply_aux

Helper function to apply auxiliary data insertion into an operation.

Type Aliases

AuxDataOperations

List of operations to be performed on storage aux data. First tuple element is the encoded data key. Second tuple element is the encoded optional data to write. If None, the key and the associated data are deleted from storage.

BackendTransaction

The transaction type used by Backend.

BadBlocks

Known bad block hashes.

FinalityNotifications

A stream of block finality notifications.

ForkBlocks

Expected hashes of blocks at given heights.

ImportNotifications

Type that implements futures::Stream of block import events.

OnFinalityAction

Callback invoked before committing the operations created during block finalization. This gives the opportunity to perform auxiliary pre-commit actions and optionally enqueue further storage write operations to be atomically performed on commit.

OnImportAction

Callback invoked before committing the operations created during block import. This gives the opportunity to perform auxiliary pre-commit actions and optionally enqueue further storage write operations to be atomically performed on commit.

StateBackendFor

Extracts the state backend type for the given backend.