yet_another_parachain_runtime::sc_client_api

Module backend

Substrate Client data backend

Structs

ClientImportOperation

Import operation wrapper.

FinalizeSummary

Finalization operation summary.

ImportSummary

Import operation summary.

KeyValueStates

Multiple key value state. States are ordered by root storage key.

KeysIter

An Iterator that iterates keys in a given block under a prefix.

PairsIter

An Iterator that iterates keys and values in a given block under a prefix.

StaleBlock

A stale block.

Enums

ImportNotificationAction

Describes which block import notification stream should be notified.

MerkleValue

Either the hash or value of a node depending on its size.

NewBlockState

State of a new block.

TrieCacheContext

Specify the desired trie cache context when calling Backend::state_at.

Traits

AuxStore

Provides access to an auxiliary database.

Backend

Client backend.

BlockImportOperation

Block insertion operation.

Finalizer

Finalize Facilities

LocalBackend

Mark for all Backend implementations, that are making use of state data, stored locally.

LockImportRun

Interface for performing operations on the backend.

StateBackend

A state backend is used to read state data and can have changes committed to it.

StorageProvider

Provides access to storage primitives

Functions

apply_aux

Helper function to apply auxiliary data insertion into an operation.

Type Aliases

BackendTransaction

The transaction type used by Backend.

StateBackendFor

Extracts the state backend type for the given backend.