yet_another_parachain_runtime::sc_client_api

Module client

A set of APIs supported by the client along with their primitives.

Structs

BlockImportNotification

Summary of an imported block

ClientInfo

Client info

FinalityNotification

Summary of a finalized block.

IoInfo

I/O statistics for client instance.

MemoryInfo

Memory statistics for client instance.

MemorySize

A wrapper to store the size of some memory.

UnpinHandle

Keeps a specific block pinned while the handle is alive. Once the last handle instance for a given block is dropped, the block is unpinned in the Backend.

UnpinHandleInner

Sends a message to the pinning-worker once dropped to unpin a block in the backend.

UsageInfo

Usage statistics for running client instance.

Enums

UnpinWorkerMessage

Message that signals notification-based pinning actions to the pinning-worker.

Traits

BlockBackend

Interface for fetching block data.

BlockOf

Figure out the block type for a given type (for now, just a Client).

BlockchainEvents

A source of blockchain events.

PreCommitActions

Interface to perform auxiliary actions before committing a block import or finality operation.

ProvideUncles

Provide a list of potential uncle headers for a given block.

Type Aliases

AuxDataOperations

List of operations to be performed on storage aux data. First tuple element is the encoded data key. Second tuple element is the encoded optional data to write. If None, the key and the associated data are deleted from storage.

BadBlocks

Known bad block hashes.

FinalityNotifications

A stream of block finality notifications.

ForkBlocks

Expected hashes of blocks at given heights.

ImportNotifications

Type that implements futures::Stream of block import events.

OnFinalityAction

Callback invoked before committing the operations created during block finalization. This gives the opportunity to perform auxiliary pre-commit actions and optionally enqueue further storage write operations to be atomically performed on commit.

OnImportAction

Callback invoked before committing the operations created during block import. This gives the opportunity to perform auxiliary pre-commit actions and optionally enqueue further storage write operations to be atomically performed on commit.