yet_another_parachain_runtime::sc_client_api

Trait BlockImportOperation

pub trait BlockImportOperation<Block>
where
    Block: Block,
{
    type State: Backend<<<Block as Block>::Header as Header>::Hashing>;

    // Required methods
    fn state(&self) -> Result<Option<&Self::State>, Error>;
    fn set_block_data(
        &mut self,
        header: <Block as Block>::Header,
        body: Option<Vec<<Block as Block>::Extrinsic>>,
        indexed_body: Option<Vec<Vec<u8>>>,
        justifications: Option<Justifications>,
        state: NewBlockState,
    ) -> Result<(), Error>;
    fn update_db_storage(
        &mut self,
        update: MemoryDB<<<Block as Block>::Header as Header>::Hashing, PrefixedKey<<<Block as Block>::Header as Header>::Hashing>, Vec<u8>>,
    ) -> Result<(), Error>;
    fn set_genesis_state(
        &mut self,
        storage: Storage,
        commit: bool,
        state_version: StateVersion,
    ) -> Result<<Block as Block>::Hash, Error>;
    fn reset_storage(
        &mut self,
        storage: Storage,
        state_version: StateVersion,
    ) -> Result<<Block as Block>::Hash, Error>;
    fn update_storage(
        &mut self,
        update: Vec<(Vec<u8>, Option<Vec<u8>>)>,
        child_update: Vec<(Vec<u8>, Vec<(Vec<u8>, Option<Vec<u8>>)>)>,
    ) -> Result<(), Error>;
    fn insert_aux<I>(&mut self, ops: I) -> Result<(), Error>
       where I: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>;
    fn mark_finalized(
        &mut self,
        hash: <Block as Block>::Hash,
        justification: Option<([u8; 4], Vec<u8>)>,
    ) -> Result<(), Error>;
    fn mark_head(&mut self, hash: <Block as Block>::Hash) -> Result<(), Error>;
    fn update_transaction_index(
        &mut self,
        index: Vec<IndexOperation>,
    ) -> Result<(), Error>;
    fn set_create_gap(&mut self, create_gap: bool);

    // Provided method
    fn update_offchain_storage(
        &mut self,
        _offchain_update: Vec<((Vec<u8>, Vec<u8>), OffchainOverlayedChange)>,
    ) -> Result<(), Error> { ... }
}

Block insertion operation.

Keeps hold if the inserted block state and data.

Required Associated Types

type State: Backend<<<Block as Block>::Header as Header>::Hashing>

Associated state backend type.

Required Methods

fn state(&self) -> Result<Option<&Self::State>, Error>

Returns pending state.

Returns None for backends with locally-unavailable state data.

fn set_block_data( &mut self, header: <Block as Block>::Header, body: Option<Vec<<Block as Block>::Extrinsic>>, indexed_body: Option<Vec<Vec<u8>>>, justifications: Option<Justifications>, state: NewBlockState, ) -> Result<(), Error>

Append block data to the transaction.

fn update_db_storage( &mut self, update: MemoryDB<<<Block as Block>::Header as Header>::Hashing, PrefixedKey<<<Block as Block>::Header as Header>::Hashing>, Vec<u8>>, ) -> Result<(), Error>

Inject storage data into the database.

fn set_genesis_state( &mut self, storage: Storage, commit: bool, state_version: StateVersion, ) -> Result<<Block as Block>::Hash, Error>

Set genesis state. If commit is false the state is saved in memory, but is not written to the database.

fn reset_storage( &mut self, storage: Storage, state_version: StateVersion, ) -> Result<<Block as Block>::Hash, Error>

Inject storage data into the database replacing any existing data.

fn update_storage( &mut self, update: Vec<(Vec<u8>, Option<Vec<u8>>)>, child_update: Vec<(Vec<u8>, Vec<(Vec<u8>, Option<Vec<u8>>)>)>, ) -> Result<(), Error>

Set storage changes.

fn insert_aux<I>(&mut self, ops: I) -> Result<(), Error>
where I: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,

Insert auxiliary keys.

Values are None if should be deleted.

fn mark_finalized( &mut self, hash: <Block as Block>::Hash, justification: Option<([u8; 4], Vec<u8>)>, ) -> Result<(), Error>

Mark a block as finalized, if multiple blocks are finalized in the same operation then they must be marked in ascending order.

fn mark_head(&mut self, hash: <Block as Block>::Hash) -> Result<(), Error>

Mark a block as new head. If both block import and set head are specified, set head overrides block import’s best block rule.

fn update_transaction_index( &mut self, index: Vec<IndexOperation>, ) -> Result<(), Error>

Add a transaction index operation.

fn set_create_gap(&mut self, create_gap: bool)

Configure whether to create a block gap if newly imported block is missing parent

Provided Methods

fn update_offchain_storage( &mut self, _offchain_update: Vec<((Vec<u8>, Vec<u8>), OffchainOverlayedChange)>, ) -> Result<(), Error>

Write offchain storage changes to the database.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.