yet_another_parachain_runtime::sc_client_api

Trait StorageProvider

pub trait StorageProvider<Block, B>
where
    Block: Block,
    B: Backend<Block>,
{
    // Required methods
    fn storage(
        &self,
        hash: <Block as Block>::Hash,
        key: &StorageKey,
    ) -> Result<Option<StorageData>, Error>;
    fn storage_hash(
        &self,
        hash: <Block as Block>::Hash,
        key: &StorageKey,
    ) -> Result<Option<<Block as Block>::Hash>, Error>;
    fn storage_keys(
        &self,
        hash: <Block as Block>::Hash,
        prefix: Option<&StorageKey>,
        start_key: Option<&StorageKey>,
    ) -> Result<KeysIter<<B as Backend<Block>>::State, Block>, Error>;
    fn storage_pairs(
        &self,
        hash: <Block as Block>::Hash,
        prefix: Option<&StorageKey>,
        start_key: Option<&StorageKey>,
    ) -> Result<PairsIter<<B as Backend<Block>>::State, Block>, Error>;
    fn child_storage(
        &self,
        hash: <Block as Block>::Hash,
        child_info: &ChildInfo,
        key: &StorageKey,
    ) -> Result<Option<StorageData>, Error>;
    fn child_storage_keys(
        &self,
        hash: <Block as Block>::Hash,
        child_info: ChildInfo,
        prefix: Option<&StorageKey>,
        start_key: Option<&StorageKey>,
    ) -> Result<KeysIter<<B as Backend<Block>>::State, Block>, Error>;
    fn child_storage_hash(
        &self,
        hash: <Block as Block>::Hash,
        child_info: &ChildInfo,
        key: &StorageKey,
    ) -> Result<Option<<Block as Block>::Hash>, Error>;
    fn closest_merkle_value(
        &self,
        hash: <Block as Block>::Hash,
        key: &StorageKey,
    ) -> Result<Option<MerkleValue<<Block as Block>::Hash>>, Error>;
    fn child_closest_merkle_value(
        &self,
        hash: <Block as Block>::Hash,
        child_info: &ChildInfo,
        key: &StorageKey,
    ) -> Result<Option<MerkleValue<<Block as Block>::Hash>>, Error>;
}

Provides access to storage primitives

Required Methods

fn storage( &self, hash: <Block as Block>::Hash, key: &StorageKey, ) -> Result<Option<StorageData>, Error>

Given a block’s Hash and a key, return the value under the key in that block.

fn storage_hash( &self, hash: <Block as Block>::Hash, key: &StorageKey, ) -> Result<Option<<Block as Block>::Hash>, Error>

Given a block’s Hash and a key, return the value under the hash in that block.

fn storage_keys( &self, hash: <Block as Block>::Hash, prefix: Option<&StorageKey>, start_key: Option<&StorageKey>, ) -> Result<KeysIter<<B as Backend<Block>>::State, Block>, Error>

Given a block’s Hash and a key prefix, returns a KeysIter iterates matching storage keys in that block.

fn storage_pairs( &self, hash: <Block as Block>::Hash, prefix: Option<&StorageKey>, start_key: Option<&StorageKey>, ) -> Result<PairsIter<<B as Backend<Block>>::State, Block>, Error>

Given a block’s Hash and a key prefix, returns an iterator over the storage keys and values in that block.

fn child_storage( &self, hash: <Block as Block>::Hash, child_info: &ChildInfo, key: &StorageKey, ) -> Result<Option<StorageData>, Error>

Given a block’s Hash, a key and a child storage key, return the value under the key in that block.

fn child_storage_keys( &self, hash: <Block as Block>::Hash, child_info: ChildInfo, prefix: Option<&StorageKey>, start_key: Option<&StorageKey>, ) -> Result<KeysIter<<B as Backend<Block>>::State, Block>, Error>

Given a block’s Hash and a key prefix and a child storage key, returns a KeysIter that iterates matching storage keys in that block.

fn child_storage_hash( &self, hash: <Block as Block>::Hash, child_info: &ChildInfo, key: &StorageKey, ) -> Result<Option<<Block as Block>::Hash>, Error>

Given a block’s Hash, a key and a child storage key, return the hash under the key in that block.

fn closest_merkle_value( &self, hash: <Block as Block>::Hash, key: &StorageKey, ) -> Result<Option<MerkleValue<<Block as Block>::Hash>>, Error>

Given a block’s Hash and a key, return the closest merkle value.

fn child_closest_merkle_value( &self, hash: <Block as Block>::Hash, child_info: &ChildInfo, key: &StorageKey, ) -> Result<Option<MerkleValue<<Block as Block>::Hash>>, Error>

Given a block’s Hash, a key and a child storage key, return the closest merkle value.