yet_another_parachain_runtime

Crate sc_consensus

Collection of common consensus specific implementations

Modules

block_import

Block import helpers.

import_queue

Import Queue primitive: something which can verify and import blocks.

metrics

Metering tools for consensus

shared_data

Provides a generic wrapper around shared data. See SharedData for more information.

Structs

BasicQueue

Interface to a basic block import queue that is importing blocks sequentially in a separate task, with plugable verification.

BlockCheckParams

Data required to check validity of a Block.

BlockImportParams

Data required to import a Block.

ImportedAux

Auxiliary data associated with an imported block result.

ImportedState

Imported state data. A vector of key-value pairs that should form a trie.

IncomingBlock

Block data used by the queue.

LongestChain

Implement Longest Chain Select implementation where ‘longest’ is defined as the highest number of blocks

Enums

BlockImportError

Block import error.

BlockImportStatus

Block import successful result.

ForkChoiceStrategy

Fork choice strategy.

ImportResult

Block import result.

JustificationImportResult

The result of importing a justification.

StateAction

Defines how a new state is computed for a given imported block.

StorageChanges

Precomputed storage.

Traits

BlockImport

Block import trait.

ImportQueue

JustificationImport

Justification import trait

JustificationSyncLink

Control the synchronization process of block justifications.

Link

Hooks that the verification queue can use to influence the synchronization algorithm.

Verifier

Verify a justification of a block

Functions

import_single_block

Single block import function.

Type Aliases

BoxBlockImport

Shared block import struct used by the queue.

BoxJustificationImport

Shared justification import struct used by the queue.

DefaultImportQueue

A commonly-used Import Queue type.