yet_another_parachain_runtime::sc_consensus

Trait JustificationImport

pub trait JustificationImport<B>
where
    B: Block,
{
    type Error: Error + Send + 'static;

    // Required methods
    fn on_start<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Vec<(<B as Block>::Hash, <<B as Block>::Header as Header>::Number)>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn import_justification<'life0, 'async_trait>(
        &'life0 mut self,
        hash: <B as Block>::Hash,
        number: <<B as Block>::Header as Header>::Number,
        justification: ([u8; 4], Vec<u8>),
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
}

Justification import trait

Required Associated Types

type Error: Error + Send + 'static

Required Methods

fn on_start<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Vec<(<B as Block>::Hash, <<B as Block>::Header as Header>::Number)>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Called by the import queue when it is started. Returns a list of justifications to request from the network.

fn import_justification<'life0, 'async_trait>( &'life0 mut self, hash: <B as Block>::Hash, number: <<B as Block>::Header as Header>::Number, justification: ([u8; 4], Vec<u8>), ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Import a Block justification and finalize the given block.