yet_another_parachain_runtime

Crate sc_consensus_aura

Aura (Authority-round) consensus in substrate.

Aura works by having a list of authorities A who are expected to roughly agree on the current time. Time is divided up into discrete slots of t seconds each. For each slot s, the author of that slot is A[s % |A|].

The author is allowed to issue one block but not more during that slot, and it will be built upon the longest valid chain that has been seen.

Blocks from future steps will be either deferred or rejected depending on how far in the future they are.

NOTE: Aura itself is designed to be generic over the crypto used.

Modules

standalone

Standalone functions used within the implementation of Aura.

Structs

AuraVerifier

A verifier for Aura blocks.

AuthoritiesTracker

AURA authorities tracker. Updates authorities based on the AURA authorities change digest in the block header.

BuildAuraWorkerParams

Parameters of build_aura_worker.

BuildVerifierParams

Parameters of build_verifier.

ImportQueueParams

Parameters of import_queue.

InherentDataProvider

Provides the slot duration inherent data for Aura.

SlotDuration

A slot duration defined in milliseconds.

SlotProportion

A unit type wrapper to express the proportion of a slot.

StartAuraParams

Parameters of start_aura.

Enums

CheckForEquivocation

Should we check for equivocation of a block author?

CompatibilityMode

Run AURA in a compatibility mode.

ConsensusLog

An consensus log item for Aura.

Error

Aura Errors

Constants

AURA_ENGINE_ID

The ConsensusEngineId of AuRa.

INHERENT_IDENTIFIER

The Aura inherent identifier.

Traits

AuraApi

API necessary for block authorship with aura.

CompatibleDigestItem

A digest item which is usable with aura consensus.

SyncOracle

An oracle for when major synchronization work is being undertaken.

Functions

build_aura_worker

Build the aura worker.

build_verifier

Build the AuraVerifier

find_pre_digest

Extract a pre-digest from a block header.

import_queue

Start an import queue for the Aura consensus algorithm.

slot_duration

Get the slot duration for Aura by reading from a runtime API at the best block’s state.

start_aura

Start the aura worker. The returned future should be run in a futures executor.

Type Aliases

AuraInherent

The type of the Aura inherent.