yet_another_parachain_runtime::sc_mixnet

Function run

pub async fn run<B, C, S, P>(
    config: Config,
    api_backend: ApiBackend,
    client: Arc<C>,
    sync: Arc<S>,
    network: Arc<dyn NetworkService>,
    protocol_name: ProtocolName,
    transaction_pool: Arc<P>,
    keystore: Option<Arc<dyn Keystore>>,
    notification_service: Box<dyn NotificationService>,
)
where
    B: Block,
    C: BlockchainEvents<B> + ProvideRuntimeApi<B> + HeaderBackend<B>,
    <C as ProvideRuntimeApi<B>>::Api: MixnetApi<B>,
    S: SyncOracle,
    P: TransactionPool<Block = B> + LocalTransactionPool<Block = B> + 'static,

Run the mixnet service. If keystore is None, the service will not attempt to register the local node as a mixnode, even if config.register is true.