yet_another_parachain_runtime::sc_network::service::traits

Trait NetworkDHTProvider

pub trait NetworkDHTProvider {
    // Required methods
    fn find_closest_peers(&self, target: PeerId);
    fn get_value(&self, key: &Key);
    fn put_value(&self, key: Key, value: Vec<u8>);
    fn put_record_to(
        &self,
        record: Record,
        peers: HashSet<PeerId>,
        update_local_storage: bool,
    );
    fn store_record(
        &self,
        key: Key,
        value: Vec<u8>,
        publisher: Option<PeerId>,
        expires: Option<Instant>,
    );
    fn start_providing(&self, key: Key);
    fn stop_providing(&self, key: Key);
    fn get_providers(&self, key: Key);
}

Provides access to the networking DHT.

Required Methods

fn find_closest_peers(&self, target: PeerId)

Start finding closest peers to the target.

fn get_value(&self, key: &Key)

Start getting a value from the DHT.

fn put_value(&self, key: Key, value: Vec<u8>)

Start putting a value in the DHT.

fn put_record_to( &self, record: Record, peers: HashSet<PeerId>, update_local_storage: bool, )

Start putting the record to peers.

If update_local_storage is true the local storage is udpated as well.

fn store_record( &self, key: Key, value: Vec<u8>, publisher: Option<PeerId>, expires: Option<Instant>, )

Store a record in the DHT memory store.

fn start_providing(&self, key: Key)

Register this node as a provider for key on the DHT.

fn stop_providing(&self, key: Key)

Deregister this node as a provider for key on the DHT.

fn get_providers(&self, key: Key)

Start getting the list of providers for key on the DHT.