yet_another_parachain_runtime::sc_network_sync::block_relay_protocol

Trait BlockDownloader

pub trait BlockDownloader<Block>:
    Debug
    + Send
    + Sync
where
    Block: Block,
{
    // Required methods
    fn protocol_name(&self) -> &ProtocolName;
    fn download_blocks<'life0, 'async_trait>(
        &'life0 self,
        who: PeerId,
        request: BlockRequest<<Block as Block>::Hash, <<Block as Block>::Header as Header>::Number>,
    ) -> Pin<Box<dyn Future<Output = Result<Result<(Vec<u8>, ProtocolName), RequestFailure>, Canceled>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn block_response_into_blocks(
        &self,
        request: &BlockRequest<<Block as Block>::Hash, <<Block as Block>::Header as Header>::Number>,
        response: Vec<u8>,
    ) -> Result<Vec<BlockData<<Block as Block>::Header, <Block as Block>::Hash, <Block as Block>::Extrinsic>>, BlockResponseError>;
}

The client side stub to download blocks from peers. This is a handle that can be used to initiate concurrent downloads.

Required Methods

fn protocol_name(&self) -> &ProtocolName

Protocol name used by block downloader.

fn download_blocks<'life0, 'async_trait>( &'life0 self, who: PeerId, request: BlockRequest<<Block as Block>::Hash, <<Block as Block>::Header as Header>::Number>, ) -> Pin<Box<dyn Future<Output = Result<Result<(Vec<u8>, ProtocolName), RequestFailure>, Canceled>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Performs the protocol specific sequence to fetch the blocks from the peer. Output: if the download succeeds, the response is a Vec<u8> which is in a format specific to the protocol implementation. The block data can be extracted from this response using BlockDownloader::block_response_into_blocks.

fn block_response_into_blocks( &self, request: &BlockRequest<<Block as Block>::Hash, <<Block as Block>::Header as Header>::Number>, response: Vec<u8>, ) -> Result<Vec<BlockData<<Block as Block>::Header, <Block as Block>::Hash, <Block as Block>::Extrinsic>>, BlockResponseError>

Parses the protocol specific response to retrieve the block data.