yet_another_parachain_runtime::sc_transaction_pool_api

Trait LocalTransactionPool

pub trait LocalTransactionPool: Send + Sync {
    type Block: Block;
    type Hash: Hash + Eq + Member + Serialize;
    type Error: From<Error> + IntoPoolError;

    // Required method
    fn submit_local(
        &self,
        at: <Self::Block as Block>::Hash,
        xt: <Self::Block as Block>::Extrinsic,
    ) -> Result<Self::Hash, Self::Error>;
}

Transaction pool interface for submitting local transactions that exposes a blocking interface for submission.

Required Associated Types

type Block: Block

Block type.

type Hash: Hash + Eq + Member + Serialize

Transaction hash type.

type Error: From<Error> + IntoPoolError

Error type.

Required Methods

fn submit_local( &self, at: <Self::Block as Block>::Hash, xt: <Self::Block as Block>::Extrinsic, ) -> Result<Self::Hash, Self::Error>

Submits the given local unverified transaction to the pool blocking the current thread for any necessary pre-verification. NOTE: It MUST NOT be used for transactions that originate from the network or RPC, since the validation is performed with TransactionSource::Local.