yet_another_parachain_runtime

Crate sp_npos_elections

A set of election algorithms to be used with a substrate runtime, typically within the staking sub-system. Notable implementation include:

seq_phragmen: Implements the Phragmén Sequential Method. An un-ranked, relatively fast election method that ensures PJR, but does not provide a constant factor approximation of the maximin problem.

ghragmms: Implements a hybrid approach inspired by Phragmén which is executed faster but it can achieve a constant factor approximation of the maximin problem, similar to that of the MMS algorithm.

balance: Implements the star balancing algorithm. This iterative process can push a solution toward being more “balanced”, which in turn can increase its score.

Terminology

This crate uses context-independent words, not to be confused with staking. This is because the election algorithms of this crate, while designed for staking, can be used in other contexts as well.

Voter: The entity casting some votes to a number of Targets. This is the same as Nominator in the context of staking. Target: The entities eligible to be voted upon. This is the same as Validator in the context of staking. Edge: A mapping from a Voter to a Target.

The goal of an election algorithm is to provide an ElectionResult. A data composed of:

winners: A flat list of identifiers belonging to those who have won the election, usually ordered in some meaningful way. They are zipped with their total backing stake.

assignment: A mapping from each voter to their winner-only targets, zipped with a ration denoting the amount of support given to that particular target.

// the winners.
let winners = vec![(1, 100), (2, 50)];
let assignments = vec![
    // A voter, giving equal backing to both 1 and 2.
    Assignment {
		who: 10,
		distribution: vec![(1, Perbill::from_percent(50)), (2, Perbill::from_percent(50))],
	},
    // A voter, Only backing 1.
    Assignment { who: 20, distribution: vec![(1, Perbill::from_percent(100))] },
];

// the combination of the two makes the election result.
let election_result = ElectionResult { winners, assignments };

The Assignment field of the election result is voter-major, i.e. it is from the perspective of the voter. The struct that represents the opposite is called a Support. This struct is usually accessed in a map-like manner, i.e. keyed by voters, therefore it is stored as a mapping called SupportMap.

Moreover, the support is built from absolute backing values, not ratios like the example above. A struct similar to Assignment that has stake value instead of ratios is called an StakedAssignment.

More information can be found at: https://arxiv.org/abs/2004.12990

Modules

balancing

Balancing algorithm implementation.

helpers

Helper methods for npos-elections.

node

(very) Basic implementation of a graph node used in the reduce algorithm.

phragmen

Implementation of the sequential-phragmen election method.

phragmms

Implementation of the PhragMMS method.

pjr

Implements functions and interfaces to check solutions for being t-PJR.

reduce

Rust implementation of the Phragmén reduce algorithm. This can be used by any off chain application to reduce cycles from the edge assignment, which will result in smaller size.

traits

Traits for the npos-election operations.

Structs

Assignment

A voter’s stake assignment among a set of targets, represented as ratios.

BalancingConfig

Utility struct to group parameters for the balancing algorithm.

Candidate

A candidate entity for the election.

Edge

A vote being casted by a Voter to a Candidate is an Edge.

ElectionResult

Final result of the election.

ElectionScore

The score of an election. This is the main measure of an election’s quality.

StakedAssignment

A voter’s stake assignment among a set of targets, represented as absolute values in the scale of ExtendedBalance.

Support

A structure to demonstrate the election result from the perspective of the candidate, i.e. how much support each candidate is receiving.

Voter

A voter entity.

Enums

Error

The errors that might occur in this crate and frame-election-provider-solution-type.

Traits

Backings

Generic representation of a support.

EvaluateSupport

Extension trait for evaluating a support map or vector.

IdentifierT

an aggregator trait for a generic type of a voter/target identifier. This usually maps to substrate’s account id.

PerThing128

Aggregator trait for a PerThing that can be multiplied by u128 (ExtendedBalance).

Functions

assignment_ratio_to_staked

Converts a vector of ratio assignments into ones with absolute budget value.

assignment_ratio_to_staked_normalized

Same as assignment_ratio_to_staked and try and do normalization.

assignment_staked_to_ratio

Converts a vector of staked assignments into ones with ratio values.

assignment_staked_to_ratio_normalized

Same as assignment_staked_to_ratio and try and do normalization.

balance

Balance the weight distribution of a given voters at most iterations times, or up until the point where the biggest difference created per iteration of all stakes is tolerance. If this is called with tolerance = 0, then exactly iterations rounds will be executed, except if no change has been made (difference = 0). tolerance and iterations are part of the BalancingConfig struct.

evaluate_support

General evaluation of a list of backings that returns an election score.

phragmms

Execute the phragmms method.

pjr_check

Check a solution to be PJR.

pjr_check_core

The internal implementation of the PJR check after having the data converted.

pretty_balance

Format a single ExtendedBalance into a pretty string with the given token symbol and decimals.

reduce

Reduce the given Vec<StakedAssignment<IdentifierT>>. This removes redundant edges without changing the overall backing of any of the elected candidates.

seq_phragmen

Execute sequential phragmen with potentially some rounds of balancing. The return type is list of winners and a weight distribution vector of all voters who contribute to the winners.

seq_phragmen_core

Core implementation of seq-phragmen.

setup_inputs

Converts raw inputs to types used in this crate.

standard_threshold

Compute the threshold corresponding to the standard PJR property

supports_to_staked_assignment

Convert some Supportss into vector of StakedAssignment

t_pjr_check

Check a solution to be t-PJR.

to_support_map

Build the support map from the assignments.

to_supports

Same as to_support_map except it returns a flat vector.

validate_pjr_challenge

Validate a challenge to an election result.

validate_t_pjr_challenge

Validate a challenge to an election result.

Type Aliases

CandidatePtr

A pointer to a candidate struct with interior mutability.

ExtendedBalance

A type in which performing operations on vote weights are safe.

SupportMap

Linkage from a winner to their Support.

Supports

A target-major representation of the the election outcome.

VoteWeight

A type which is used in the API of this crate as a numeric weight of a vote, most often the stake of the voter. It is always converted to ExtendedBalance for computation.