yet_another_parachain_runtime::staging_xcm_builder

Trait DescribeLocation

pub trait DescribeLocation {
    // Required method
    fn describe_location(location: &Location) -> Option<Vec<u8>>;
}

Means of converting a location into a stable and unique descriptive identifier.

Required Methods

fn describe_location(location: &Location) -> Option<Vec<u8>>

Create a description of the given location if possible. No two locations should have the same descriptor.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.