yet_another_parachain_runtime::staging_xcm_builder

Trait EnsureDelivery

pub trait EnsureDelivery {
    // Required method
    fn ensure_successful_delivery(
        origin_ref: &Location,
        dest: &Location,
        fee_reason: FeeReason,
    ) -> (Option<FeesMode>, Option<Assets>);
}

Trait for a type which ensures all requirements for successful delivery with XCM transport layers.

Required Methods

fn ensure_successful_delivery( origin_ref: &Location, dest: &Location, fee_reason: FeeReason, ) -> (Option<FeesMode>, Option<Assets>)

Prepare all requirements for successful XcmSender: SendXcm passing (accounts, balances, channels …). Returns:

possible FeesMode which is expected to be set to executor

possible Assets which are expected to be subsume to the Holding Register

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.