yet_another_parachain_runtime::staging_xcm_executor::traits

Module prelude

Structs

Identity

A structure that performs identity conversion.

JustTry

A structure that performs standard conversion using the standard Rust conversion traits.

WithOriginFilter

Enums

Error

Errors associated with MatchesFungibles operation.

FeeReason

Context under which a fee is paid.

LockError

Traits

AssetExchange

A service for exchanging assets.

AssetLock

Define a handler for notification of an asset being locked and for the unlock instruction.

ClaimAssets

Define any handlers for the AssetClaim instruction.

ConvertOrigin

A converter trait for origin types.

DropAssets

Define a handler for when some non-empty AssetsInHolding value should be dropped.

Enact

EventEmitter

Defines the event emitter for the XCM executor. This trait allows implementations to emit events related to XCM handling, including successful sends and failure cases.

ExportXcm

Utility for delivering a message to a system under a different (non-local) consensus with a spoofed origin. This essentially defines the behaviour of the ExportMessage XCM instruction.

FeeManager

Handle stuff to do with taking fees in certain XCM instructions.

MatchesFungible

MatchesFungibles

MatchesInstance

Unique instances matcher trait.

MatchesNonFungible

MatchesNonFungibles

OnResponse

Define what needs to be done upon receiving a query response.

ProcessTransaction

Provides mechanisms for transactional processing of XCM instructions.

ShouldExecute

Trait to determine whether the execution engine should actually execute a given XCM.

TransactAsset

Facility for asset transacting.

VersionChangeNotifier

Trait for a type which handles notifying a destination of XCM version changes.

WeightBounds

Determine the weight of an XCM message.

WeightTrader

Charge for weight in order to execute XCM.

Functions

export_xcm

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as mutable references into T::send_xcm.

validate_export

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as mutable references into T::send_xcm.