yet_another_parachain_runtime::staging_xcm_executor::traits

Trait XcmAssetTransfers

pub trait XcmAssetTransfers {
    type IsReserve: ContainsPair<Asset, Location>;
    type IsTeleporter: ContainsPair<Asset, Location>;
    type AssetTransactor: TransactAsset;

    // Provided method
    fn determine_for(
        asset: &Asset,
        dest: &Location,
    ) -> Result<TransferType, Error> { ... }
}

A trait for identifying asset transfer type based on IsTeleporter and IsReserve configurations.

Required Associated Types

type IsReserve: ContainsPair<Asset, Location>

Combinations of (Asset, Location) pairs which we trust as reserves. Meaning reserve-based-transfers are to be used for assets matching this filter.

type IsTeleporter: ContainsPair<Asset, Location>

Combinations of (Asset, Location) pairs which we trust as teleporters. Meaning teleports are to be used for assets matching this filter.

type AssetTransactor: TransactAsset

How to withdraw and deposit an asset.

Provided Methods

fn determine_for(asset: &Asset, dest: &Location) -> Result<TransferType, Error>

Determine transfer type to be used for transferring asset from local chain to dest.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.