yet_another_parachain_runtime::substrate_rpc_client

Struct WsClient

pub struct WsClient { /* private fields */ }

Generic asynchronous client.

Implementations

impl Client

pub fn builder() -> ClientBuilder

Create a builder for the server.

pub fn is_connected(&self) -> bool

Checks if the client is connected to the target.

pub async fn disconnect_reason(&self) -> Error

This is similar to Client::on_disconnect but it can be used to get the reason why the client was disconnected but it’s not cancel-safe.

The typical use-case is that this method will be called after Client::on_disconnect has returned in a “select loop”.

Cancel-safety

This method is cancel-safe

pub async fn on_disconnect(&self)

Completes when the client is disconnected or the client’s background task encountered an error. If the client is already disconnected, the future produced by this method will complete immediately.

Cancel safety

This method is cancel safe.