Crate yet_another_parachain_runtime

Re-exports

pub use polkadot_sdk::staging_parachain_info as parachain_info;

Modules

api

asset_test_utils

Module contains predefined test-case scenarios for Runtime with various assets.

assets_common

binary_merkle_tree

This crate implements a simple binary Merkle Tree utilities required for inter-op with Ethereum bridge & Solidity contract.

bp_header_chain

Defines traits which represent a common interface for Substrate pallets which want to incorporate bridge functionality.

bp_messages

Primitives of messages module.

bp_parachains

Primitives of parachains module.

bp_polkadot_core

Primitives of the Polkadot-like chains.

bp_relayers

Primitives of messages module.

bp_runtime

Primitives that may be used at (bridges) runtime level.

bp_test_utils

Utilities for testing runtime code.

bp_xcm_bridge_hub

Primitives of the xcm-bridge-hub pallet.

bp_xcm_bridge_hub_router

Primitives of the xcm-bridge-hub-router pallet.

bridge_hub_common

bridge_hub_test_utils

Module contains predefined test-case scenarios for “BridgeHub” Runtimes.

bridge_runtime_common

Common types/functions that may be used by runtimes of all bridged chains.

cumulus_client_bootnodes

Parachain bootnodes advertisement and discovery service.

cumulus_client_cli

Cumulus CLI library.

cumulus_client_collator

Cumulus Collator implementation for Substrate.

cumulus_client_consensus_aura

The AuRa consensus algorithm for parachains.

cumulus_client_consensus_common

cumulus_client_consensus_relay_chain

The relay-chain provided consensus algorithm for parachains.

cumulus_client_network

Parachain specific networking

cumulus_client_parachain_inherent

Client side code for generating the parachain inherent.

cumulus_client_pov_recovery

Parachain PoV recovery

cumulus_client_service

Cumulus service

cumulus_pallet_aura_ext

Cumulus extension pallet for AuRa

cumulus_pallet_dmp_queue

This pallet used to implement a message queue for downward messages from the relay-chain.

cumulus_pallet_parachain_system

cumulus-pallet-parachain-system is a base pallet for Cumulus-based parachains.

cumulus_pallet_parachain_system_proc_macro

cumulus_pallet_session_benchmarking

Benchmarks for the Session Pallet.

cumulus_pallet_solo_to_para

cumulus_pallet_weight_reclaim

Pallet and transaction extensions to reclaim PoV proof size weight after an extrinsic has been applied.

cumulus_pallet_xcm

Pallet for stuff specific to parachains’ usage of XCM. Right now that’s just the origin used by parachains when receiving Transact messages from other parachains or the Relay chain which must be natively represented.

cumulus_pallet_xcmp_queue

A pallet which uses the XCMP transport layer to handle both incoming and outgoing XCM message sending and dispatch, queuing, signalling and backpressure. To do so, it implements:

cumulus_ping

Pallet to spam the XCM/UMP.

cumulus_primitives_aura

Core primitives for Aura in Cumulus.

cumulus_primitives_core

Cumulus related core primitive types and traits.

cumulus_primitives_parachain_inherent

Cumulus parachain inherent

cumulus_primitives_proof_size_hostfunction

Tools for reclaiming PoV weight in parachain runtimes.

cumulus_primitives_storage_weight_reclaim

Mechanism to reclaim PoV proof size weight after an extrinsic has been applied.

cumulus_primitives_timestamp

Cumulus timestamp related primitives.

cumulus_primitives_utility

Helper datatypes for cumulus. This includes the ParentAsUmp routing type which will route messages into an UpwardMessageSender if the destination is Parent.

cumulus_relay_chain_inprocess_interface

cumulus_relay_chain_interface

cumulus_relay_chain_minimal_node

cumulus_relay_chain_rpc_interface

cumulus_relay_chain_streams

Common utilities for transforming relay chain streams.

cumulus_test_relay_sproof_builder

emulated_integration_tests_common

ethereum_standards

Ethereum standards.

fork_tree

Utility library for managing tree-like ordered data with logic for pruning the tree while finalizing nodes.

frame_benchmarking

Macro for benchmarking a FRAME runtime.

frame_benchmarking_cli

Contains the root BenchmarkCmd command and exports its sub-commands.

frame_benchmarking_pallet_pov

End-to-end testing pallet for PoV benchmarking. Should only be deployed in a testing runtime.

frame_election_provider_solution_type

Proc macro for a npos solution type.

frame_election_provider_support

Primitive traits for providing election functionality.

frame_executive

Executive Module

frame_metadata_hash_extension

The CheckMetadataHash transaction extension.

frame_remote_externalities

Remote Externalities

frame_support

Support code for the runtime.

frame_support_procedural

Proc macro of Support code for the runtime.

frame_support_procedural_tools

Proc macro helpers for procedural macros

frame_support_procedural_tools_derive

Use to derive parsing for parsing struct.

frame_system

System Pallet

frame_system_benchmarking

Frame System benchmarks.

frame_system_rpc_runtime_api

Runtime API definition required by System RPC extensions.

frame_try_runtime

Supporting types for try-runtime, testing and dry-running commands.

generate_bags

Support code to ease the process of generating bag thresholds.

mmr_gadget

MMR offchain gadget

mmr_rpc

Node-specific RPC methods for interaction with Merkle Mountain Range pallet.

pallet_alliance

Alliance Pallet

pallet_asset_conversion_ops

Asset Conversion Operations Suite.

pallet_asset_conversion_tx_payment

Asset Conversion Transaction Payment Pallet

pallet_asset_rate

Asset Rate Pallet

pallet_asset_rewards

FRAME Staking Rewards Pallet

pallet_asset_tx_payment

Asset Transaction Payment Pallet

pallet_assets_freezer

Assets Freezer Pallet

pallet_assets_holder

Assets Holder Pallet

pallet_assets_precompiles

pallet_atomic_swap

Atomic Swap

pallet_authority_discovery

Authority discovery pallet.

pallet_authorship

Authorship tracking for FRAME runtimes.

pallet_babe

Consensus extension module for BABE consensus. Collects on-chain randomness from VRF outputs and manages epoch transitions.

pallet_bags_list

Made with Substrate, for Polkadot.

pallet_beefy

pallet_beefy_mmr

A BEEFY+MMR pallet combo.

pallet_bounties

Bounties Module ( pallet-bounties )

pallet_bridge_grandpa

Substrate GRANDPA Pallet

pallet_bridge_messages

Runtime module that allows sending and receiving messages using lane concept:

pallet_bridge_parachains

Parachains finality module.

pallet_bridge_relayers

Runtime module that is used to store relayer rewards and (in the future) to coordinate relations between relayers.

pallet_broker

Pallet Broker

pallet_child_bounties

Child Bounties Pallet ( pallet-child-bounties )

pallet_collator_selection

Collator Selection pallet.

pallet_collective

Collective system: Members of a set of account IDs can make their collective feelings known through dispatched calls from one of two specialized origins.

pallet_collective_content

Managed Collective Content Pallet

pallet_contracts

Contracts Pallet

pallet_contracts_mock_network

pallet_contracts_proc_macro

Procedural macros used in the contracts module.

pallet_contracts_uapi

External C API to communicate with substrate contracts runtime module.

pallet_conviction_voting

Voting Pallet

pallet_core_fellowship

Additional logic for the Core Fellowship. This determines salary, registers activity/passivity and handles promotion and demotion periods.

pallet_delegated_staking

Delegated Staking Pallet

pallet_democracy

Democracy Pallet

pallet_derivatives

The purpose of the pallet-derivatives is to cover the following derivative asset support scenarios:

pallet_dev_mode

pallet_dummy_dim

A simple interface to interact with a Proof-of-Personhood system.

pallet_election_provider_multi_block

Multi-phase, multi-block, election provider pallet.

pallet_election_provider_multi_phase

Multi phase, offchain election provider pallet.

pallet_election_provider_support_benchmarking

Election provider support pallet benchmarking.

pallet_elections_phragmen

Phragmén Election Module.

pallet_fast_unstake

Made with Substrate, for Polkadot.

pallet_glutton

WARNING

pallet_grandpa

GRANDPA Consensus module for runtime.

pallet_identity

Identity Pallet

pallet_im_online

I’m online Pallet

pallet_indices

An index is a short form of an address. This module handles allocation of indices for a newly created accounts.

pallet_insecure_randomness_collective_flip

DO NOT USE IN PRODUCTION

pallet_lottery

A lottery pallet that uses participation in the network to purchase tickets.

pallet_membership

Membership Module

pallet_meta_tx

Meta Tx (Meta Transaction) Pallet

pallet_migrations

pallet-migrations

pallet_mixnet

This pallet is responsible for determining the current mixnet session and phase, and the mixnode set for each session.

pallet_mmr

Merkle Mountain Range

pallet_multi_asset_bounties

Made with Substrate, for Polkadot.

pallet_multisig

Multisig pallet

pallet_nft_fractionalization

NFT Fractionalization Pallet

pallet_nfts

Nfts Module

pallet_nfts_runtime_api

Runtime API definition for the FRAME NFTs pallet.

pallet_nis

Non-Interactive Staking (NIS) Pallet

pallet_node_authorization

Node authorization pallet

pallet_nomination_pools

Nomination Pools for Staking Delegation

pallet_nomination_pools_benchmarking

Benchmarks for the nomination pools coupled with the staking and bags list pallets.

pallet_nomination_pools_runtime_api

Runtime API definition for nomination-pools pallet. Currently supports only one rpc endpoint.

pallet_offences

Offences Pallet

pallet_offences_benchmarking

Offences pallet benchmarking.

pallet_oracle

Oracle

pallet_oracle_runtime_api

Runtime API definition for the oracle pallet.

pallet_origin_restriction

Origin restriction pallet and transaction extension

pallet_paged_list

Made with Substrate, for DotSama.

pallet_parameters

⚠️ WARNING ⚠️

pallet_people

People Pallet

pallet_preimage

Preimage Pallet

pallet_proxy

Proxy Pallet

pallet_ranked_collective

Ranked collective system.

pallet_recovery

Recovery Pallet

pallet_referenda

Referenda Pallet

pallet_remark

Remark storage pallet. Indexes remarks and stores them off chain.

pallet_revive

Revive Pallet

pallet_revive_proc_macro

Procedural macros used in pallet-revive.

pallet_revive_uapi

External C API to communicate with Polkadot SDK’s pallet-revive module.

pallet_root_offences

Root Offences Pallet

pallet_root_testing

Root Testing Pallet

pallet_safe_mode

Safe Mode

pallet_salary

Make periodic payment to members of a ranked collective according to rank.

pallet_scheduler

Made with Substrate, for Polkadot.

pallet_scored_pool

Scored Pool Pallet

pallet_session

Session Pallet

pallet_session_benchmarking

Offences pallet benchmarking.

pallet_skip_feeless_payment

Skip Feeless Payment Pallet

pallet_society

Society Pallet

pallet_staking

Staking Pallet

pallet_staking_async

Staking Async Pallet

pallet_staking_async_ah_client

The client for AssetHub, intended to be used in the relay chain.

pallet_staking_async_rc_client

The client for the relay chain, intended to be used in AssetHub.

pallet_staking_async_reward_fn

Useful function for inflation for nominated proof of stake.

pallet_staking_async_runtime_api

Runtime API definition for the staking pallet.

pallet_staking_reward_curve

Proc macro to generate the reward curve functions and tests.

pallet_staking_reward_fn

Useful function for inflation for nominated proof of stake.

pallet_staking_runtime_api

Runtime API definition for the staking pallet.

pallet_state_trie_migration

Pallet State Trie Migration

pallet_statement

Supporting pallet for the statement store.

pallet_sudo

Made with Substrate, for Polkadot.

pallet_timestamp

Made with Substrate, for Polkadot.

pallet_tips

Tipping Pallet ( pallet-tips )

pallet_transaction_payment

Transaction Payment Pallet

pallet_transaction_payment_rpc

RPC interface for the transaction payment pallet.

pallet_transaction_payment_rpc_runtime_api

Runtime API definition for transaction payment pallet.

pallet_transaction_storage

Transaction storage pallet. Indexes transactions and manages storage proofs.

pallet_treasury

Made with Substrate, for Polkadot.

pallet_tx_pause

Transaction Pause

pallet_uniques

Unique (Items) Module

pallet_utility

Utility Pallet

pallet_verify_signature

Transaction extension which validates a signature against a payload constructed from a call and the rest of the transaction extension pipeline.

pallet_vesting

Vesting Pallet

pallet_xcm_benchmarks

Pallet that serves no other purpose than benchmarking raw messages Xcm.

pallet_xcm_bridge_hub

Module that adds XCM support to bridge pallets. The pallet allows to dynamically open and close bridges between local (to this pallet location) and remote XCM destinations.

pallet_xcm_bridge_hub_router

Pallet that may be used instead of SovereignPaidRemoteExporter in the XCM router configuration. The main thing that the pallet offers is the dynamic message fee, that is computed based on the bridge queues state. It starts exponentially increasing if the queue between this chain and the sibling/child bridge hub is congested.

pallet_xcm_precompiles

parachains_common

parachains_common_types

parachains_runtimes_test_utils

polkadot_approval_distribution

ApprovalDistribution implementation.

polkadot_availability_bitfield_distribution

The bitfield distribution

polkadot_availability_distribution

polkadot_availability_recovery

Availability Recovery Subsystem of Polkadot.

polkadot_cli

Polkadot CLI library.

polkadot_collator_protocol

The Collator Protocol allows collators and validators talk to each other. This subsystem implements both sides of the collator protocol.

polkadot_core_primitives

Core Polkadot types.

polkadot_dispute_distribution

Sending and receiving of DisputeRequests.

polkadot_erasure_coding

As part of Polkadot’s availability system, certain pieces of data for each block are required to be kept available.

polkadot_gossip_support

This subsystem is responsible for keeping track of session changes and issuing a connection request to the relevant validators on every new session.

polkadot_network_bridge

The Network Bridge Subsystem - protocol multiplexer for Polkadot.

polkadot_node_collation_generation

The collation generation subsystem is the interface between polkadot and the collators.

polkadot_node_core_approval_voting

The Approval Voting Subsystem.

polkadot_node_core_approval_voting_parallel

The Approval Voting Parallel Subsystem.

polkadot_node_core_av_store

Implements a AvailabilityStoreSubsystem.

polkadot_node_core_backing

Implements the CandidateBackingSubsystem.

polkadot_node_core_bitfield_signing

The bitfield signing subsystem produces SignedAvailabilityBitfields once per block.

polkadot_node_core_candidate_validation

The Candidate Validation subsystem.

polkadot_node_core_chain_api

Implements the Chain API Subsystem

polkadot_node_core_chain_selection

Implements the Chain Selection Subsystem.

polkadot_node_core_dispute_coordinator

Implements the dispute coordinator subsystem.

polkadot_node_core_parachains_inherent

The parachain inherent data provider

polkadot_node_core_prospective_parachains

Implementation of the Prospective Parachains subsystem - this tracks and handles prospective parachain fragments and informs other backing-stage subsystems of work to be done.

polkadot_node_core_provisioner

The provisioner is responsible for assembling a relay chain block from a set of available parachain candidates of its choice.

polkadot_node_core_pvf

The PVF validation host. Responsible for coordinating preparation and execution of PVFs.

polkadot_node_core_pvf_checker

Implements the PVF pre-checking subsystem.

polkadot_node_core_pvf_common

Contains functionality related to PVFs that is shared by the PVF host and the PVF workers.

polkadot_node_core_pvf_execute_worker

Contains the logic for executing PVFs. Used by the polkadot-execute-worker binary.

polkadot_node_core_pvf_prepare_worker

Contains the logic for preparing PVFs. Used by the polkadot-prepare-worker binary.

polkadot_node_core_runtime_api

Implements the Runtime API Subsystem

polkadot_node_metrics

Metrics helpers

polkadot_node_network_protocol

Network protocol types for parachains.

polkadot_node_primitives

Primitive types used on the node-side.

polkadot_node_subsystem

Subsystem accumulation.

polkadot_node_subsystem_types

Subsystem trait definitions and message types.

polkadot_node_subsystem_util

Utility module for subsystems

polkadot_omni_node_lib

Polkadot Omni Node Library

polkadot_parachain_primitives

Defines primitive types for creating or validating a parachain.

polkadot_primitives_test_helpers

A set of primitive constructors, to aid in crafting meaningful testcase while reducing repetition.

polkadot_rpc

Polkadot-specific RPCs implementation.

polkadot_runtime_common

Common runtime code for the Relay Chain, e.g. Rococo, Westend, Polkadot, Kusama …

polkadot_runtime_metrics

Runtime metric interface similar to native Prometheus metrics.

polkadot_runtime_parachains

Runtime modules for parachains code.

polkadot_sdk_frame

FRAME

polkadot_statement_distribution

The Statement Distribution Subsystem.

polkadot_statement_table

The statement table.

sc_allocator

Collection of allocator implementations.

sc_authority_discovery

Substrate authority discovery.

sc_basic_authorship

Basic implementation of block-authoring logic.

sc_block_builder

Substrate block builder

sc_chain_spec

This crate includes structs and utilities for defining configuration files (known as chain specification) for both runtime and node.

sc_chain_spec_derive

Macros to derive chain spec extension traits implementation.

sc_cli

Substrate CLI library.

sc_client_api

Substrate client interfaces.

sc_client_db

Client backend that is backed by a database.

sc_consensus

Collection of common consensus specific implementations

sc_consensus_aura

Aura (Authority-round) consensus in substrate.

sc_consensus_babe

BABE (Blind Assignment for Blockchain Extension)

sc_consensus_babe_rpc

RPC api for babe.

sc_consensus_beefy

sc_consensus_beefy_rpc

RPC API for BEEFY.

sc_consensus_epochs

Generic utilities for epoch-based consensus engines.

sc_consensus_grandpa

Integration of the GRANDPA finality gadget into substrate.

sc_consensus_grandpa_rpc

RPC API for GRANDPA.

sc_consensus_manual_seal

A manual sealing engine: the engine listens for rpc calls to seal blocks and create forks. This is suitable for a testing environment.

sc_consensus_pow

Proof of work consensus for Substrate.

sc_consensus_slots

Slots functionality for Substrate.

sc_executor

A crate that provides means of executing/dispatching calls into the runtime.

sc_executor_common

A set of common definitions that are needed for defining execution engines.

sc_executor_polkavm

sc_executor_wasmtime

Defines a WasmRuntime that uses the Wasmtime JIT to execute.

sc_informant

Console informant. Prints sync progress and block events. Runs on the calling thread.

sc_keystore

Keystore (and session key management) for ed25519 based chains like Polkadot.

sc_mixnet

Substrate mixnet service. This implements the Substrate Mix Network Specification.

sc_network

Substrate-specific P2P networking.

sc_network_common

Common data structures of the networking layer.

sc_network_gossip

Polite gossiping.

sc_network_light

Light client data structures of the networking layer.

sc_network_statement

Statement handling to plug on top of the network service.

sc_network_sync

Blockchain syncing implementation in Substrate.

sc_network_transactions

Transactions handling to plug on top of the network service.

sc_network_types

sc_offchain

Substrate offchain workers.

sc_proposer_metrics

Prometheus basic proposer metrics.

sc_rpc

Substrate RPC implementation.

sc_rpc_api

Substrate RPC interfaces.

sc_rpc_server

Substrate RPC servers.

sc_rpc_spec_v2

Substrate JSON-RPC interface v2.

sc_runtime_utilities

Substrate client runtime utilities.

sc_service

Substrate service. Starts a thread that spins up the network, client, and extrinsic pool. Manages communication between them.

sc_state_db

State database maintenance. Handles canonicalization and pruning in the database.

sc_statement_store

Disk-backed statement store.

sc_storage_monitor

sc_sync_state_rpc

A RPC handler to create sync states for light clients.

sc_sysinfo

This crate contains the code necessary to gather basic hardware and software telemetry information about the node on which we’re running.

sc_telemetry

Substrate’s client telemetry is a part of substrate that allows ingesting telemetry data with for example Polkadot telemetry.

sc_tracing

Instrumentation implementation for substrate.

sc_tracing_proc_macro

sc_transaction_pool

Substrate transaction pool implementation.

sc_transaction_pool_api

Transaction pool client facing API.

sc_utils

Utilities Primitives for Substrate

slot_range_helper

A helper macro for generating SlotRange enum.

sp_api_proc_macro

Macros for declaring and implementing runtime apis.

sp_authority_discovery

Runtime Api to help discover authorities.

sp_consensus

Common utilities for building and using consensus engines in substrate.

sp_consensus_babe

Primitives for BABE.

sp_consensus_beefy

Primitives for BEEFY protocol.

sp_consensus_pow

Primitives for Substrate Proof-of-Work (PoW) consensus.

sp_consensus_slots

Primitives for slots-based consensus engines.

sp_core_hashing

This package has been deprecated. Please use sp-crypto-hashing.

sp_core_hashing_proc_macro

This package has been deprecated. Please use sp-crypto-hashing-proc-macro.

sp_crypto_ec_utils

This crate offers elliptic curves types which are compatible with the Arkworks library functionalities.

sp_database

The main database trait, allowing Substrate to store data persistently.

sp_debug_derive

Macros to derive runtime debug implementation.

sp_externalities

Substrate externalities abstraction

sp_keystore

Keystore traits

sp_maybe_compressed_blob

Handling of blobs that may be compressed, based on an 8-byte magic identifier at the head.

sp_mixnet

Substrate mixnet types and runtime interface.

sp_npos_elections

A set of election algorithms to be used with a substrate runtime, typically within the staking sub-system. Notable implementation include:

sp_panic_handler

Custom panic hook with bug report link

sp_rpc

Substrate RPC primitives and utilities.

sp_runtime_interface

Substrate runtime interface

sp_runtime_interface_proc_macro

This crate provides procedural macros for usage within the context of the Substrate runtime interface.

sp_session

Substrate core types around sessions.

sp_staking

A crate which contains primitives that are useful for implementation that uses staking approaches in general. Definitions related to sessions, slashing, etc go here.

sp_state_machine

Substrate state machine implementation.

sp_statement_store

A crate which contains statement-store primitives.

sp_timestamp

Substrate core types and inherents for timestamps.

sp_transaction_pool

Transaction pool runtime facing API.

sp_transaction_storage_proof

Storage proof primitives. Contains types and basic code to extract storage proofs for indexed transactions.

sp_trie

Utility functions to interact with Substrate’s Base-16 Modified Merkle Patricia tree (“trie”).

sp_version_proc_macro

A proc-macro that generates a custom wasm section from a given RuntimeVersion declaration

sp_weights

Primitives for transaction weighting.

staging_chain_spec_builder

Chain Spec Builder

staging_node_inspect

A CLI extension for substrate node, adding sub-command to pretty print debug info about blocks and extrinsics.

staging_parachain_info

Minimal Pallet that injects a ParachainId into Runtime storage from

staging_tracking_allocator

Tracking/limiting global allocator. Calculates the peak allocation between two checkpoints for the whole process. Accepts an optional limit and a failure handler which is called if the limit is overflown.

staging_xcm_builder

XCM-Builder

staging_xcm_executor

subkey

Subkey

substrate_bip39

substrate_build_script_utils

Crate with utility functions for build.rs scripts.

substrate_frame_rpc_support

Combines sc_rpc_api::state::StateApiClient with [frame_support::storage::generator] traits to provide strongly typed chain state queries over rpc.

substrate_frame_rpc_system

System FRAME specific RPC methods.

substrate_rpc_client

Shared JSON-RPC client related code and abstractions.

substrate_state_trie_migration_rpc

Rpc for state migration.

substrate_wasm_builder

Wasm builder is a utility for building a project as a Wasm binary

testnet_parachains_constants

tracing_gum

A wrapper around tracing macros, to provide semi automatic traceID annotation without codebase turnover.

tracing_gum_proc_macro

Generative part of tracing-gum. See tracing-gum for usage documentation.

xcm_procedural

Procedural macros used in XCM.

xcm_runtime_apis

Various runtime APIs to support XCM processing and manipulation.

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

AsEnsureOriginWithArg

BlockExecutionWeight

Weight of executing an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockHashCount

BoundedSlice

A bounded slice.

ConstBool

Const getter for a basic type.

ConstU8

Const getter for a basic type.

ConstU32

Const getter for a basic type.

ConstU64

Const getter for a basic type.

ConstantMultiplier

Implementor of WeightToFee that uses a constant multiplier.

DmpSink

EitherOfDiverse

“OR gate” implementation of EnsureOrigin allowing for different Success types for L and R, with them combined using an Either type.

ExistentialDeposit

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

InMemoryDbWeight

IsInVec

Trivial utility for implementing Contains/OrderedMembership with a Vec.

MessageQueueServiceWeight

Offset

PalletId

A pallet identifier. These are per pallet and should be stored in a registry somewhere.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Period

Permill

A fixed point representation of a number in the range [0, 1].

PotId

QueueFootprint

The resource footprint of a queue.

RelayOrigin

RemoveCollectiveFlip

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

StakingAdminBodyId

TransactionByteFee

TransformOrigin

Transform the origin of an EnqueueMessage via C::convert.

VerifySignatureBenchmarkHelper

Version

Weight

WeightToFee

WeightToFeeCoefficient

One coefficient and its position in the WeightToFee.

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

DispatchClass

A generalized group of dispatch types.

Everything

A Contains implementation that contains every value.

Nothing

A Contains implementation that contains no value.

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

DAYS

EPOCH_DURATION_IN_BLOCKS

HOURS

MILLISECS_PER_BLOCK

MINUTES

NANOYAP

RUNTIME_API_VERSIONS

SLOT_DURATION

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

WEIGHT_REF_TIME_PER_SECOND

YAP

Traits

BuildStorage

Complex storage builder stuff.

Contains

A trait for querying whether a type can be said to “contain” a value.

HandleMessage

Handles incoming messages for a single origin.

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

WeightToFeePolynomial

A trait that describes the weight to fee calculation as polynomial.

Functions

build_state

Build GenesisConfig from a JSON blob not using any defaults and store it in the storage. For more info refer to sp_genesis_builder::GenesisBuilder::build_state.

get_preset

Get the default GenesisConfig as a JSON blob if name is None.

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

We allow root and the StakingAdmin to execute privileged collator selection operations.

CumulusXcm

Executive

Executive: handles dispatch to the various modules.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

MessageQueue

Nonce

Index of a transaction in the chain.

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Session

SessionConfig

SessionHandlers

SignedBlock

A Block signed with a Justification

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

TxExtension

The TransactionExtension to the basic transaction logic.

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

VerifySignature

WeightReclaim

WeightToFeeCoefficients

A list of coefficients that represent a polynomial.

XcmpQueue

Attribute Macros

derive_impl

This attribute can be used to derive a full implementation of a trait based on a local partial impl and an external impl containing defaults that can be overridden in the local impl.