Crate yet_another_parachain_runtime

Modules

api

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

AsEnsureOriginWithArg

BlockExecutionWeight

Weight of executing an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockHashCount

BoundedSlice

A bounded slice.

ConstBool

Const getter for a basic type.

ConstU8

Const getter for a basic type.

ConstU32

Const getter for a basic type.

ConstU64

Const getter for a basic type.

ConstantMultiplier

Implementor of WeightToFee that uses a constant multiplier.

DmpSink

EitherOfDiverse

“OR gate” implementation of EnsureOrigin allowing for different Success types for L and R, with them combined using an Either type.

ExistentialDeposit

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

InMemoryDbWeight

IsInVec

Trivial utility for implementing Contains/OrderedMembership with a Vec.

MessageQueueServiceWeight

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

QueueFootprint

The resource footprint of a queue.

RelayOrigin

RemoveCollectiveFlip

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

TransactionByteFee

TransformOrigin

Transform the origin of an EnqueueMessage via C::convert.

VerifySignatureBenchmarkHelper

Version

Weight

WeightToFee

WeightToFeeCoefficient

One coefficient and its position in the WeightToFee.

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

DispatchClass

A generalized group of dispatch types.

Everything

A Contains implementation that contains every value.

Nothing

A Contains implementation that contains no value.

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

DAYS

EPOCH_DURATION_IN_BLOCKS

HOURS

MILLISECS_PER_BLOCK

MINUTES

NANOYAP

RUNTIME_API_VERSIONS

SLOT_DURATION

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

WEIGHT_REF_TIME_PER_SECOND

YAP

Traits

BuildStorage

Complex storage builder stuff.

Contains

A trait for querying whether a type can be said to “contain” a value.

HandleMessage

Handles incoming messages for a single origin.

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

WeightToFeePolynomial

A trait that describes the weight to fee calculation as polynomial.

Functions

build_state

Build GenesisConfig from a JSON blob not using any defaults and store it in the storage. For more info refer to sp_genesis_builder::GenesisBuilder::build_state.

get_preset

Get the default GenesisConfig as a JSON blob if name is None.

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CumulusXcm

Executive

Executive: handles dispatch to the various modules.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

MessageQueue

Nonce

Index of a transaction in the chain.

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

SessionHandlers

SignedBlock

A Block signed with a Justification

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

TxExtension

The TransactionExtension to the basic transaction logic.

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

VerifySignature

WeightReclaim

WeightToFeeCoefficients

A list of coefficients that represent a polynomial.

XcmpQueue

Attribute Macros

derive_impl

This attribute can be used to derive a full implementation of a trait based on a local partial impl and an external impl containing defaults that can be overridden in the local impl.