yet_another_parachain_runtime::ethereum_standards

Module IERC20

@dev Interface of the ERC-20 standard as defined in the ERC.

interface IERC20 {
    event Transfer(address indexed from, address indexed to, uint256 value);
    event Approval(address indexed owner, address indexed spender, uint256 value);
    function totalSupply() external view returns (uint256);
    function balanceOf(address account) external view returns (uint256);
    function transfer(address to, uint256 value) external returns (bool);
    function allowance(address owner, address spender) external view returns (uint256);
    function approve(address spender, uint256 value) external returns (bool);
    function transferFrom(address from, address to, uint256 value) external returns (bool);
}

Structs

Approval

@dev Emitted when the allowance of a spender for an owner is set by a call to {approve}. value is the new allowance. Event with signature Approval(address,address,uint256) and selector 0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925.

Transfer

@dev Emitted when value tokens are moved from one account (from) to another (to).

allowanceCall

@dev Returns the remaining number of tokens that spender will be allowed to spend on behalf of owner through {transferFrom}. This is zero by default.

allowanceReturn

@dev Returns the remaining number of tokens that spender will be allowed to spend on behalf of owner through {transferFrom}. This is zero by default.

approveCall

@dev Sets a value amount of tokens as the allowance of spender over the caller’s tokens.

approveReturn

@dev Sets a value amount of tokens as the allowance of spender over the caller’s tokens.

balanceOfCall

@dev Returns the value of tokens owned by account. Function with signature balanceOf(address) and selector 0x70a08231.

balanceOfReturn

@dev Returns the value of tokens owned by account. Container type for the return parameters of the balanceOf(address) function.

totalSupplyCall

@dev Returns the value of tokens in existence. Function with signature totalSupply() and selector 0x18160ddd.

totalSupplyReturn

@dev Returns the value of tokens in existence. Container type for the return parameters of the totalSupply() function.

transferCall

@dev Moves a value amount of tokens from the caller’s account to to.

transferFromCall

@dev Moves a value amount of tokens from from to to using the allowance mechanism. value is then deducted from the caller’s allowance.

transferFromReturn

@dev Moves a value amount of tokens from from to to using the allowance mechanism. value is then deducted from the caller’s allowance.

transferReturn

@dev Moves a value amount of tokens from the caller’s account to to.

Enums

IERC20Calls

Container for all the IERC20 function calls.

IERC20Events

Container for all the IERC20 events.